SEC Company Filings

SEC Company Filings

company_filings(
  x,
  ownership = FALSE,
  type = "",
  before = "",
  count = 40,
  page = 1
)

Arguments

x
either a stock ticker, CIK number, or XML document for a company page

ownership
boolean for inclusion of company change filings

type
Type of filing to fetch. NOTE: due to the way the SEC EDGAR system works, it is actually is a 'starts-with' search, so for instance specifying 'type = "10-K" will return "10-K/A" and "10-K405" filings as well. To ensure you only get the type you want, best practice would be to filter the results.

before
yyyymmdd format of latest filing to fetch

count
Number of filings to fetch per page. Valid options are 10, 20, 40, 80, or 100. Other values will result in the closest count.

page
Which page of results to return.

Value

A dataframe of company filings

Examples

# \donttest{
company_filings("AAPL", before = "20170810")

#> No encoding supplied: defaulting to UTF-8.

#> Error in check_result(res): EDGAR request blocked from Undeclared Automated Tool.
#> Please visit https://www.sec.gov/developer for best practices.
#> See https://mwaldstein.github.io/edgarWebR/index.html#ethical-use--fair-access for your responsibilities
#> Consider also setting the environment variable 'EDGARWEBR_USER_AGENT

# }