SEC Filing Details

The SEC generates a html page as an index for every filing it receives containing all the meta-information about the filing. We extract 3 main types of information:

Filing Information

Filing date, accepted date, etc.

Documents

All the documents included in the filing

Filers

Companies included in the filing

Funds

Funds included in the filing

filing_details(x)

# S3 method for character
filing_details(x)

# S3 method for xml_node
filing_details(x)

Arguments

x
URL to a SEC filing index page

Value

A list with the following components:

information

A data.frame as returned by filing_information

documents

A data.frame as returned by filing_documents

filers

A data.frame as returned by filing_filers

funds

A data.frame as returned by filing_funds

Details

For a company, there is typically a single filer and no funds, but many filings for funds get more complicated - e.g. 400+ funds with 100's of companies

NOTE: This can get process intensive for large fund pages. If you don't need all components, try just using filing_info

Examples

# \donttest{
# Typically you'd get the URL from one of the search functions
x <- paste0("https://www.sec.gov/Archives/edgar/data/",
            "712515/000071251517000063/0000712515-17-000063-index.htm")
filing_details(x)

#> $information
#>   type                              description     accession_number
#> 1 10-Q Quarterly report [Sections 13 or 15(d)]: 0000712515-17-000063
#>   filing_date       accepted_date documents period_date changed_date
#> 1  2017-08-08 2017-08-08 17:05:57        87  2017-06-30         <NA>
#>   effective_date   bytes
#> 1           <NA> 8947076
#> 
#> $documents
#>    seq                                                         description
#> 1    1                                       QUARTERLY REPORT ON FORM 10-Q
#> 2    2                                        AWARENESS LETTER OF KPMG LLP
#> 3    3 CERTIFICATION OF CHIEF EXECUTIVE OFFICER PURSUANT TO RULE 13A-14(A)
#> 4    4 CERTIFICATION OF CHIEF FINANCIAL OFFICER PURSUANT TO RULE 13A-14(A)
#> 5    5    CERTIFICATION OF CHIEF EXECUTIVE OFFICER PURSUANT TO SECTION 906
#> 6    6    CERTIFICATION OF CHIEF FINANCIAL OFFICER PURSUANT TO SECTION 906
#> 7   NA                                       Complete submission text file
#> 8    7                                              XBRL INSTANCE DOCUMENT
#> 9    8                             XBRL TAXONOMY EXTENSION SCHEMA DOCUMENT
#> 10   9               XBRL TAXONOMY EXTENSION CALCULATION LINKBASE DOCUMENT
#> 11  10                XBRL TAXONOMY EXTENSION DEFINITION LINKBASE DOCUMENT
#> 12  11                     XBRL TAXONOMY EXTENSION LABEL LINKBASE DOCUMENT
#> 13  12              XBRL TAXONOMY EXTENSION PRESENTATION LINKBASE DOCUMENT
#>                          document
#> 1      ea6302017-q1fy1810qdoc.htm
#> 2  ex-151kpmgawarenessletter6.htm
#> 3    ex-311sec302ceocert63017.htm
#> 4    ex-312sec302cfocert63017.htm
#> 5    ex-321sec906ceocert63017.htm
#> 6    ex-322sec906cfocert63017.htm
#> 7        0000712515-17-000063.txt
#> 8                 ea-20170630.xml
#> 9                 ea-20170630.xsd
#> 10            ea-20170630_cal.xml
#> 11            ea-20170630_def.xml
#> 12            ea-20170630_lab.xml
#> 13            ea-20170630_pre.xml
#>                                                                                                href
#> 1      https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea6302017-q1fy1810qdoc.htm
#> 2  https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ex-151kpmgawarenessletter6.htm
#> 3    https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ex-311sec302ceocert63017.htm
#> 4    https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ex-312sec302cfocert63017.htm
#> 5    https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ex-321sec906ceocert63017.htm
#> 6    https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ex-322sec906cfocert63017.htm
#> 7        https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/0000712515-17-000063.txt
#> 8                 https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630.xml
#> 9                 https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630.xsd
#> 10            https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630_cal.xml
#> 11            https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630_def.xml
#> 12            https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630_lab.xml
#> 13            https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630_pre.xml
#>          type    size
#> 1        10-Q 1428163
#> 2     EX-15.1    4013
#> 3     EX-31.1   14907
#> 4     EX-31.2   15281
#> 5     EX-32.1    7059
#> 6     EX-32.2    7248
#> 7             8947076
#> 8  EX-101.INS 2088048
#> 9  EX-101.SCH   74416
#> 10 EX-101.CAL   92568
#> 11 EX-101.DEF  479648
#> 12 EX-101.LAB  888418
#> 13 EX-101.PRE  574659
#> 
#> $filers
#>            mailing_address_1     mailing_address_2 mailing_address_3
#> 1 209 REDWOOD SHORES PARKWAY REDWOOD CITY CA 94065              <NA>
#>   mailing_address_4         business_address_1    business_address_2
#> 1              <NA> 209 REDWOOD SHORES PARKWAY REDWOOD CITY CA 94065
#>   business_address_3 business_address_4         company_name company_cik
#> 1       650-628-1500               <NA> ELECTRONIC ARTS INC.  0000712515
#>                                                        company_filings_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000712515&action=getcompany
#>   company_irs_number company_incorporation_state company_fiscal_year_end
#> 1          942838567                          DE                    0331
#>   filing_type filing_act
#> 1        10-Q         34
#>                                                               file_number_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?filenum=000-17948&action=getcompany
#>   file_number film_number sic_code
#> 1   000-17948   171015593     7372
#>                                                                            sic_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&SIC=7372&owner=include
#> 
#> $funds
#> data frame with 0 columns and 0 rows
#> 

# }