SEC Filing Information

The SEC generates a html page as an index for every filing it receives containing all the meta-information about the filing.

filing_information(x)

# S3 method for character
filing_information(x)

# S3 method for xml_node
filing_information(x)

Arguments

x
URL or xml_document for a SEC filing index page

Value

A dataframe with all the parsed meta-info on the filing

Details

Information returned:

type

description

accession_number

filing_date

accepted_date

documents

period_date

changed_date

effective_date

filing_bytes

Not all details are valid for all filings, but the column will always be present

If you know you're going to want all the details of a filing, including documents funds and filers, look at `filing_details`

Examples

# Typically you'd get the URL from one of the search functions
x <- paste0("https://www.sec.gov/Archives/edgar/data/",
            "933691/000119312517247698/0001193125-17-247698-index.htm")
filing_information(x)

#>   type           description     accession_number filing_date
#> 1  497 Definitive materials: 0001193125-17-247698  2017-08-04
#>         accepted_date documents period_date changed_date effective_date  bytes
#> 1 2017-08-03 18:23:11        15        <NA>         <NA>     2017-08-04 628369