SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

full_text('intel')

#>    filing_date                                                         name
#> 1   2018-01-03       INFORMATION TABLE of 13F-HR for NEVILLE RODIE SHAW INC
#> 2   2018-01-03                   N-30B-2 for POWERSHARES QQQ TRUST SERIES 1
#> 3   2018-01-03       INFORMATION TABLE of 13F-HR for BEACON FINANCIAL GROUP
#>                                                                                                     href
#> 1      https://www.sec.gov/Archives/edgar/data/71210/000007121018000001/xslForm13F_X01/nrs13fdec2017.xml
#> 2                   https://www.sec.gov/Archives/edgar/data/1067839/000119312518001208/d432220dn30b2.htm
#> 3  https://www.sec.gov/Archives/edgar/data/1591379/000159137918000001/xslForm13F_X01/122917INFOTABLE.xml
#>                               company_name     cik  sic
#> 1                   NEVILLE RODIE SHAW INC   71210 <NA>
#> 2           POWERSHARES QQQ TRUST SERIES 1 1067839 <NA>
#> 3                   BEACON FINANCIAL GROUP 1591379 <NA>
#>                                                                                                                                                                                                                                                                                                        content
#> 1     3M COMPANY COM 88579Y101 5754 24446 SH SOLE 5545 0 18901 3M COMPANY COM 88579Y101 635 2700 SH OTR 0 0 2700 ABBOTT LABS COM 002824100 2319 40629 SH SOLE 20424 0 20205 ABBOTT LABS COM 002824100 1481 25948 SH OTR 11998 0 13950 ABBVIE INC COM 00287Y109 8902 92050 SH SOLE 25455 0 66595 ABBVIE INC COM
#> 2                                                                                                                                      September 30, 2017 2017 Annual Report to Shareholders QQQ PowerShares QQQ Trust SM , Series 1 Table of Contents PowerShares QQQ Trust SM , Series 1 1 2 4 5 6 7 8 13 14
#> 3   ABBOTT LABS COM 002824100 1370 24011 SH SOLE NONE 24011 0 0 ABBVIE INC COM 00287Y109 918 9495 SH SOLE NONE 9495 0 0 AFLAC INC COM 001055102 433 4938 SH SOLE NONE 4938 0 0 ALPHABET INC CL A 02079K305 892 847 SH SOLE NONE 847 0 0 ALPHABET INC CL C 02079K107 1031 985 SH SOLE NONE 985 0 0 ALTRIA GROUP
#>                                                                                          parent_href
#> 1    https://www.sec.gov/Archives/edgar/data/71210/000007121018000001/xslForm13F_X01/primary_doc.xml
#> 2                                                                                               <NA>
#> 3       https://www.sec.gov/Archives/edgar/data/1708131/000105640418000051/csc17sa8_absee-201712.htm
#>                                                                                           index_href
#> 1    https://www.sec.gov/Archives/edgar/data/71210/000007121018000001/0000071210-18-000001-index.htm
#> 2                                                                                               <NA>
#> 3  https://www.sec.gov/Archives/edgar/data/1708131/000105640418000051/0001056404-18-000051-index.htm
#>  [ reached getOption("max.print") -- omitted 7 rows ]