SEC Header Search

Seaches filing headers going back to 1994 excluding the most recent day using the interface https://www.sec.gov/cgi-bin/srch-edgar.

header_search(q, page = 1, from = 1994, to = 2017)

Arguments

q
The search string. Documentation https://www.sec.gov/edgar/searchedgar/search_help.htm

page
Which results page to return (default: 1)

from
Start year (default: 1994)

to
End year (default: Current year)

Value

A dataframe of funds found including the following columns -

company_name

filing_href

form

filing_date

size

Examples

header_search("company-name = Apple")

#>                           company_name
#> 1                 ADS APPLE VALLEY INC
#> 2                 ADS APPLE VALLEY INC
#> 3                 ADS APPLE VALLEY INC
#> 4                 ADS APPLE VALLEY INC
#> 5                 ADS APPLE VALLEY INC
#> 6     ADS APPLE VALLEY LTD PARTNERSHIP
#>                                                                                          filing_href
#> 1  https://www.sec.gov/Archives/edgar/data/1281227/000110465907053957/0001104659-07-053957-index.htm
#> 2  https://www.sec.gov/Archives/edgar/data/1281227/000110465907053944/0001104659-07-053944-index.htm
#> 3  https://www.sec.gov/Archives/edgar/data/1281227/000095011605002515/0000950116-05-002515-index.htm
#> 4  https://www.sec.gov/Archives/edgar/data/1281227/000095011605002291/0000950116-05-002291-index.htm
#> 5  https://www.sec.gov/Archives/edgar/data/1281227/000095011604001921/0000950116-04-001921-index.htm
#> 6  https://www.sec.gov/Archives/edgar/data/1281225/000110465907053944/0001104659-07-053944-index.htm
#>       form filing_date    size
#> 1   POSASR  2007-07-16 1032070
#> 2   POS AM  2007-07-16 1095382
#> 3    S-3/A  2005-07-22 1497372
#> 4      S-3  2005-06-28 1961178
#> 5    S-4/A  2004-06-17 3237214
#> 6   POS AM  2007-07-16 1095382
#>  [ reached getOption("max.print") -- omitted 75 rows ]