Parse Submission

Raw SEC filings are sent in a SGML file - this parses that master submission into component documents, with content lines in list column 'TEXT'.

parse_submission(uri)

Arguments

uri
- URL to a SEC complete submission text file

Value

a dataframe with one row per document.

Details

Most of the time the information you need along with the specific files will be available by using filing_documents, but there are scenarios where you may want to access the full contents of the master submission -

Old Submissions

Older submissions are not parsed into component documents by the SEC so access requires parsing the main filing

Full Document List

The SEC only provides what it considers the relevant documents, but filings often include many more ancillary files

Efficient Downloading

If you're fetching many documents from a filing over many filings, there can be efficiency gains from just downloading a single file.

*NOTE:* non-text documents are uuencoded and need a separate decoder to be viewed.

Examples

parse_submission(paste0('https://www.sec.gov/Archives/edgar/data/',
                 '37996/000003799617000084/0000037996-17-000084.txt'))[ ,
                   c('SEQUENCE', 'TYPE', 'DESCRIPTION', 'FILENAME')]

#>   SEQUENCE    TYPE DESCRIPTION                        FILENAME
#> 1        1     8-K         8-K       ceostrategicupdate8-k.htm
#> 2        2   EX-99  EXHIBIT 99    exhibit99ceostrategicupd.htm
#> 3        3 GRAPHIC        <NA> exhibit99ceostrategicupd001.jpg
#> 4        4 GRAPHIC        <NA> exhibit99ceostrategicupd002.jpg
#> 5        5 GRAPHIC        <NA> exhibit99ceostrategicupd003.jpg
#> 6        6 GRAPHIC        <NA> exhibit99ceostrategicupd004.jpg
#> 7        7 GRAPHIC        <NA> exhibit99ceostrategicupd005.jpg