SEC Variable Insurance Search

Provides access to the results of the SEC's Variable Insurance Product search tool available here

variable_insurance_search(term)

variable_insurance_fast_search(identifier)

Arguments

term
Search term to search for in a company, fund or contract name

identifier
A Series, Class/Contract ID, Ticker Symbol or CIK

Value

A dataframe of products found including the following columns -

class_id

class_filings_href

class_name

class_ticker

series_id

series_filings_href

series_name

series_funds_href

cik

cik_name

cik_filings_href

cik_funds_href

Functions

variable_insurance_fast_search: Performs a 'Fast Search' based on an identifier

Examples

variable_insurance_search("precious metals")

#>      class_id
#> 1  C000000347
#> 2  C000000348
#> 3  C000000349
#> 4  C000188944
#> 5  C000191892
#> 6  C000071332
#> 7  C000000350
#> 8  C000078594
#> 9  C000098001
#> 10 C000098002
#> 11 C000098003
#> 12 C000098004
#> 13 C000177601
#> 14 C000098005
#> 15 C000032636
#> 16 C000032635
#> 17 C000032634
#> 18 C000032633
#> 19 C000033472
#> 20 C000033471
#> 21 C000122777
#> 22 C000113088
#> 23 C000016746
#> 24 C000016747
#> 25 C000016748
#> 26 C000016749
#> 27 C000016750
#> 28 C000063733
#> 29 C000009495
#> 30 C000009496
#> 31 C000009497
#> 32 C000009498
#> 33 C000018489
#> 34 C000018486
#> 35 C000018487
#> 36 C000018488
#> 37 C000128741
#> 38 C000100708
#> 39 C000138497
#> 40 C000068252
#> 41 C000120416
#> 42 C000118120
#> 43 C000118121
#> 44 C000012271
#> 45 C000008403
#> 46 C000008404
#> 47 C000011103
#> 48 C000012431
#> 49 C000012306
#> 50 C000012307
#> 51 C000038839
#> 52 C000038815
#> 53 C000038864
#> 54 C000008793
#> 55 C000008794
#> 56 C000008790
#> 57 C000008791
#> 58 C000042979
#> 59 C000008792
#> 60 C000010580
#> 61 C000010581
#> 62 C000010579
#> 63 C000177738
#> 64 C000010578
#> 65 C000010412
#> 66 C000074490
#> 67 C000074491
#> 68 C000117583
#> 69 C000012994
#> 70 C000130162
#> 71 C000008672
#> 72 C000091158
#> 73 C000066842
#> 74 C000034879
#> 75 C000092800
#> 76 C000089457
#> 77 C000089458
#> 78 C000089459
#> 79 C000089460
#>                                                                       class_filings_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000000347&action=getcompany&scd=filings
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000000348&action=getcompany&scd=filings
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000000349&action=getcompany&scd=filings
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000188944&action=getcompany&scd=filings
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000191892&action=getcompany&scd=filings
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000071332&action=getcompany&scd=filings
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000000350&action=getcompany&scd=filings
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000078594&action=getcompany&scd=filings
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000098001&action=getcompany&scd=filings
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000098002&action=getcompany&scd=filings
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000098003&action=getcompany&scd=filings
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000098004&action=getcompany&scd=filings
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000177601&action=getcompany&scd=filings
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000098005&action=getcompany&scd=filings
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000032636&action=getcompany&scd=filings
#> 16 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000032635&action=getcompany&scd=filings
#> 17 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000032634&action=getcompany&scd=filings
#> 18 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000032633&action=getcompany&scd=filings
#> 19 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000033472&action=getcompany&scd=filings
#> 20 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000033471&action=getcompany&scd=filings
#> 21 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000122777&action=getcompany&scd=filings
#> 22 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000113088&action=getcompany&scd=filings
#> 23 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000016746&action=getcompany&scd=filings
#> 24 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000016747&action=getcompany&scd=filings
#> 25 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000016748&action=getcompany&scd=filings
#> 26 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000016749&action=getcompany&scd=filings
#> 27 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000016750&action=getcompany&scd=filings
#> 28 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000063733&action=getcompany&scd=filings
#> 29 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000009495&action=getcompany&scd=filings
#> 30 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000009496&action=getcompany&scd=filings
#> 31 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000009497&action=getcompany&scd=filings
#> 32 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000009498&action=getcompany&scd=filings
#> 33 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000018489&action=getcompany&scd=filings
#> 34 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000018486&action=getcompany&scd=filings
#> 35 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000018487&action=getcompany&scd=filings
#> 36 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000018488&action=getcompany&scd=filings
#> 37 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000128741&action=getcompany&scd=filings
#> 38 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000100708&action=getcompany&scd=filings
#> 39 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000138497&action=getcompany&scd=filings
#> 40 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000068252&action=getcompany&scd=filings
#> 41 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000120416&action=getcompany&scd=filings
#> 42 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000118120&action=getcompany&scd=filings
#> 43 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000118121&action=getcompany&scd=filings
#> 44 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000012271&action=getcompany&scd=filings
#> 45 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000008403&action=getcompany&scd=filings
#> 46 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000008404&action=getcompany&scd=filings
#> 47 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000011103&action=getcompany&scd=filings
#> 48 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000012431&action=getcompany&scd=filings
#> 49 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000012306&action=getcompany&scd=filings
#> 50 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000012307&action=getcompany&scd=filings
#> 51 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000038839&action=getcompany&scd=filings
#> 52 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000038815&action=getcompany&scd=filings
#> 53 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000038864&action=getcompany&scd=filings
#> 54 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000008793&action=getcompany&scd=filings
#> 55 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000008794&action=getcompany&scd=filings
#> 56 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000008790&action=getcompany&scd=filings
#> 57 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000008791&action=getcompany&scd=filings
#> 58 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000042979&action=getcompany&scd=filings
#> 59 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000008792&action=getcompany&scd=filings
#> 60 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000010580&action=getcompany&scd=filings
#> 61 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000010581&action=getcompany&scd=filings
#> 62 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000010579&action=getcompany&scd=filings
#> 63 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000177738&action=getcompany&scd=filings
#> 64 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000010578&action=getcompany&scd=filings
#> 65 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000010412&action=getcompany&scd=filings
#> 66 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000074490&action=getcompany&scd=filings
#> 67 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000074491&action=getcompany&scd=filings
#> 68 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000117583&action=getcompany&scd=filings
#> 69 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000012994&action=getcompany&scd=filings
#> 70 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000130162&action=getcompany&scd=filings
#> 71 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000008672&action=getcompany&scd=filings
#> 72 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000091158&action=getcompany&scd=filings
#> 73 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000066842&action=getcompany&scd=filings
#> 74 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000034879&action=getcompany&scd=filings
#> 75 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000092800&action=getcompany&scd=filings
#> 76 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000089457&action=getcompany&scd=filings
#> 77 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000089458&action=getcompany&scd=filings
#> 78 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000089459&action=getcompany&scd=filings
#> 79 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000089460&action=getcompany&scd=filings
#>                                                     class_name class_ticker
#> 1                                                      Class A        IGDAX
#> 2                                                      Class B        IGDBX
#> 3                                                      Class C        IGDCX
#> 4                                                     Class R6             
#> 5                                                      Class T             
#> 6                                                      Class Y        IGDYX
#> 7                                               Investor Class        FGLDX
#> 8  Jefferies|TR/J CRB Global Precious Metals Equity Index Fund             
#> 9                                                      Class A        SGDAX
#> 10                                                     Class B        SGDBX
#> 11                                                     Class C        SGDCX
#> 12                                                     Class S        SCGDX
#> 13                                                     Class T        SGDTX
#> 14                                         Institutional Class        SGDIX
#> 15                                              Investor Class             
#> 16                                               Service Class             
#> 17                                              Investor Class             
#> 18                                               Service Class             
#> 19                          VP Precious Metals Bear 1.25X Fund             
#> 20                          VP Precious Metals Bull 1.25X Fund             
#> 21               Fairfax Gold and Precious Metals Fund Class I        GOLLX
#> 22               Fairfax Gold and Precious Metals Fund Class N        GOLMX
#> 23                                                     Class A        SGDAX
#> 24                                                  Class AARP        SGLDX
#> 25                                                     Class B        SGDBX
#> 26                                                     Class C        SGDCX
#> 27                                                     Class S        SCGDX
#> 28                                         Institutional Class             
#> 29                                                           A        EKWAX
#> 30                                                           B        EKWBX
#> 31                                                           C        EKWCX
#> 32                                                           I        EKWYX
#> 33                                               ADVISOR CLASS        FGADX
#> 34                                                     CLASS A        FKRCX
#> 35                                                     CLASS B        FAGPX
#> 36                                                     CLASS C        FRGOX
#> 37                                                    CLASS R6        FGPMX
#> 38                     IQ Global Precious Metals Small Cap ETF         JEWL
#> 39           ( )PowerShares Precious Metals Strategy Portfolio             
#> 40                 Invesco Global Gold and Precious Metals ETF         PSAU
#> 41                                                     CLASS P             
#> 42                                                     CLASS I             
#> 43                                                     CLASS P             
#> 44                 PRECIOUS METALS ULTRASECTOR PROFUND CLASS A             
#> 45          PRECIOUS METALS ULTRASECTOR PROFUND INVESTOR CLASS        PMPIX
#> 46           PRECIOUS METALS ULTRASECTOR PROFUND SERVICE CLASS        PMPSX
#> 47                                  PROFUND VP PRECIOUS METALS             
#> 48                            PROFUND VP SHORT PRECIOUS METALS             
#> 49                SHORT PRECIOUS METALS PROFUND INVESTOR CLASS        SPPIX
#> 50                 SHORT PRECIOUS METALS PROFUND SERVICE CLASS        SPPSX
#> 51                             ProShares Short Precious Metals             
#> 52                             ProShares Ultra Precious Metals             
#> 53                        ProShares UltraShort Precious Metals             
#> 54         RiverSource Precious Metals and Mining Fund Class A        INPMX
#> 55         RiverSource Precious Metals and Mining Fund Class B        INPBX
#> 56         RiverSource Precious Metals and Mining Fund Class C        RPMCX
#> 57         RiverSource Precious Metals and Mining Fund Class I             
#> 58        RiverSource Precious Metals and Mining Fund Class R4        AEVYX
#> 59         RiverSource Precious Metals and Mining Fund Class Y        AEVYX
#> 60                                                     Class A        RYMNX
#> 61                                                     Class C        RYZCX
#> 62                                                     Class H        RYMPX
#> 63                                                     Class T             
#> 64                                              Investor Class        RYPMX
#> 65                                            Variable Annuity             
#> 66                                              Class I Shares        DWGOX
#> 67                                             Class II Shares             
#> 68                                        Institutional Shares             
#> 69                               Gold and Precious Metals Fund        USERX
#> 70                                  Institutional Class Shares        USEIX
#> 71                           Precious Metals and Minerals Fund        USAGX
#> 72            Precious Metals and Minerals Fund Adviser Shares        UPMMX
#> 73      Precious Metals and Minerals Fund Institutional Shares        UIPMX
#> 74                    Precious Metals and Minerals Fund Shares        USAGX
#> 75                                         Administrator Class        EKWDX
#> 76                                                     Class A        EKWAX
#> 77                                                     Class B        EKWBX
#> 78                                                     Class C        EKWCX
#> 79                                         Institutional Class        EKWYX
#>     series_id
#> 1  S000000157
#> 2  S000000157
#> 3  S000000157
#> 4  S000000157
#> 5  S000000157
#> 6  S000000157
#> 7  S000000157
#> 8  S000026191
#> 9  S000031521
#> 10 S000031521
#> 11 S000031521
#> 12 S000031521
#> 13 S000031521
#> 14 S000031521
#> 15 S000011955
#> 16 S000011955
#> 17 S000011954
#> 18 S000011954
#> 19 S000012297
#> 20 S000012296
#> 21 S000036961
#> 22 S000036961
#> 23 S000006096
#> 24 S000006096
#> 25 S000006096
#> 26 S000006096
#> 27 S000006096
#> 28 S000006096
#> 29 S000003434
#> 30 S000003434
#> 31 S000003434
#> 32 S000003434
#> 33 S000006839
#> 34 S000006839
#> 35 S000006839
#> 36 S000006839
#> 37 S000006839
#> 38 S000032651
#> 39 S000044513
#> 40 S000023333
#> 41 S000039143
#> 42 S000038270
#> 43 S000038270
#> 44 S000003094
#> 45 S000003094
#> 46 S000003094
#> 47 S000003976
#> 48 S000004533
#> 49 S000004469
#> 50 S000004469
#> 51 S000014280
#> 52 S000014256
#> 53 S000014305
#> 54 S000003299
#> 55 S000003299
#> 56 S000003299
#> 57 S000003299
#> 58 S000003299
#> 59 S000003299
#> 60 S000003798
#> 61 S000003798
#> 62 S000003798
#> 63 S000003798
#> 64 S000003798
#> 65 S000003743
#> 66 S000025038
#> 67 S000025038
#> 68 S000025038
#> 69 S000004783
#> 70 S000004783
#> 71 S000003254
#> 72 S000012910
#> 73 S000012910
#> 74 S000012910
#> 75 S000029099
#> 76 S000029099
#> 77 S000029099
#> 78 S000029099
#> 79 S000029099
#>                                                                      series_filings_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000000157&action=getcompany&scd=filings
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000000157&action=getcompany&scd=filings
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000000157&action=getcompany&scd=filings
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000000157&action=getcompany&scd=filings
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000000157&action=getcompany&scd=filings
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000000157&action=getcompany&scd=filings
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000000157&action=getcompany&scd=filings
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000026191&action=getcompany&scd=filings
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000031521&action=getcompany&scd=filings
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000031521&action=getcompany&scd=filings
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000031521&action=getcompany&scd=filings
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000031521&action=getcompany&scd=filings
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000031521&action=getcompany&scd=filings
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000031521&action=getcompany&scd=filings
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000011955&action=getcompany&scd=filings
#> 16 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000011955&action=getcompany&scd=filings
#> 17 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000011954&action=getcompany&scd=filings
#> 18 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000011954&action=getcompany&scd=filings
#> 19 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000012297&action=getcompany&scd=filings
#> 20 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000012296&action=getcompany&scd=filings
#> 21 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000036961&action=getcompany&scd=filings
#> 22 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000036961&action=getcompany&scd=filings
#> 23 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000006096&action=getcompany&scd=filings
#> 24 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000006096&action=getcompany&scd=filings
#> 25 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000006096&action=getcompany&scd=filings
#> 26 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000006096&action=getcompany&scd=filings
#> 27 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000006096&action=getcompany&scd=filings
#> 28 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000006096&action=getcompany&scd=filings
#> 29 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003434&action=getcompany&scd=filings
#> 30 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003434&action=getcompany&scd=filings
#> 31 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003434&action=getcompany&scd=filings
#> 32 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003434&action=getcompany&scd=filings
#> 33 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000006839&action=getcompany&scd=filings
#> 34 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000006839&action=getcompany&scd=filings
#> 35 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000006839&action=getcompany&scd=filings
#> 36 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000006839&action=getcompany&scd=filings
#> 37 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000006839&action=getcompany&scd=filings
#> 38 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000032651&action=getcompany&scd=filings
#> 39 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000044513&action=getcompany&scd=filings
#> 40 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000023333&action=getcompany&scd=filings
#> 41 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000039143&action=getcompany&scd=filings
#> 42 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000038270&action=getcompany&scd=filings
#> 43 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000038270&action=getcompany&scd=filings
#> 44 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003094&action=getcompany&scd=filings
#> 45 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003094&action=getcompany&scd=filings
#> 46 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003094&action=getcompany&scd=filings
#> 47 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003976&action=getcompany&scd=filings
#> 48 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000004533&action=getcompany&scd=filings
#> 49 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000004469&action=getcompany&scd=filings
#> 50 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000004469&action=getcompany&scd=filings
#> 51 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000014280&action=getcompany&scd=filings
#> 52 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000014256&action=getcompany&scd=filings
#> 53 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000014305&action=getcompany&scd=filings
#> 54 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003299&action=getcompany&scd=filings
#> 55 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003299&action=getcompany&scd=filings
#> 56 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003299&action=getcompany&scd=filings
#> 57 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003299&action=getcompany&scd=filings
#> 58 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003299&action=getcompany&scd=filings
#> 59 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003299&action=getcompany&scd=filings
#> 60 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003798&action=getcompany&scd=filings
#> 61 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003798&action=getcompany&scd=filings
#> 62 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003798&action=getcompany&scd=filings
#> 63 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003798&action=getcompany&scd=filings
#> 64 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003798&action=getcompany&scd=filings
#> 65 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003743&action=getcompany&scd=filings
#> 66 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000025038&action=getcompany&scd=filings
#> 67 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000025038&action=getcompany&scd=filings
#> 68 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000025038&action=getcompany&scd=filings
#> 69 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000004783&action=getcompany&scd=filings
#> 70 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000004783&action=getcompany&scd=filings
#> 71 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000003254&action=getcompany&scd=filings
#> 72 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000012910&action=getcompany&scd=filings
#> 73 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000012910&action=getcompany&scd=filings
#> 74 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000012910&action=getcompany&scd=filings
#> 75 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000029099&action=getcompany&scd=filings
#> 76 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000029099&action=getcompany&scd=filings
#> 77 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000029099&action=getcompany&scd=filings
#> 78 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000029099&action=getcompany&scd=filings
#> 79 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000029099&action=getcompany&scd=filings
#>                                                    series_name
#> 1                          INVESCO Gold & Precious Metals Fund
#> 2                          INVESCO Gold & Precious Metals Fund
#> 3                          INVESCO Gold & Precious Metals Fund
#> 4                          INVESCO Gold & Precious Metals Fund
#> 5                          INVESCO Gold & Precious Metals Fund
#> 6                          INVESCO Gold & Precious Metals Fund
#> 7                          INVESCO Gold & Precious Metals Fund
#> 8  Jefferies|TR/J CRB Global Precious Metals Equity Index Fund
#> 9                         Deutsche Gold & Precious Metals Fund
#> 10                        Deutsche Gold & Precious Metals Fund
#> 11                        Deutsche Gold & Precious Metals Fund
#> 12                        Deutsche Gold & Precious Metals Fund
#> 13                        Deutsche Gold & Precious Metals Fund
#> 14                        Deutsche Gold & Precious Metals Fund
#> 15                                Precious Metals Bear 2X Fund
#> 16                                Precious Metals Bear 2X Fund
#> 17                                Precious Metals Bull 2X Fund
#> 18                                Precious Metals Bull 2X Fund
#> 19                          VP Precious Metals Bear 1.25X Fund
#> 20                          VP Precious Metals Bull 1.25X Fund
#> 21                       Fairfax Gold and Precious Metals Fund
#> 22                       Fairfax Gold and Precious Metals Fund
#> 23                             DWS Gold & Precious Metals Fund
#> 24                             DWS Gold & Precious Metals Fund
#> 25                             DWS Gold & Precious Metals Fund
#> 26                             DWS Gold & Precious Metals Fund
#> 27                             DWS Gold & Precious Metals Fund
#> 28                             DWS Gold & Precious Metals Fund
#> 29                              Evergreen Precious Metals Fund
#> 30                              Evergreen Precious Metals Fund
#> 31                              Evergreen Precious Metals Fund
#> 32                              Evergreen Precious Metals Fund
#> 33                        FRANKLIN GOLD & PRECIOUS METALS FUND
#> 34                        FRANKLIN GOLD & PRECIOUS METALS FUND
#> 35                        FRANKLIN GOLD & PRECIOUS METALS FUND
#> 36                        FRANKLIN GOLD & PRECIOUS METALS FUND
#> 37                        FRANKLIN GOLD & PRECIOUS METALS FUND
#> 38                     IQ Global Precious Metals Small Cap ETF
#> 39           ( )PowerShares Precious Metals Strategy Portfolio
#> 40                 Invesco Global Gold and Precious Metals ETF
#> 41                                     PF Precious Metals Fund
#> 42                                   PRECIOUS METALS PORTFOLIO
#> 43                                   PRECIOUS METALS PORTFOLIO
#> 44                         PRECIOUS METALS ULTRASECTOR PROFUND
#> 45                         PRECIOUS METALS ULTRASECTOR PROFUND
#> 46                         PRECIOUS METALS ULTRASECTOR PROFUND
#> 47                                  PROFUND VP PRECIOUS METALS
#> 48                            PROFUND VP SHORT PRECIOUS METALS
#> 49                               SHORT PRECIOUS METALS PROFUND
#> 50                               SHORT PRECIOUS METALS PROFUND
#> 51                             ProShares Short Precious Metals
#> 52                             ProShares Ultra Precious Metals
#> 53                        ProShares UltraShort Precious Metals
#> 54                 RiverSource Precious Metals and Mining Fund
#> 55                 RiverSource Precious Metals and Mining Fund
#> 56                 RiverSource Precious Metals and Mining Fund
#> 57                 RiverSource Precious Metals and Mining Fund
#> 58                 RiverSource Precious Metals and Mining Fund
#> 59                 RiverSource Precious Metals and Mining Fund
#> 60                                        Precious Metals Fund
#> 61                                        Precious Metals Fund
#> 62                                        Precious Metals Fund
#> 63                                        Precious Metals Fund
#> 64                                        Precious Metals Fund
#> 65                                        Precious Metals Fund
#> 66                         Dynamic Gold & Precious Metals Fund
#> 67                         Dynamic Gold & Precious Metals Fund
#> 68                         Dynamic Gold & Precious Metals Fund
#> 69                               Gold and Precious Metals Fund
#> 70                               Gold and Precious Metals Fund
#> 71                      USAA Precious Metals and Minerals Fund
#> 72                           Precious Metals and Minerals Fund
#> 73                           Precious Metals and Minerals Fund
#> 74                           Precious Metals and Minerals Fund
#> 75                            Wells Fargo Precious Metals Fund
#> 76                            Wells Fargo Precious Metals Fund
#> 77                            Wells Fargo Precious Metals Fund
#> 78                            Wells Fargo Precious Metals Fund
#> 79                            Wells Fargo Precious Metals Fund
#>    series_funds_href        cik
#> 1               <NA> 0000725781
#> 2               <NA> 0000725781
#> 3               <NA> 0000725781
#> 4               <NA> 0000725781
#> 5               <NA> 0000725781
#> 6               <NA> 0000725781
#> 7               <NA> 0000725781
#> 8               <NA> 0001414040
#> 9               <NA> 0000088048
#> 10              <NA> 0000088048
#> 11              <NA> 0000088048
#> 12              <NA> 0000088048
#> 13              <NA> 0000088048
#> 14              <NA> 0000088048
#> 15              <NA> 0001040587
#> 16              <NA> 0001040587
#> 17              <NA> 0001040587
#> 18              <NA> 0001040587
#> 19              <NA> 0001102060
#> 20              <NA> 0001102060
#> 21              <NA> 0000889392
#> 22              <NA> 0000889392
#> 23              <NA> 0000832946
#> 24              <NA> 0000832946
#> 25              <NA> 0000832946
#> 26              <NA> 0000832946
#> 27              <NA> 0000832946
#> 28              <NA> 0000832946
#> 29              <NA> 0001046027
#> 30              <NA> 0001046027
#> 31              <NA> 0001046027
#> 32              <NA> 0001046027
#> 33              <NA> 0000083293
#> 34              <NA> 0000083293
#> 35              <NA> 0000083293
#> 36              <NA> 0000083293
#> 37              <NA> 0000083293
#> 38              <NA> 0001415995
#> 39              <NA> 0001595386
#> 40              <NA> 0001378872
#> 41              <NA> 0001137761
#> 42              <NA> 0000813900
#> 43              <NA> 0000813900
#> 44              <NA> 0001039803
#> 45              <NA> 0001039803
#> 46              <NA> 0001039803
#> 47              <NA> 0001039803
#> 48              <NA> 0001039803
#> 49              <NA> 0001039803
#> 50              <NA> 0001039803
#> 51              <NA> 0001174610
#> 52              <NA> 0001174610
#> 53              <NA> 0001174610
#> 54              <NA> 0000755222
#> 55              <NA> 0000755222
#> 56              <NA> 0000755222
#> 57              <NA> 0000755222
#> 58              <NA> 0000755222
#> 59              <NA> 0000755222
#> 60              <NA> 0000899148
#> 61              <NA> 0000899148
#> 62              <NA> 0000899148
#> 63              <NA> 0000899148
#> 64              <NA> 0000899148
#> 65              <NA> 0001064046
#> 66              <NA> 0001365151
#> 67              <NA> 0001365151
#> 68              <NA> 0001365151
#> 69              <NA> 0000101507
#> 70              <NA> 0000101507
#> 71              <NA> 0000745903
#> 72              <NA> 0000908695
#> 73              <NA> 0000908695
#> 74              <NA> 0000908695
#> 75              <NA> 0001081400
#> 76              <NA> 0001081400
#> 77              <NA> 0001081400
#> 78              <NA> 0001081400
#> 79              <NA> 0001081400
#>                                                         cik_name
#> 1                        AIM SECTOR FUNDS (INVESCO SECTOR FUNDS)
#> 2                        AIM SECTOR FUNDS (INVESCO SECTOR FUNDS)
#> 3                        AIM SECTOR FUNDS (INVESCO SECTOR FUNDS)
#> 4                        AIM SECTOR FUNDS (INVESCO SECTOR FUNDS)
#> 5                        AIM SECTOR FUNDS (INVESCO SECTOR FUNDS)
#> 6                        AIM SECTOR FUNDS (INVESCO SECTOR FUNDS)
#> 7                        AIM SECTOR FUNDS (INVESCO SECTOR FUNDS)
#> 8                                                 ALPS ETF Trust
#> 9                                  DEUTSCHE DWS SECURITIES TRUST
#> 10                                 DEUTSCHE DWS SECURITIES TRUST
#> 11                                 DEUTSCHE DWS SECURITIES TRUST
#> 12                                 DEUTSCHE DWS SECURITIES TRUST
#> 13                                 DEUTSCHE DWS SECURITIES TRUST
#> 14                                 DEUTSCHE DWS SECURITIES TRUST
#> 15                                                DIREXION FUNDS
#> 16                                                DIREXION FUNDS
#> 17                                                DIREXION FUNDS
#> 18                                                DIREXION FUNDS
#> 19                                      DIREXION INSURANCE TRUST
#> 20                                      DIREXION INSURANCE TRUST
#> 21                                            DOMINION FUNDS INC
#> 22                                            DOMINION FUNDS INC
#> 23                                        DWS MUTUAL FUNDS, INC.
#> 24                                        DWS MUTUAL FUNDS, INC.
#> 25                                        DWS MUTUAL FUNDS, INC.
#> 26                                        DWS MUTUAL FUNDS, INC.
#> 27                                        DWS MUTUAL FUNDS, INC.
#> 28                                        DWS MUTUAL FUNDS, INC.
#> 29                                 EVERGREEN INTERNATIONAL TRUST
#> 30                                 EVERGREEN INTERNATIONAL TRUST
#> 31                                 EVERGREEN INTERNATIONAL TRUST
#> 32                                 EVERGREEN INTERNATIONAL TRUST
#> 33                          FRANKLIN GOLD & PRECIOUS METALS FUND
#> 34                          FRANKLIN GOLD & PRECIOUS METALS FUND
#> 35                          FRANKLIN GOLD & PRECIOUS METALS FUND
#> 36                          FRANKLIN GOLD & PRECIOUS METALS FUND
#> 37                          FRANKLIN GOLD & PRECIOUS METALS FUND
#> 38                                             IndexIQ ETF Trust
#> 39 Invesco Actively Managed Exchange-Traded Commodity Fund Trust
#> 40                         Invesco Exchange-Traded Fund Trust II
#> 41                                    PACIFIC FUNDS SERIES TRUST
#> 42                                           PACIFIC SELECT FUND
#> 43                                           PACIFIC SELECT FUND
#> 44                                                      PROFUNDS
#> 45                                                      PROFUNDS
#> 46                                                      PROFUNDS
#> 47                                                      PROFUNDS
#> 48                                                      PROFUNDS
#> 49                                                      PROFUNDS
#> 50                                                      PROFUNDS
#> 51                                               PROSHARES TRUST
#> 52                                               PROSHARES TRUST
#> 53                                               PROSHARES TRUST
#> 54                             RIVERSOURCE SELECTED SERIES, INC.
#> 55                             RIVERSOURCE SELECTED SERIES, INC.
#> 56                             RIVERSOURCE SELECTED SERIES, INC.
#> 57                             RIVERSOURCE SELECTED SERIES, INC.
#> 58                             RIVERSOURCE SELECTED SERIES, INC.
#> 59                             RIVERSOURCE SELECTED SERIES, INC.
#> 60                                            RYDEX SERIES FUNDS
#> 61                                            RYDEX SERIES FUNDS
#> 62                                            RYDEX SERIES FUNDS
#> 63                                            RYDEX SERIES FUNDS
#> 64                                            RYDEX SERIES FUNDS
#> 65                                          RYDEX VARIABLE TRUST
#> 66                                    Scotia Institutional Funds
#> 67                                    Scotia Institutional Funds
#> 68                                    Scotia Institutional Funds
#> 69                                     US GLOBAL INVESTORS FUNDS
#> 70                                     US GLOBAL INVESTORS FUNDS
#> 71                                         USAA INVESTMENT TRUST
#> 72                                       USAA MUTUAL FUNDS TRUST
#> 73                                       USAA MUTUAL FUNDS TRUST
#> 74                                       USAA MUTUAL FUNDS TRUST
#> 75                                       WELLS FARGO FUNDS TRUST
#> 76                                       WELLS FARGO FUNDS TRUST
#> 77                                       WELLS FARGO FUNDS TRUST
#> 78                                       WELLS FARGO FUNDS TRUST
#> 79                                       WELLS FARGO FUNDS TRUST
#>                                                             cik_filings_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001414040&action=getcompany
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001040587&action=getcompany
#> 16 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001040587&action=getcompany
#> 17 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001040587&action=getcompany
#> 18 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001040587&action=getcompany
#> 19 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001102060&action=getcompany
#> 20 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001102060&action=getcompany
#> 21 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000889392&action=getcompany
#> 22 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000889392&action=getcompany
#> 23 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany
#> 24 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany
#> 25 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany
#> 26 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany
#> 27 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany
#> 28 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany
#> 29 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001046027&action=getcompany
#> 30 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001046027&action=getcompany
#> 31 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001046027&action=getcompany
#> 32 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001046027&action=getcompany
#> 33 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000083293&action=getcompany
#> 34 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000083293&action=getcompany
#> 35 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000083293&action=getcompany
#> 36 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000083293&action=getcompany
#> 37 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000083293&action=getcompany
#> 38 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001415995&action=getcompany
#> 39 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001595386&action=getcompany
#> 40 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001378872&action=getcompany
#> 41 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001137761&action=getcompany
#> 42 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000813900&action=getcompany
#> 43 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000813900&action=getcompany
#> 44 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany
#> 45 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany
#> 46 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany
#> 47 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany
#> 48 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany
#> 49 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany
#> 50 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany
#> 51 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001174610&action=getcompany
#> 52 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001174610&action=getcompany
#> 53 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001174610&action=getcompany
#> 54 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany
#> 55 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany
#> 56 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany
#> 57 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany
#> 58 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany
#> 59 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany
#> 60 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000899148&action=getcompany
#> 61 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000899148&action=getcompany
#> 62 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000899148&action=getcompany
#> 63 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000899148&action=getcompany
#> 64 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000899148&action=getcompany
#> 65 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001064046&action=getcompany
#> 66 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001365151&action=getcompany
#> 67 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001365151&action=getcompany
#> 68 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001365151&action=getcompany
#> 69 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000101507&action=getcompany
#> 70 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000101507&action=getcompany
#> 71 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000745903&action=getcompany
#> 72 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000908695&action=getcompany
#> 73 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000908695&action=getcompany
#> 74 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000908695&action=getcompany
#> 75 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001081400&action=getcompany
#> 76 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001081400&action=getcompany
#> 77 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001081400&action=getcompany
#> 78 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001081400&action=getcompany
#> 79 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001081400&action=getcompany
#>                                                                          cik_funds_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany&scd=series
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany&scd=series
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany&scd=series
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany&scd=series
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany&scd=series
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany&scd=series
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000725781&action=getcompany&scd=series
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001414040&action=getcompany&scd=series
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany&scd=series
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany&scd=series
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany&scd=series
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany&scd=series
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany&scd=series
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000088048&action=getcompany&scd=series
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001040587&action=getcompany&scd=series
#> 16 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001040587&action=getcompany&scd=series
#> 17 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001040587&action=getcompany&scd=series
#> 18 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001040587&action=getcompany&scd=series
#> 19 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001102060&action=getcompany&scd=series
#> 20 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001102060&action=getcompany&scd=series
#> 21 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000889392&action=getcompany&scd=series
#> 22 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000889392&action=getcompany&scd=series
#> 23 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany&scd=series
#> 24 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany&scd=series
#> 25 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany&scd=series
#> 26 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany&scd=series
#> 27 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany&scd=series
#> 28 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000832946&action=getcompany&scd=series
#> 29 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001046027&action=getcompany&scd=series
#> 30 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001046027&action=getcompany&scd=series
#> 31 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001046027&action=getcompany&scd=series
#> 32 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001046027&action=getcompany&scd=series
#> 33 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000083293&action=getcompany&scd=series
#> 34 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000083293&action=getcompany&scd=series
#> 35 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000083293&action=getcompany&scd=series
#> 36 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000083293&action=getcompany&scd=series
#> 37 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000083293&action=getcompany&scd=series
#> 38 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001415995&action=getcompany&scd=series
#> 39 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001595386&action=getcompany&scd=series
#> 40 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001378872&action=getcompany&scd=series
#> 41 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001137761&action=getcompany&scd=series
#> 42 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000813900&action=getcompany&scd=series
#> 43 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000813900&action=getcompany&scd=series
#> 44 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany&scd=series
#> 45 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany&scd=series
#> 46 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany&scd=series
#> 47 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany&scd=series
#> 48 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany&scd=series
#> 49 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany&scd=series
#> 50 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001039803&action=getcompany&scd=series
#> 51 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001174610&action=getcompany&scd=series
#> 52 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001174610&action=getcompany&scd=series
#> 53 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001174610&action=getcompany&scd=series
#> 54 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany&scd=series
#> 55 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany&scd=series
#> 56 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany&scd=series
#> 57 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany&scd=series
#> 58 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany&scd=series
#> 59 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000755222&action=getcompany&scd=series
#> 60 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000899148&action=getcompany&scd=series
#> 61 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000899148&action=getcompany&scd=series
#> 62 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000899148&action=getcompany&scd=series
#> 63 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000899148&action=getcompany&scd=series
#> 64 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000899148&action=getcompany&scd=series
#> 65 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001064046&action=getcompany&scd=series
#> 66 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001365151&action=getcompany&scd=series
#> 67 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001365151&action=getcompany&scd=series
#> 68 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001365151&action=getcompany&scd=series
#> 69 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000101507&action=getcompany&scd=series
#> 70 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000101507&action=getcompany&scd=series
#> 71 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000745903&action=getcompany&scd=series
#> 72 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000908695&action=getcompany&scd=series
#> 73 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000908695&action=getcompany&scd=series
#> 74 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000908695&action=getcompany&scd=series
#> 75 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001081400&action=getcompany&scd=series
#> 76 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001081400&action=getcompany&scd=series
#> 77 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001081400&action=getcompany&scd=series
#> 78 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001081400&action=getcompany&scd=series
#> 79 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0001081400&action=getcompany&scd=series

variable_insurance_fast_search("VMFVX")

#>     class_id
#> 1 C000092041
#>                                                                      class_filings_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?CIK=C000092041&action=getcompany&scd=filings
#>             class_name class_ticker  series_id
#> 1 Institutional Shares        VMFVX S000030015
#>                                                                     series_filings_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?CIK=S000030015&action=getcompany&scd=filings
#>                                 series_name series_funds_href        cik
#> 1 Vanguard S&P Mid-Cap 400 Value Index Fund              <NA> 0000891190
#>                 cik_name
#> 1 VANGUARD ADMIRAL FUNDS
#>                                                            cik_filings_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000891190&action=getcompany
#>                                                                         cik_funds_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?CIK=0000891190&action=getcompany&scd=series