SEC Company Details

For a given company, either by ticker, CIK, or pre-fetched page, we extract 2 sets of information:

Company Information

Filing date, accepted date, etc.

Filings

Companies included in the filing

company_details(
  x,
  ownership = FALSE,
  type = "",
  before = "",
  count = 40,
  page = 1
)

Arguments

x
either a stock ticker, CIK number, or XML document for a company page

ownership
boolean for inclusion of company change filings

type
Type of filing to fetch. NOTE: due to the way the SEC EDGAR system works, it is actually is a 'starts-with' search, so for instance specifying 'type = "10-K" will return "10-K/A" and "10-K405" filings as well. To ensure you only get the type you want, best practice would be to filter the results.

before
yyyymmdd format of latest filing to fetch

count
Number of filings to fetch per page. Valid options are 10, 20, 40, 80, or 100. Other values will result in the closest count.

page
Which page of results to return.

Value

A list with the following components

information

data.frame as returned by company_information

filings

data.frame as returned by company_filings

Examples

# \donttest{
company_details("AAPL", before = "20170810")

#> $information
#>         name        cik fiscal_year_end
#> 1 Apple Inc. 0000320193            0925
#>                                                                                       company_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000320193&owner=exclude&count=40
#>    sic      sic_description state_location state_incorporation mailing_city
#> 1 3571 ELECTRONIC COMPUTERS             CA                  CA    CUPERTINO
#>   mailing_state mailing_zip     mailing_street mailing_street2 business_city
#> 1            CA       95014 ONE APPLE PARK WAY            <NA>     CUPERTINO
#>   business_state business_zip    business_street business_street2
#> 1             CA        95014 ONE APPLE PARK WAY             <NA>
#>   business_phone
#> 1 (408) 996-1010
#> 
#> $filings
#>    accession_number act file_number filing_date accepted_date
#> 1              <NA>  34   001-36743  2017-08-02    2017-08-02
#> 2              <NA>  34   001-36743  2017-08-01    2017-08-01
#> 3              <NA>  34   001-36743  2017-06-20    2017-06-20
#> 4              <NA>  33  333-210983  2017-06-14    2017-06-14
#> 5              <NA>  34  333-210983  2017-06-14    2017-06-13
#> 6              <NA>  33  333-210983  2017-06-13    2017-06-13
#> 7              <NA>  34   001-36743  2017-05-25    2017-05-25
#> 8              <NA>  34   001-36743  2017-05-24    2017-05-24
#> 9              <NA>  34   001-36743  2017-05-24    2017-05-24
#> 10             <NA>  33  333-210983  2017-05-18    2017-05-18
#> 11             <NA>  34  333-210983  2017-05-17    2017-05-17
#> 12             <NA>  33  333-210983  2017-05-17    2017-05-17
#> 13             <NA>  34   001-36743  2017-05-11    2017-05-11
#> 14             <NA>  33  333-210983  2017-05-05    2017-05-05
#> 15             <NA>  34   001-36743  2017-05-05    2017-05-05
#> 16             <NA>  34  333-210983  2017-05-05    2017-05-04
#> 17             <NA>  33  333-210983  2017-05-04    2017-05-04
#> 18             <NA>  34   001-36743  2017-05-03    2017-05-03
#> 19             <NA>  34   001-36743  2017-05-02    2017-05-02
#> 20             <NA>  34   001-36743  2017-03-03    2017-03-03
#> 21             <NA>  34   001-36743  2017-03-01    2017-03-01
#> 22             <NA>  33  333-210983  2017-02-15    2017-02-15
#> 23             <NA>  34  333-210983  2017-02-14    2017-02-14
#> 24             <NA>  33  333-210983  2017-02-14    2017-02-13
#> 25             <NA>  34   001-36743  2017-02-09    2017-02-09
#> 26             <NA>  34   005-33632  2017-02-09    2017-02-09
#> 27             <NA>  33  333-210983  2017-02-03    2017-02-03
#> 28             <NA>  34  333-210983  2017-02-03    2017-02-02
#> 29             <NA>  33  333-210983  2017-02-02    2017-02-02
#> 30             <NA>  34   001-36743  2017-02-01    2017-02-01
#> 31             <NA>  34   001-36743  2017-01-31    2017-01-31
#> 32             <NA>  34   001-36743  2017-01-31    2017-01-31
#> 33             <NA>  34   005-33632  2017-01-19    2017-01-19
#> 34             <NA>  34   001-36743  2017-01-06    2017-01-06
#> 35             <NA>  34   001-36743  2017-01-06    2017-01-06
#> 36             <NA>  34   001-36743  2016-12-15    2016-12-15
#> 37             <NA>  34   001-36743  2016-12-05    2017-01-03
#> 38             <NA>  34   001-36743  2016-10-27    2017-02-16
#> 39             <NA>  34   001-36743  2016-10-26    2016-11-01
#> 40             <NA>  34   001-36743  2016-10-26    2016-10-26
#>                                                                                                href
#> 1  https://www.sec.gov/Archives/edgar/data/320193/000032019317000009/0000320193-17-000009-index.htm
#> 2  https://www.sec.gov/Archives/edgar/data/320193/000032019317000006/0000320193-17-000006-index.htm
#> 3  https://www.sec.gov/Archives/edgar/data/320193/000119312517208226/0001193125-17-208226-index.htm
#> 4  https://www.sec.gov/Archives/edgar/data/320193/000119312517203548/0001193125-17-203548-index.htm
#> 5  https://www.sec.gov/Archives/edgar/data/320193/000119312517202772/0001193125-17-202772-index.htm
#> 6  https://www.sec.gov/Archives/edgar/data/320193/000119312517201958/0001193125-17-201958-index.htm
#> 7  https://www.sec.gov/Archives/edgar/data/320193/999999999717006550/9999999997-17-006550-index.htm
#> 8  https://www.sec.gov/Archives/edgar/data/320193/000119312517181871/0001193125-17-181871-index.htm
#> 9  https://www.sec.gov/Archives/edgar/data/320193/000119312517181867/0001193125-17-181867-index.htm
#> 10 https://www.sec.gov/Archives/edgar/data/320193/000119312517175330/0001193125-17-175330-index.htm
#> 11 https://www.sec.gov/Archives/edgar/data/320193/000119312517173678/0001193125-17-173678-index.htm
#> 12 https://www.sec.gov/Archives/edgar/data/320193/000119312517172693/0001193125-17-172693-index.htm
#> 13 https://www.sec.gov/Archives/edgar/data/320193/000119312517167400/0001193125-17-167400-index.htm
#> 14 https://www.sec.gov/Archives/edgar/data/320193/000119312517160948/0001193125-17-160948-index.htm
#> 15 https://www.sec.gov/Archives/edgar/data/320193/000119312517159397/0001193125-17-159397-index.htm
#> 16 https://www.sec.gov/Archives/edgar/data/320193/000119312517159324/0001193125-17-159324-index.htm
#> 17 https://www.sec.gov/Archives/edgar/data/320193/000119312517157385/0001193125-17-157385-index.htm
#> 18 https://www.sec.gov/Archives/edgar/data/320193/000162828017004790/0001628280-17-004790-index.htm
#> 19 https://www.sec.gov/Archives/edgar/data/320193/000162828017004663/0001628280-17-004663-index.htm
#> 20 https://www.sec.gov/Archives/edgar/data/320193/000119312517069853/0001193125-17-069853-index.htm
#> 21 https://www.sec.gov/Archives/edgar/data/320193/000119312517064019/0001193125-17-064019-index.htm
#> 22 https://www.sec.gov/Archives/edgar/data/320193/000119312517044564/0001193125-17-044564-index.htm
#> 23 https://www.sec.gov/Archives/edgar/data/320193/000119312517042083/0001193125-17-042083-index.htm
#> 24 https://www.sec.gov/Archives/edgar/data/320193/000119312517041233/0001193125-17-041233-index.htm
#> 25 https://www.sec.gov/Archives/edgar/data/320193/000119312517036283/0001193125-17-036283-index.htm
#> 26 https://www.sec.gov/Archives/edgar/data/320193/000093247117000777/0000932471-17-000777-index.htm
#> 27 https://www.sec.gov/Archives/edgar/data/320193/000119312517030832/0001193125-17-030832-index.htm
#> 28 https://www.sec.gov/Archives/edgar/data/320193/000119312517029548/0001193125-17-029548-index.htm
#> 29 https://www.sec.gov/Archives/edgar/data/320193/000119312517028315/0001193125-17-028315-index.htm
#> 30 https://www.sec.gov/Archives/edgar/data/320193/000162828017000717/0001628280-17-000717-index.htm
#> 31 https://www.sec.gov/Archives/edgar/data/320193/000162828017000663/0001628280-17-000663-index.htm
#> 32 https://www.sec.gov/Archives/edgar/data/320193/000121465917000547/0001214659-17-000547-index.htm
#> 33 https://www.sec.gov/Archives/edgar/data/320193/000021545717001275/0000215457-17-001275-index.htm
#> 34 https://www.sec.gov/Archives/edgar/data/320193/000119312517003764/0001193125-17-003764-index.htm
#> 35 https://www.sec.gov/Archives/edgar/data/320193/000119312517003753/0001193125-17-003753-index.htm
#> 36 https://www.sec.gov/Archives/edgar/data/320193/000162828016022047/0001628280-16-022047-index.htm
#> 37 https://www.sec.gov/Archives/edgar/data/320193/999999999717000002/9999999997-17-000002-index.htm
#> 38 https://www.sec.gov/Archives/edgar/data/320193/999999999717000809/9999999997-17-000809-index.htm
#> 39 https://www.sec.gov/Archives/edgar/data/320193/999999999716026607/9999999997-16-026607-index.htm
#> 40 https://www.sec.gov/Archives/edgar/data/320193/000162828016020309/0001628280-16-020309-index.htm
#>        type film_number
#> 1      10-Q   171000359
#> 2       8-K    17997093
#> 3       8-K    17921171
#> 4     424B2    17911569
#> 5       FWP    17910189
#> 6     424B2    17907933
#> 7   CERTNYS    17010416
#> 8    8-A12B    17867595
#> 9       8-K    17867578
#> 10    424B2    17855370
#> 11      FWP    17852308
#> 12    424B2    17850239
#> 13      8-K    17835158
#> 14    424B2    17819168
#> 15       SD    17816013
#> 16      FWP    17815909
#> 17    424B2    17811979
#> 18     10-Q    17809653
#> 19      8-K    17805508
#> 20      8-K    17664239
#> 21      8-K    17651574
#> 22    424B2    17613002
#> 23      FWP    17602904
#> 24    424B2    17601961
#> 25      8-K    17588362
#> 26 SC 13G/A    17586504
#> 27    424B2    17572868
#> 28      FWP    17570099
#> 29    424B2    17566778
#> 30     10-Q    17565317
#> 31      8-K    17561519
#> 32  PX14A6G    17559872
#> 33 SC 13G/A    17536781
#> 34  DEFA14A    17512909
#> 35  DEF 14A    17512896
#> 36      8-K   162054250
#> 37   NO ACT    16005019
#> 38   NO ACT    20160005
#> 39   NO ACT    16005137
#> 40     10-K   161953070
#>                                                                        form_name
#> 1                                        Quarterly report [Sections 13 or 15(d)]
#> 2                                                                 Current report
#> 3                                                                 Current report
#> 4                                                    Prospectus [Rule 424(b)(2)]
#> 5         Filing under Securities Act Rules 163/433 of free writing prospectuses
#> 6                                                    Prospectus [Rule 424(b)(2)]
#> 7  Certification by the New York Stock Exchange approving securities for listing
#> 8                                     Registration of securities [Section 12(b)]
#> 9                                                                 Current report
#> 10                                                   Prospectus [Rule 424(b)(2)]
#> 11        Filing under Securities Act Rules 163/433 of free writing prospectuses
#> 12                                                   Prospectus [Rule 424(b)(2)]
#> 13                                                                Current report
#> 14                                                   Prospectus [Rule 424(b)(2)]
#> 15                                                 Specialized disclosure report
#> 16        Filing under Securities Act Rules 163/433 of free writing prospectuses
#> 17                                                   Prospectus [Rule 424(b)(2)]
#> 18                                       Quarterly report [Sections 13 or 15(d)]
#> 19                                                                Current report
#> 20                                                                Current report
#> 21                                                                Current report
#> 22                                                   Prospectus [Rule 424(b)(2)]
#> 23        Filing under Securities Act Rules 163/433 of free writing prospectuses
#> 24                                                   Prospectus [Rule 424(b)(2)]
#> 25                                                                Current report
#> 26               Statement of acquisition of beneficial ownership by individuals
#> 27                                                   Prospectus [Rule 424(b)(2)]
#> 28        Filing under Securities Act Rules 163/433 of free writing prospectuses
#> 29                                                   Prospectus [Rule 424(b)(2)]
#> 30                                       Quarterly report [Sections 13 or 15(d)]
#> 31                                                                Current report
#> 32                     Notice of exempt solicitation submitted by non-management
#> 33               Statement of acquisition of beneficial ownership by individuals
#> 34  Additional definitive proxy soliciting materials and Rule 14(a)(12) material
#> 35                                             Other definitive proxy statements
#> 36                                                                Current report
#> 37                                                              No Action Letter
#> 38                                                              No Action Letter
#> 39                                                              No Action Letter
#> 40                        Annual report [Section 13 and 15(d), not S-K Item 405]
#>            description   size
#> 1                 <NA>  10 MB
#> 2  items 2.02 and 9.01 406 KB
#> 3  items 8.01 and 9.01 217 KB
#> 4                 <NA> 426 KB
#> 5                 <NA>  16 KB
#> 6                 <NA> 422 KB
#> 7                 <NA>   1 KB
#> 8                 <NA>  17 KB
#> 9  items 8.01 and 9.01 309 KB
#> 10                <NA> 469 KB
#> 11                <NA>  30 KB
#> 12                <NA> 462 KB
#> 13 items 8.01 and 9.01 514 KB
#> 14                <NA> 460 KB
#> 15                <NA> 835 KB
#> 16                <NA>  54 KB
#> 17                <NA> 428 KB
#> 18                <NA>   9 MB
#> 19 items 2.02 and 9.01 430 KB
#> 20 items 8.01 and 9.01 234 KB
#> 21           item 5.07  49 KB
#> 22                <NA> 476 KB
#> 23                <NA>  13 KB
#> 24                <NA> 476 KB
#> 25 items 8.01 and 9.01 675 KB
#> 26                <NA>  37 KB
#> 27                <NA> 544 KB
#> 28                <NA>  96 KB
#> 29                <NA> 498 KB
#> 30                <NA>   8 MB
#> 31 items 2.02 and 9.01 390 KB
#> 32                <NA>  28 KB
#> 33                <NA>  13 KB
#> 34                <NA> 674 KB
#> 35                <NA>   5 MB
#> 36 items 5.03 and 9.01 248 KB
#> 37                <NA>   1 KB
#> 38                <NA>   1 KB
#> 39                <NA>   1 KB
#> 40                <NA>  13 MB
#> 

# }