SEC Company Search

Provides access to the SEC Company Name Search from here using a company's formal name rather than its common name.

company_search(
  x,
  match = "start",
  file_number = FALSE,
  state = "",
  country = "",
  sic = "",
  ownership = FALSE,
  type = "",
  count = 40,
  page = 1
)

Arguments

x
Name of company to search or file number

match
(default: 'start') Either 'start' or 'contains' for where in the company name to search

file_number
(default: FALSE) if set to TRUE, x is treated as a file number

state
(default: '') Limit to a specific state of registration using 2-letter state abbreviations. Special values:

X1

The United States

A0

Alberta, Canada

A1

British Columbia, Canada

A2

Manitoba, Canada

A3

New Brunswick, Canada

A4

Newfoundland, Canada

A5

Nova Scotia, Canada

A6

Ontario, Canada

A7

Prince Edward Island, Canada

A8

Quebec, Canada

A9

Saskatchewan, Canada

B0

Yukon, Canada

country
2-character country code. The mapping is non-obvious, so unfortunately the best way to find it is to examine the company search page.

sic
SIC Code

ownership
boolean for inclusion of company change filings

type
Limit to companies with a given filing type - e.g. 'N-PX'

count
Number of filings to fetch per page. Valid options are 10, 20, 40, 80, or 100. Other values will result in the closest count.

page
Which page of results to return.

Value

A dataframe of companies

cik

company_href

name

location

location_href

formerly

sic

sic_description

sic_href

Details

Note On 'Fast Search' -- The SEC Company Search page also includes a 'Fast Search' function to "search" by CIK or Stock Ticker. This doesn't actually search, but rather goes directly to the company details page if found. If you have a company's CIK or Ticker, use the company_information, company_filings, or company_details functions.

Examples

# \donttest{
company_search("Intel")

#>                                                   name        cik
#> 1                          INTEL CAPITAL (CAYMAN) CORP 0001165733
#> 2                                   INTEL CAPITAL CORP 0001129302
#> 3                                           INTEL CORP 0000050863
#> 4                                     INTEL FOUNDATION 0001422503
#> 5                                  INTEL OVERSEAS CORP 0000768545
#> 6                        Intel Technology Asia Pte Ltd 0001504543
#> 7                                     IntelAccount Inc 0001330442
#> 8                                     IntelAegis, Inc. 0001795529
#> 9                                       INTELAGARD INC 0001007001
#> 10                                       Intelage Corp 0001761317
#> 11                                        NEXGRID, LLC 0001495887
#> 12                                   IntelaSight, Inc. 0001318139
#> 13                                     INTELATRADE.COM 0001109248
#> 14                                    IntelaTrak, Inc. 0001617224
#> 15                               Uniontown Energy Inc. 0001310482
#> 16                         ICG HOLDINGS CANADA CO /CO/ 0000786343
#> 17                                    ICG HOLDINGS INC 0001001131
#> 18                                         Intele Inc. 0001535797
#> 19                                  INTELECOM CORP INC 0001141741
#> 20                        Intelecom Data Systems, Inc. 0001522997
#> 21                                    INTELECOM INC/IN 0001374166
#> 22                             INTELECOM SOLUTIONS INC 0001169428
#> 23                           TERAFORCE TECHNOLOGY CORP 0000316672
#> 24                                IntElect Medical Inc 0001402706
#> 25                      Intelective Communications Inc 0001416837
#> 26                                     INTELEFILM CORP 0000882160
#> 27                    Intelemedia Communications, Inc. 0001342517
#> 28                                      INTELEPEER INC 0001423664
#> 29                                       INTELEPIX LLC 0001452405
#> 30                     INTELETECH  CORP OF AMERICA INC 0001178710
#> 31                        IntelGenx Technologies Corp. 0001098880
#> 32                                        INTELIBS INC 0001437567
#> 33                     INTELICARE HEALTH SERVICES, LLC 0001591726
#> 34                                      INTELICIS CORP 0001354587
#> 35                           INTELICLOUD HOLDINGS INC. 0001534353
#> 36                        INTELICLOUD TECHNOLOGY, INC. 0001507834
#> 37 Intelicoat Technologies Image Products Matthews LLC 0001501758
#> 38                                  INTRATEL GROUP LTD 0000931758
#> 39                        INTELIDATA TECHNOLOGIES CORP 0001021810
#> 40                                  INTELILABS COM INC 0001085257
#>    fiscal_year_end company_href  sic state_location state_incorporation
#> 1             1231         <NA> <NA>             E9                  E9
#> 2             1231         <NA> <NA>             DE                  DE
#> 3             1228         <NA> 3674             CA                  DE
#> 4             1231         <NA> <NA>             OR                  OR
#> 5             1231         <NA> 3674             CA                  CA
#> 6             1231         <NA> <NA>             U0                  U0
#> 7             <NA>         <NA> <NA>             MN                <NA>
#> 8             1231         <NA> <NA>             WI                  WI
#> 9             <NA>         <NA> <NA>             CO                <NA>
#> 10            1231         <NA> <NA>             WA                  DE
#> 11            1231         <NA> <NA>             VA                  VA
#> 12            1231         <NA> <NA>             AZ                  WA
#> 13            <NA>         <NA> <NA>             CT                  CT
#> 14            1231         <NA> <NA>             NH                  DE
#> 15            1031         <NA> 1311             A1                  NV
#> 16            1231         <NA> 4813             CO                  DE
#> 17            1231         <NA> 4813             CO                  CO
#> 18            1231         <NA> <NA>             NJ                  DE
#> 19            <NA>         <NA> <NA>           <NA>                <NA>
#> 20            1231         <NA> <NA>             MA                  RI
#> 21            <NA>         <NA> <NA>             IN                <NA>
#> 22            <NA>         <NA> <NA>             NY                <NA>
#> 23            1231         <NA> 3576             TX                  DE
#> 24            <NA>         <NA> <NA>             OH                  de
#> 25            <NA>         <NA> <NA>             CA                  CA
#> 26            1231         <NA> 7812             MN                  MN
#> 27            1231         <NA> <NA>             TX                  TX
#> 28            <NA>         <NA> 7385             CA                  DE
#> 29            <NA>         <NA> <NA>             NY                  NY
#> 30            <NA>         <NA> <NA>             FL                <NA>
#> 31            1231         <NA> 2834             A8                  DE
#> 32            <NA>         <NA> <NA>             NJ                  DE
#> 33            1231         <NA> <NA>             FL                  FL
#> 34            <NA>         <NA> <NA>             CA                <NA>
#> 35            1231         <NA> <NA>             CA                  DE
#> 36            1031         <NA> <NA>             CA                  NV
#> 37            1231         <NA> <NA>             SC                  DE
#> 38            1231         <NA> 4813             FL                  DE
#> 39            1231         <NA> 7373             VA                  DE
#> 40            0630         <NA> 7389             FL                  DE
#>    formerly            mailing_city mailing_state mailing_zip
#> 1      <NA> GEORGETOWN GRAND CAYMAN            E9    KY1-1102
#> 2      <NA>              WILMINGTON            DE       19801
#> 3      <NA>             SANTA CLARA            CA       95054
#> 4      <NA>               HILLSBORO            OR       97124
#> 5      <NA>             SANTA CLARA            CA       95052
#> 6      <NA>            PWC BUILDING            U0      048424
#> 7      <NA>             MINNEAPOLIS            MN       55455
#> 8      <NA>               MILWAUKEE            WI       53203
#> 9      <NA>                    <NA>          <NA>        <NA>
#> 10     <NA>                BELLEVUE            WA       98004
#> 11     <NA>            LOCUST GROVE            VA       22508
#> 12     <NA>                    MESA            AZ       85210
#> 13     <NA>                 NORWALK            CT       06855
#> 14     <NA>                HENNIKER            NH       03242
#> 15     <NA>               VANCOUVER            A1     V6C 3E2
#> 16     <NA>               ENGLEWOOD            CO       80112
#> 17     <NA>               ENGLEWOOD            CO       80112
#> 18     <NA>            ROBBINSVILLE            NJ       08691
#> 19     <NA>                    <NA>          <NA>        <NA>
#> 20     <NA>              BURLINGTON            MA       01803
#> 21     <NA>            INDIANOPOLIS            IN       46201
#> 22     <NA>                    <NA>          <NA>        <NA>
#> 23     <NA>              RICHARDSON            TX       75081
#> 24     <NA>               Cleveland            OH       44114
#> 25     <NA>             Penn Valley            CA       95946
#> 26     <NA>             MINNEAPOLIS            MN       55416
#> 27     <NA>                   PLANO            TX       75075
#> 28     <NA>               SAN MATEO            CA       94403
#> 29     <NA>               GLEN COVE            NY       11542
#> 30     <NA>                    <NA>          <NA>        <NA>
#> 31     <NA>           SAINT LAURENT            A8     H4S 1Y2
#> 32     <NA>                  SUMMIT            NJ       07901
#> 33     <NA>           BOYNTON BEACH            FL       33472
#> 34     <NA>             SANTA CLARA            CA       95054
#> 35     <NA>           NEWPORT BEACH            CA       92660
#> 36     <NA>           NEWPORT BEACH            CA       92660
#> 37     <NA>             SPARTANBURG            SC       29302
#> 38     <NA>                METAIRIE            LA       70005
#> 39     <NA>                  RESTON            VA       20191
#> 40     <NA>                   MIAMI            FL       33131
#>                           mailing_street1
#> 1         STERLING TRUST (CAYMAN) LIMITED
#> 2           THE CORPORATION TRUST COMPANY
#> 3               2200 MISSION COLLEGE BLVD
#> 4  5200 NE ELAM YOUNG PARKWAY, MS AG6-601
#> 5                                    <NA>
#> 6                  8 CROSS STREET # 11-00
#> 7                         7701 FRANCE AVE
#> 8             840 N. OLD WORLD 3RD STREET
#> 9                                    <NA>
#> 10                        10400 NE 4TH ST
#> 11       4444 GERMANNA HIGHWAY, SUITE 330
#> 12              1201 SOUTH ALMA SCHOOL RD
#> 13                      14 DUCK POND ROAD
#> 14              388 WESTERN AVE., UNIT A3
#> 15            999 CANADA PLACE, SUITE 404
#> 16               161 INVERNESS DRIVE WEST
#> 17               161 INVERNESS DRIVE WEST
#> 18                         6 THOMPSON WAY
#> 19             P O BOX 63 1 RUE DE BUGNON
#> 20                          5 WALL STREET
#> 21                   2421 E WASHINGTON ST
#> 22                                   <NA>
#> 23                     1240 CAMPBELL ROAD
#> 24                         1300 E 9th Str
#> 25             11348 Pleasant Valley Road
#> 26                    5501 EXCELSIOR BLVD
#> 27         1255 W. 15TH STREET, SUITE 700
#> 28                      2855 CAMPUS DRIVE
#> 29                         70 GLEN STREET
#> 30                                   <NA>
#> 31                            6420 ABRAMS
#> 32                     246 BLACKBURN ROAD
#> 33                          8198 JOG ROAD
#> 34               4633 OLD IRONSIDES DRIVE
#> 35                      4600 CAMPUS DRIVE
#> 36                        895 DOVE STREET
#> 37                    3070 SOUTHPORT ROAD
#> 38                           1109 ANDREWS
#> 39             11600 SUNRISE VALLEY DRIVE
#> 40           701 BRICKELL AVENUE STE 3120
#>                            mailing_street2           business_city
#> 1  WHITEHALL HOUSE 238 NORTH CHURCH STREET GEORGETOWN GRAND CAYMAN
#> 2                       1209 ORANGE STREET              WILMINGTON
#> 3                                RNB-4-151             SANTA CLARA
#> 4                                     <NA>               HILLSBORO
#> 5                2200 MISSION COLLEGE BLVD             SANTA CLARA
#> 6                                     <NA>            PWC BUILDING
#> 7                                SUITE 200             MINNEAPOLIS
#> 8                                SUITE 500               MILWAUKEE
#> 9                                     <NA>              BROOMFIELD
#> 10                                    <NA>                BELLEVUE
#> 11                                    <NA>            LOCUST GROVE
#> 12                              SUITE 4450                    MESA
#> 13                                    <NA>                 NORWALK
#> 14                                    <NA>                HENNIKER
#> 15                                    <NA>               VANCOUVER
#> 16                             PO BOX 6742               ENGLEWOOD
#> 17                              SUITE 1610               ENGLEWOOD
#> 18                                    <NA>            ROBBINSVILLE
#> 19             CRANS VAAD SWITZERLAND 1299                    <NA>
#> 20                                    <NA>              BURLINGTON
#> 21                                    <NA>            INDIANOPOLIS
#> 22                                    <NA>               PLAINVIEW
#> 23                                    <NA>              RICHARDSON
#> 24                                    <NA>               Cleveland
#> 25                               Suite 251             Penn Valley
#> 26                                    <NA>             MINNEAPOLIS
#> 27                                    <NA>                   PLANO
#> 28                               SUITE 200               SAN MATEO
#> 29                               SUITE 360               GLEN COVE
#> 30                                    <NA>         FORT LAUDERDALE
#> 31                                    <NA>           SAINT LAURENT
#> 32                                    <NA>                  SUMMIT
#> 33                               SUITE 201           BOYNTON BEACH
#> 34                               SUITE 130             SANTA CLARA
#> 35                               SUITE 106           NEWPORT BEACH
#> 36                               SUITE 300           NEWPORT BEACH
#> 37                                    <NA>             SPARTANBURG
#> 38                                    <NA>              CLEARWATER
#> 39                               SUITE 100                  RESTON
#> 40                                    <NA>                   MIAMI
#>    business_state business_zip                       business_street1
#> 1              E9     KY1-1102        STERLING TRUST (CAYMAN) LIMITED
#> 2              DE        19801          THE CORPORATION TRUST COMPANY
#> 3              CA        95054              2200 MISSION COLLEGE BLVD
#> 4              OR        97124 5200 NE ELAM YOUNG PARKWAY, MS AG6-601
#> 5              CA        95051              2200 MISSION COLLEGE BLVD
#> 6              U0       048424                 8 CROSS STREET # 11-00
#> 7              MN        55455                        7701 FRANCE AVE
#> 8              WI        53203            840 N. OLD WORLD 3RD STREET
#> 9              CO        80020                         590 BURBANK ST
#> 10             WA        98004                        10400 NE 4TH ST
#> 11             VA        22508       4444 GERMANNA HIGHWAY, SUITE 330
#> 12             AZ        85210              1201 SOUTH ALMA SCHOOL RD
#> 13             CT        06855                      14 DUCK POND ROAD
#> 14             NH        03242              388 WESTERN AVE., UNIT A3
#> 15             A1      V6C 3E2            999 CANADA PLACE, SUITE 404
#> 16             CO        80112               161 INVERNESS DRIVE WEST
#> 17             CO        80112               161 INVERNESS DRIVE WEST
#> 18             NJ        08691                         6 THOMPSON WAY
#> 19           <NA>         <NA>             P O BOX 63 1 RUE DE BUGNON
#> 20             MA        01803                          5 WALL STREET
#> 21             IN        46201                   2421 E WASHINGTON ST
#> 22             NY        11803                  1600 OLD COUNTRY ROAD
#> 23             TX        75081                     1240 CAMPBELL ROAD
#> 24             OH        44114                         1300 E 9th Str
#> 25             CA        95946             11348 Pleasant Valley Road
#> 26             MN        55416                    5501 EXCELSIOR BLVD
#> 27             TX        75075         1255 W. 15TH STREET, SUITE 700
#> 28             CA        94403                      2855 CAMPUS DRIVE
#> 29             NY        11542                         70 GLEN STREET
#> 30             FL        33301                  ONE BROWARD BOULEVARD
#> 31             A8      H4S 1Y2                            6420 ABRAMS
#> 32             NJ        07901                     246 BLACKBURN ROAD
#> 33             FL        33472                          8198 JOG ROAD
#> 34             CA        95054               4633 OLD IRONSIDES DRIVE
#> 35             CA        92660                      4600 CAMPUS DRIVE
#> 36             CA        92660                        895 DOVE STREET
#> 37             SC        29302                    3070 SOUTHPORT ROAD
#> 38             FL        34621                       28050 US HWY 19N
#> 39             VA        20191             11600 SUNRISE VALLEY DRIVE
#> 40             FL        33131                       701 BRICKELL AVE
#>                           business_street2     business_phone
#> 1  WHITEHALL HOUSE 238 NORTH CHURCH STREET         4087658080
#> 2                       1209 ORANGE STREET         4087658080
#> 3                                RNB-4-151         4087658080
#> 4                                     <NA>     (408) 765-8080
#> 5                                     <NA>         4089878080
#> 6                                     <NA>        65 62131000
#> 7                                SUITE 200       612-920-1747
#> 8                                SUITE 500       312-218-5363
#> 9                                SUITE 220       303-309-6309
#> 10                                    <NA>       206.201.4389
#> 11                                    <NA> 888 556 0911 X1002
#> 12                              SUITE 4450       480-307-8700
#> 13                                    <NA>       203 838-8470
#> 14                                    <NA>     (339) 933-2793
#> 15                                    <NA>     (702) 583-5533
#> 16                            P O BOX 6742         3034145431
#> 17                            P O BOX 6742         3034145643
#> 18                                    <NA>     (609) 802-2273
#> 19             CRANS VAAD SWITZERLAND 1299               <NA>
#> 20                                    <NA>     (404) 815-0770
#> 21                                    <NA>       317-639-6545
#> 22                               PLAINVIEW       516-847-7700
#> 23                                    <NA>         9723672100
#> 24                                    <NA>       216-916-6924
#> 25                               Suite 251     (800) 804-5007
#> 26                                    <NA>         6129258840
#> 27                                    <NA>       972 994 0700
#> 28                               SUITE 200       650-235-8509
#> 29                               SUITE 360  516-801-6460 X 11
#> 30                               SUITE 700       954-745-5820
#> 31                                    <NA>       514-331-7440
#> 32                                    <NA>       908-273-2696
#> 33                               SUITE 201       561-877-2132
#> 34                               SUITE 130               <NA>
#> 35                               SUITE 106       949-309-2406
#> 36                               SUITE 300       949-309-2406
#> 37                                    <NA>       864-596-7115
#> 38                                 STE 202         8137979000
#> 39                               SUITE 100         7032593000
#> 40                              SUITE 3120         3055390900

# }