SEC Company Search

Provides access to the SEC Company Name Search from here using a company's formal name rather than its common name.

company_search(
  x,
  match = "start",
  file_number = FALSE,
  state = "",
  country = "",
  sic = "",
  ownership = FALSE,
  type = "",
  count = 40,
  page = 1
)

Arguments

x
Name of company to search or file number

match
(default: 'start') Either 'start' or 'contains' for where in the company name to search

file_number
(default: FALSE) if set to TRUE, x is treated as a file number

state
(default: '') Limit to a specific state of registration using 2-letter state abbreviations. Special values:

X1

The United States

A0

Alberta, Canada

A1

British Columbia, Canada

A2

Manitoba, Canada

A3

New Brunswick, Canada

A4

Newfoundland, Canada

A5

Nova Scotia, Canada

A6

Ontario, Canada

A7

Prince Edward Island, Canada

A8

Quebec, Canada

A9

Saskatchewan, Canada

B0

Yukon, Canada

country
2-character country code. The mapping is non-obvious, so unfortunately the best way to find it is to examine the company search page.

sic
SIC Code

ownership
boolean for inclusion of company change filings

type
Limit to companies with a given filing type - e.g. 'N-PX'

count
Number of filings to fetch per page. Valid options are 10, 20, 40, 80, or 100. Other values will result in the closest count.

page
Which page of results to return.

Value

A dataframe of companies

cik

company_href

name

location

location_href

formerly

sic

sic_description

sic_href

Details

Note On 'Fast Search' -- The SEC Company Search page also includes a 'Fast Search' function to "search" by CIK or Stock Ticker. This doesn't actually search, but rather goes directly to the company details page if found. If you have a company's CIK or Ticker, use the company_information, company_filings, or company_details functions.

Examples

# \donttest{
try(company_search("Intel"))

#> No encoding supplied: defaulting to UTF-8.

#> Error in check_result(res) : 
#>   EDGAR request blocked from Undeclared Automated Tool.
#> Please visit https://www.sec.gov/developer for best practices.
#> See https://mwaldstein.github.io/edgarWebR/index.html#ethical-use--fair-access for your responsibilities
#> Consider also setting the environment variable 'EDGARWEBR_USER_AGENT

# }