SEC Notice of Effectiveness

Returns the current Notice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

effectiveness()

#>   registration_number
#> 1        \n333-228592
#> 2        \n333-229093
#> 3        \n333-225028
#> 4        \n333-229262
#> 5        \n333-228521
#> 6        \n333-228767
#> 7        \n333-228782
#> 8        \n333-227992
#> 9        \n333-174001
#>                                                                                               file_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-228592&owner=exclude&count=200
#> 2 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-229093&owner=exclude&count=200
#> 3 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-225028&owner=exclude&count=200
#> 4 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-229262&owner=exclude&count=200
#> 5 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-228521&owner=exclude&count=200
#> 6 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-228767&owner=exclude&count=200
#> 7 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-228782&owner=exclude&count=200
#> 8 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-227992&owner=exclude&count=200
#> 9 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-174001&owner=exclude&count=200
#>                                   registrant
#> 1             AMERICAN HONDA RECEIVABLES LLC
#> 2                     BIOHITECH GLOBAL, INC.
#> 3                     JMAX INTERNATIONAL Ltd
#> 4              NORTHERN DYNASTY MINERALS LTD
#> 5                        Nano Dimension Ltd.
#> 6                      Upland Software, Inc.
#> 7                             Zymeworks Inc.
#> 8 GUGGENHEIM DEFINED PORTFOLIOS, SERIES 1841
#> 9               CAMBIUM LEARNING GROUP, INC.
#>                                                                                     registrant_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000890975&owner=include&count=200
#> 2 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001590383&owner=include&count=200
#> 3 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001735664&owner=include&count=200
#> 4 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001164771&owner=include&count=200
#> 5 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001643303&owner=include&count=200
#> 6 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001505155&owner=include&count=200
#> 7 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001403752&owner=include&count=200
#> 8 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001752081&owner=include&count=200
#> 9 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001466815&owner=include&count=200
#>                division
#> 1   Corporation Finance
#> 2   Corporation Finance
#> 3   Corporation Finance
#> 4   Corporation Finance
#> 5   Corporation Finance
#> 6   Corporation Finance
#> 7   Corporation Finance
#> 8 Investment Management
#> 9   Corporation Finance
#>                                                       type      effective_date
#> 1                              Securities Act Registration 2019-01-31 17:00:00
#> 2                              Securities Act Registration 2019-01-31 10:00:00
#> 3                              Securities Act Registration 2019-01-31 10:00:00
#> 4                              Securities Act Registration 2019-01-31 17:30:00
#> 5                              Securities Act Registration 2019-01-31 17:00:00
#> 6                              Securities Act Registration 2019-01-31 16:00:00
#> 7                              Securities Act Registration 2019-01-31 10:00:00
#> 8                              Securities Act Registration 2019-01-31 12:15:00
#> 9 Post-effective Amendments to Securities Act Registration 2019-01-31 00:00:00
#>   filing_date
#> 1        <NA>
#> 2        <NA>
#> 3        <NA>
#> 4        <NA>
#> 5        <NA>
#> 6        <NA>
#> 7        <NA>
#> 8        <NA>
#> 9  2018-12-18