SEC Notice of Effectiveness

Returns the current Notice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

effectiveness()

#>    registration_number
#> 1      \n333-226265-08
#> 2      \n333-226265-07
#> 3      \n333-226265-06
#> 4         \n333-226868
#> 5         \n333-226194
#> 6         \n333-226342
#> 7      \n333-226265-05
#> 8         \n333-226500
#> 9         \n333-226803
#> 10     \n333-226265-09
#> 11     \n333-226265-04
#> 12     \n333-226265-03
#> 13        \n333-226265
#> 14     \n333-226265-01
#> 15     \n333-226265-02
#> 16        \n333-226162
#> 17        \n333-157143
#> 18        \n333-157144
#> 19        \n333-162411
#> 20        \n333-225121
#>                                                                                                   file_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226265-08&owner=exclude&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226265-07&owner=exclude&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226265-06&owner=exclude&count=200
#> 4     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226868&owner=exclude&count=200
#> 5     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226194&owner=exclude&count=200
#> 6     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226342&owner=exclude&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226265-05&owner=exclude&count=200
#> 8     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226500&owner=exclude&count=200
#> 9     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226803&owner=exclude&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226265-09&owner=exclude&count=200
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226265-04&owner=exclude&count=200
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226265-03&owner=exclude&count=200
#> 13    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226265&owner=exclude&count=200
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226265-01&owner=exclude&count=200
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226265-02&owner=exclude&count=200
#> 16    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226162&owner=exclude&count=200
#> 17    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-157143&owner=exclude&count=200
#> 18    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-157144&owner=exclude&count=200
#> 19    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-162411&owner=exclude&count=200
#> 20    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-225121&owner=exclude&count=200
#>                                     registrant
#> 1                            Burleson Sand LLC
#> 2                Burleson Water Resources, LLC
#> 3                   Esquisto Resources II, LLC
#> 4                            HealthLynked Corp
#> 5      National Energy Services Reunited Corp.
#> 6              People's United Financial, Inc.
#> 7                   Petromax E&P Burleson, LLC
#> 8                    Synthetic Biologics, Inc.
#> 9  TRANSCONTINENTAL GAS PIPE LINE COMPANY, LLC
#> 10                     WHCC Infrastructure LLC
#> 11                             WHE AcqCo., LLC
#> 12                          WHR Eagle Ford LLC
#> 13         WildHorse Resource Development Corp
#> 14                 WildHorse Resources II, LLC
#> 15 WildHorse Resources Management Company, LLC
#> 16                               Forward Funds
#> 17                   FARMERS CAPITAL BANK CORP
#> 18                   FARMERS CAPITAL BANK CORP
#> 19                   FARMERS CAPITAL BANK CORP
#> 20                      Realm Therapeutics plc
#>                                                                                      registrant_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001743139&owner=include&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001715294&owner=include&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001715347&owner=include&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001680139&owner=include&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001698514&owner=include&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001378946&owner=include&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001715352&owner=include&count=200
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000894158&owner=include&count=200
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000099250&owner=include&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001749832&owner=include&count=200
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001715307&owner=include&count=200
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001715333&owner=include&count=200
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001681714&owner=include&count=200
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001715288&owner=include&count=200
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001715276&owner=include&count=200
#> 16 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000889188&owner=include&count=200
#> 17 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000713095&owner=include&count=200
#> 18 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000713095&owner=include&count=200
#> 19 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000713095&owner=include&count=200
#> 20 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001718903&owner=include&count=200
#>                 division
#> 1    Corporation Finance
#> 2    Corporation Finance
#> 3    Corporation Finance
#> 4    Corporation Finance
#> 5    Corporation Finance
#> 6    Corporation Finance
#> 7    Corporation Finance
#> 8    Corporation Finance
#> 9    Corporation Finance
#> 10   Corporation Finance
#> 11   Corporation Finance
#> 12   Corporation Finance
#> 13   Corporation Finance
#> 14   Corporation Finance
#> 15   Corporation Finance
#> 16 Investment Management
#> 17   Corporation Finance
#> 18   Corporation Finance
#> 19   Corporation Finance
#> 20   Corporation Finance
#>                                                        type      effective_date
#> 1                               Securities Act Registration 2018-08-22 16:00:00
#> 2                               Securities Act Registration 2018-08-22 16:00:00
#> 3                               Securities Act Registration 2018-08-22 16:00:00
#> 4                               Securities Act Registration 2018-08-22 16:30:00
#> 5                               Securities Act Registration 2018-08-22 16:00:00
#> 6                               Securities Act Registration 2018-08-22 09:30:00
#> 7                               Securities Act Registration 2018-08-22 16:00:00
#> 8                               Securities Act Registration 2018-08-22 16:00:00
#> 9                               Securities Act Registration 2018-08-22 09:00:00
#> 10                              Securities Act Registration 2018-08-22 16:00:00
#> 11                              Securities Act Registration 2018-08-22 16:00:00
#> 12                              Securities Act Registration 2018-08-22 16:00:00
#> 13                              Securities Act Registration 2018-08-22 16:00:00
#> 14                              Securities Act Registration 2018-08-22 16:00:00
#> 15                              Securities Act Registration 2018-08-22 16:00:00
#> 16                              Securities Act Registration 2018-08-22 16:00:00
#> 17 Post-effective Amendments to Securities Act Registration 2018-08-22 00:00:00
#> 18 Post-effective Amendments to Securities Act Registration 2018-08-22 00:00:00
#> 19 Post-effective Amendments to Securities Act Registration 2018-08-22 00:00:00
#> 20 Post-effective Amendments to Securities Act Registration 2018-08-22 00:00:00
#>    filing_date
#> 1         <NA>
#> 2         <NA>
#> 3         <NA>
#> 4         <NA>
#> 5         <NA>
#> 6         <NA>
#> 7         <NA>
#> 8         <NA>
#> 9         <NA>
#> 10        <NA>
#> 11        <NA>
#> 12        <NA>
#> 13        <NA>
#> 14        <NA>
#> 15        <NA>
#> 16        <NA>
#> 17  2018-08-20
#> 18  2018-08-20
#> 19  2018-08-20
#> 20  2018-08-22