SEC Notice of Effectiveness

Returns the current Notice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

# \donttest{
effectiveness()

#>    registration_number
#> 1         \n333-237389
#> 2         \n333-237496
#> 3         \n333-237614
#> 4         \n333-237592
#> 5         \n333-237236
#> 6         \n333-237332
#> 7      \n333-236768-01
#> 8         \n333-236768
#> 9         \n333-237187
#> 10        \n333-236645
#> 11        \n333-236644
#> 12        \n333-236646
#> 13        \n333-218400
#> 14        \n333-235412
#> 15        \n333-227519
#> 16        \n333-215642
#>                                                                                                   file_href
#> 1     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237389&owner=exclude&count=200
#> 2     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237496&owner=exclude&count=200
#> 3     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237614&owner=exclude&count=200
#> 4     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237592&owner=exclude&count=200
#> 5     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237236&owner=exclude&count=200
#> 6     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237332&owner=exclude&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-236768-01&owner=exclude&count=200
#> 8     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-236768&owner=exclude&count=200
#> 9     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237187&owner=exclude&count=200
#> 10    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-236645&owner=exclude&count=200
#> 11    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-236644&owner=exclude&count=200
#> 12    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-236646&owner=exclude&count=200
#> 13    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-218400&owner=exclude&count=200
#> 14    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235412&owner=exclude&count=200
#> 15    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-227519&owner=exclude&count=200
#> 16    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-215642&owner=exclude&count=200
#>                          registrant
#> 1            Akari Therapeutics Plc
#> 2     Bioceres Crop Solutions Corp.
#> 3              ENERGY FOCUS, INC/DE
#> 4  Hancock Jaffe Laboratories, Inc.
#> 5            Kearny Financial Corp.
#> 6                    Verastem, Inc.
#> 7                WLKP Finance Corp.
#> 8     Westlake Chemical Partners LP
#> 9                Zosano Pharma Corp
#> 10    COLUMBIA FUNDS SERIES TRUST I
#> 11      COLUMBIA FUNDS SERIES TRUST
#> 12   Columbia Funds Series Trust II
#> 13        Benitec Biopharma LTD/ADR
#> 14        Benitec Biopharma LTD/ADR
#> 15          Harvest Oil & Gas Corp.
#> 16                     Mimecast Ltd
#>                                                                                      registrant_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001541157&owner=include&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001769484&owner=include&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000924168&owner=include&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001661053&owner=include&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001617242&owner=include&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001526119&owner=include&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001698982&owner=include&count=200
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001604665&owner=include&count=200
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001587221&owner=include&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000773757&owner=include&count=200
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001097519&owner=include&count=200
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001352280&owner=include&count=200
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001552795&owner=include&count=200
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001552795&owner=include&count=200
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001361937&owner=include&count=200
#> 16 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001644675&owner=include&count=200
#>                 division
#> 1    Corporation Finance
#> 2    Corporation Finance
#> 3    Corporation Finance
#> 4    Corporation Finance
#> 5    Corporation Finance
#> 6    Corporation Finance
#> 7    Corporation Finance
#> 8    Corporation Finance
#> 9    Corporation Finance
#> 10 Investment Management
#> 11 Investment Management
#> 12 Investment Management
#> 13   Corporation Finance
#> 14   Corporation Finance
#> 15   Corporation Finance
#> 16   Corporation Finance
#>                                                        type      effective_date
#> 1                               Securities Act Registration 2020-04-16 16:30:00
#> 2                               Securities Act Registration 2020-04-16 17:25:00
#> 3                               Securities Act Registration 2020-04-16 16:30:00
#> 4                               Securities Act Registration 2020-04-16 16:30:00
#> 5                               Securities Act Registration 2020-04-16 16:00:00
#> 6                               Securities Act Registration 2020-04-16 16:00:00
#> 7                               Securities Act Registration 2020-04-16 16:30:00
#> 8                               Securities Act Registration 2020-04-16 16:30:00
#> 9                               Securities Act Registration 2020-04-16 16:00:00
#> 10                              Securities Act Registration 2020-04-16 12:00:00
#> 11                              Securities Act Registration 2020-04-16 12:00:00
#> 12                              Securities Act Registration 2020-04-16 12:00:00
#> 13 Post-effective Amendments to Securities Act Registration 2020-04-16 00:00:00
#> 14 Post-effective Amendments to Securities Act Registration 2020-04-16 00:00:00
#> 15 Post-effective Amendments to Securities Act Registration 2020-04-16 00:00:00
#> 16 Post-effective Amendments to Securities Act Registration 2020-04-16 00:00:00
#>    filing_date
#> 1         <NA>
#> 2         <NA>
#> 3         <NA>
#> 4         <NA>
#> 5         <NA>
#> 6         <NA>
#> 7         <NA>
#> 8         <NA>
#> 9         <NA>
#> 10        <NA>
#> 11        <NA>
#> 12        <NA>
#> 13  2020-04-15
#> 14  2020-04-15
#> 15  2020-04-14
#> 16  2020-04-14

# }