SEC Notice of Effectiveness

Returns the current Noice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

effectiveness()

#>    registration_number
#> 1         \n333-220295
#> 2         \n333-222475
#> 3         \n333-224990
#> 4         \n333-224881
#> 5         \n333-223956
#> 6         \n333-224930
#> 7         \n333-222829
#> 8         \n333-224732
#> 9         \n333-224505
#> 10        \n333-224488
#> 11        \n333-224884
#> 12        \n333-222581
#> 13        \n333-225077
#> 14        \n333-224493
#> 15        \n333-224951
#> 16        \n333-224880
#> 17        \n333-224153
#> 18        \n333-213451
#> 19        \n333-217272
#> 20        \n333-223255
#> 21        \n333-208174
#>                                                                                                file_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-220295&owner=exclude&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-222475&owner=exclude&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224990&owner=exclude&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224881&owner=exclude&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-223956&owner=exclude&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224930&owner=exclude&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-222829&owner=exclude&count=200
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224732&owner=exclude&count=200
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224505&owner=exclude&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224488&owner=exclude&count=200
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224884&owner=exclude&count=200
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-222581&owner=exclude&count=200
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-225077&owner=exclude&count=200
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224493&owner=exclude&count=200
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224951&owner=exclude&count=200
#> 16 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224880&owner=exclude&count=200
#> 17 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224153&owner=exclude&count=200
#> 18 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-213451&owner=exclude&count=200
#> 19 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-217272&owner=exclude&count=200
#> 20 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-223255&owner=exclude&count=200
#> 21 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-208174&owner=exclude&count=200
#>                                registrant
#> 1               AIRBORNE WIRELESS NETWORK
#> 2                        Altice USA, Inc.
#> 3  Bluerock Residential Growth REIT, Inc.
#> 4                     CHF Solutions, Inc.
#> 5                                CLPS Inc
#> 6                        DYNATRONICS CORP
#> 7                             EXENT CORP.
#> 8                            EuroDry Ltd.
#> 9                          GreenSky, Inc.
#> 10          Kiniksa Pharmaceuticals, Ltd.
#> 11           MACKINAC FINANCIAL CORP /MI/
#> 12                            PAVmed Inc.
#> 13  Sailpoint Technologies Holdings, Inc.
#> 14              Scholar Rock Holding Corp
#> 15               TENAX THERAPEUTICS, INC.
#> 16                            Tocagen Inc
#> 17              TEMPLETON GROWTH FUND INC
#> 18               NIOCORP DEVELOPMENTS LTD
#> 19               NIOCORP DEVELOPMENTS LTD
#> 20                  Nautilus Holdco, Inc.
#> 21              ZAIS Group Holdings, Inc.
#>                                                                                      registrant_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001537258&owner=include&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001702780&owner=include&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001442626&owner=include&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001506492&owner=include&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001724542&owner=include&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000720875&owner=include&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001726744&owner=include&count=200
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001731388&owner=include&count=200
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001712923&owner=include&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001730430&owner=include&count=200
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000036506&owner=include&count=200
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001624326&owner=include&count=200
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001627857&owner=include&count=200
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001727196&owner=include&count=200
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000034956&owner=include&count=200
#> 16 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001419041&owner=include&count=200
#> 17 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000805664&owner=include&count=200
#> 18 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001512228&owner=include&count=200
#> 19 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001512228&owner=include&count=200
#> 20 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001732004&owner=include&count=200
#> 21 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001562214&owner=include&count=200
#>                 division
#> 1    Corporation Finance
#> 2    Corporation Finance
#> 3    Corporation Finance
#> 4    Corporation Finance
#> 5    Corporation Finance
#> 6    Corporation Finance
#> 7    Corporation Finance
#> 8    Corporation Finance
#> 9    Corporation Finance
#> 10   Corporation Finance
#> 11   Corporation Finance
#> 12   Corporation Finance
#> 13   Corporation Finance
#> 14   Corporation Finance
#> 15   Corporation Finance
#> 16   Corporation Finance
#> 17 Investment Management
#> 18   Corporation Finance
#> 19   Corporation Finance
#> 20   Corporation Finance
#> 21   Corporation Finance
#>                                                        type effective_date
#> 1                               Securities Act Registration     1527107400
#> 2                               Securities Act Registration     1527105600
#> 3                               Securities Act Registration     1527105600
#> 4                               Securities Act Registration     1527109200
#> 5                               Securities Act Registration     1527109200
#> 6                               Securities Act Registration     1527105600
#> 7                               Securities Act Registration     1527094800
#> 8                               Securities Act Registration     1527080400
#> 9                               Securities Act Registration     1527105600
#> 10                              Securities Act Registration     1527105600
#> 11                              Securities Act Registration     1527105600
#> 12                              Securities Act Registration     1527107400
#> 13                              Securities Act Registration     1527107400
#> 14                              Securities Act Registration     1527105600
#> 15                              Securities Act Registration     1527107400
#> 16                              Securities Act Registration     1527105600
#> 17                              Securities Act Registration     1527096600
#> 18 Post-effective Amendments to Securities Act Registration     1527048000
#> 19 Post-effective Amendments to Securities Act Registration     1527048000
#> 20 Post-effective Amendments to Securities Act Registration     1527048000
#> 21 Post-effective Amendments to Securities Act Registration     1527048000
#>    filing_date
#> 1           NA
#> 2           NA
#> 3           NA
#> 4           NA
#> 5           NA
#> 6           NA
#> 7           NA
#> 8           NA
#> 9           NA
#> 10          NA
#> 11          NA
#> 12          NA
#> 13          NA
#> 14          NA
#> 15          NA
#> 16          NA
#> 17          NA
#> 18  1526443200
#> 19  1526443200
#> 20  1526529600
#> 21  1526616000