SEC Notice of Effectiveness

Returns the current Notice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

# \donttest{
effectiveness()

#>    registration_number
#> 1         \n333-235748
#> 2         \n333-234105
#> 3         \n333-234363
#> 4         \n333-235557
#> 5         \n333-235857
#> 6         \n333-234351
#> 7         \n333-235803
#> 8         \n333-234354
#> 9         \n333-235850
#> 10        \n333-235713
#> 11        \n333-234250
#> 12        \n333-235496
#> 13        \n333-234468
#> 14        \n333-224178
#> 15        \n333-227302
#>                                                                                                file_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235748&owner=exclude&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-234105&owner=exclude&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-234363&owner=exclude&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235557&owner=exclude&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235857&owner=exclude&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-234351&owner=exclude&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235803&owner=exclude&count=200
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-234354&owner=exclude&count=200
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235850&owner=exclude&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235713&owner=exclude&count=200
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-234250&owner=exclude&count=200
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235496&owner=exclude&count=200
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-234468&owner=exclude&count=200
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224178&owner=exclude&count=200
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-227302&owner=exclude&count=200
#>                            registrant
#> 1       GARDNER DENVER HOLDINGS, INC.
#> 2                HEAT BIOLOGICS, INC.
#> 3                               I-Mab
#> 4                               I-Mab
#> 5                          LIZHI INC.
#> 6                          LIZHI INC.
#> 7  PETROBRAS - PETROLEO BRASILEIRO SA
#> 8           Phoenix Tree Holdings Ltd
#> 9           Phoenix Tree Holdings Ltd
#> 10    TriState Capital Holdings, Inc.
#> 11            Velocity Financial, LLC
#> 12                            FT 8492
#> 13                     SmartTrust 460
#> 14    Carolina Trust BancShares, Inc.
#> 15    Carolina Trust BancShares, Inc.
#>                                                                                      registrant_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001699150&owner=include&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001476963&owner=include&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001778016&owner=include&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001778016&owner=include&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001783407&owner=include&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001783407&owner=include&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001119639&owner=include&count=200
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001785154&owner=include&count=200
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001785154&owner=include&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001380846&owner=include&count=200
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001692376&owner=include&count=200
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001793644&owner=include&count=200
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001792925&owner=include&count=200
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001676667&owner=include&count=200
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001676667&owner=include&count=200
#>                 division
#> 1    Corporation Finance
#> 2    Corporation Finance
#> 3    Corporation Finance
#> 4    Corporation Finance
#> 5    Corporation Finance
#> 6    Corporation Finance
#> 7    Corporation Finance
#> 8    Corporation Finance
#> 9    Corporation Finance
#> 10   Corporation Finance
#> 11   Corporation Finance
#> 12 Investment Management
#> 13 Investment Management
#> 14   Corporation Finance
#> 15   Corporation Finance
#>                                                        type      effective_date
#> 1                               Securities Act Registration 2020-01-16 16:00:00
#> 2                               Securities Act Registration 2020-01-16 09:00:00
#> 3                               Securities Act Registration 2020-01-16 16:00:00
#> 4                               Securities Act Registration 2020-01-16 16:00:00
#> 5                               Securities Act Registration 2020-01-16 16:00:00
#> 6                               Securities Act Registration 2020-01-16 16:00:00
#> 7                               Securities Act Registration 2020-01-16 16:00:00
#> 8                               Securities Act Registration 2020-01-16 16:00:00
#> 9                               Securities Act Registration 2020-01-16 16:00:00
#> 10                              Securities Act Registration 2020-01-16 16:00:00
#> 11                              Securities Act Registration 2020-01-16 14:00:00
#> 12                              Securities Act Registration 2020-01-16 11:00:00
#> 13                              Securities Act Registration 2020-01-16 15:00:00
#> 14 Post-effective Amendments to Securities Act Registration 2020-01-16 00:00:00
#> 15 Post-effective Amendments to Securities Act Registration 2020-01-16 00:00:00
#>    filing_date
#> 1         <NA>
#> 2         <NA>
#> 3         <NA>
#> 4         <NA>
#> 5         <NA>
#> 6         <NA>
#> 7         <NA>
#> 8         <NA>
#> 9         <NA>
#> 10        <NA>
#> 11        <NA>
#> 12        <NA>
#> 13        <NA>
#> 14  2020-01-10
#> 15  2020-01-10

# }