SEC Notice of Effectiveness

Returns the current Notice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

effectiveness()

#>    registration_number
#> 1         \n333-226652
#> 2         \n333-226316
#> 3         \n333-226506
#> 4         \n333-224931
#> 5         \n333-226789
#> 6         \n333-226270
#> 7         \n333-226528
#> 8         \n333-226820
#> 9         \n333-226470
#> 10        \n333-221993
#>                                                                                                file_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226652&owner=exclude&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226316&owner=exclude&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226506&owner=exclude&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224931&owner=exclude&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226789&owner=exclude&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226270&owner=exclude&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226528&owner=exclude&count=200
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226820&owner=exclude&count=200
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-226470&owner=exclude&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-221993&owner=exclude&count=200
#>                             registrant
#> 1        BELLICUM PHARMACEUTICALS, INC
#> 2               CX Network Group, Inc.
#> 3            FIRST MIDWEST BANCORP INC
#> 4              Kisses From Italy, Inc.
#> 5                Loop Industries, Inc.
#> 6  Megalith Financial Acquisition Corp
#> 7            OLD NATIONAL BANCORP /IN/
#> 8       Ocean Power Technologies, Inc.
#> 9                  PEOPLES BANCORP INC
#> 10                        Ashford Inc.
#>                                                                                      registrant_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001358403&owner=include&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001502557&owner=include&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000702325&owner=include&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001608092&owner=include&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001504678&owner=include&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001725872&owner=include&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000707179&owner=include&count=200
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001378140&owner=include&count=200
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000318300&owner=include&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001604738&owner=include&count=200
#>               division                                                     type
#> 1  Corporation Finance                              Securities Act Registration
#> 2  Corporation Finance                              Securities Act Registration
#> 3  Corporation Finance                              Securities Act Registration
#> 4  Corporation Finance                              Securities Act Registration
#> 5  Corporation Finance                              Securities Act Registration
#> 6  Corporation Finance                              Securities Act Registration
#> 7  Corporation Finance                              Securities Act Registration
#> 8  Corporation Finance                              Securities Act Registration
#> 9  Corporation Finance                              Securities Act Registration
#> 10 Corporation Finance Post-effective Amendments to Securities Act Registration
#>         effective_date filing_date
#> 1  2018-08-23 16:30:00        <NA>
#> 2  2018-08-23 10:00:00        <NA>
#> 3  2018-08-23 16:00:00        <NA>
#> 4  2018-08-23 17:00:00        <NA>
#> 5  2018-08-23 16:00:00        <NA>
#> 6  2018-08-23 16:00:00        <NA>
#> 7  2018-08-23 15:00:00        <NA>
#> 8  2018-08-23 17:30:00        <NA>
#> 9  2018-08-23 16:00:00        <NA>
#> 10 2018-08-23 00:00:00  2018-08-16