SEC Notice of Effectiveness

Returns the current Notice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

effectiveness()

#>   registration_number
#> 1        \n333-235751
#> 2        \n333-235606
#> 3        \n333-235740
#> 4        \n333-235351
#> 5        \n333-235465
#> 6        \n333-235393
#> 7        \n333-234341
#>                                                                                               file_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235751&owner=exclude&count=200
#> 2 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235606&owner=exclude&count=200
#> 3 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235740&owner=exclude&count=200
#> 4 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235351&owner=exclude&count=200
#> 5 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235465&owner=exclude&count=200
#> 6 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235393&owner=exclude&count=200
#> 7 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-234341&owner=exclude&count=200
#>                      registrant
#> 1         DELCATH SYSTEMS, INC.
#> 2             GENESIS ENERGY LP
#> 3             Green Plains Inc.
#> 4       Menlo Therapeutics Inc.
#> 5 SOUTHERN FIRST BANCSHARES INC
#> 6                       FT 8485
#> 7                SmartTrust 458
#>                                                                                     registrant_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000872912&owner=include&count=200
#> 2 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001022321&owner=include&count=200
#> 3 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001309402&owner=include&count=200
#> 4 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001566044&owner=include&count=200
#> 5 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001090009&owner=include&count=200
#> 6 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001792858&owner=include&count=200
#> 7 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001790304&owner=include&count=200
#>                division                        type      effective_date
#> 1   Corporation Finance Securities Act Registration 2020-01-07 16:01:00
#> 2   Corporation Finance Securities Act Registration 2020-01-07 16:30:00
#> 3   Corporation Finance Securities Act Registration 2020-01-07 16:00:00
#> 4   Corporation Finance Securities Act Registration 2020-01-07 16:30:00
#> 5   Corporation Finance Securities Act Registration 2020-01-07 09:00:00
#> 6 Investment Management Securities Act Registration 2020-01-07 13:00:00
#> 7 Investment Management Securities Act Registration 2020-01-07 11:50:00
#>   filing_date
#> 1        <NA>
#> 2        <NA>
#> 3        <NA>
#> 4        <NA>
#> 5        <NA>
#> 6        <NA>
#> 7        <NA>