SEC Notice of Effectiveness

Returns the current Notice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

# \donttest{
effectiveness()

#>    registration_number
#> 1         \n333-252898
#> 2         \n333-255129
#> 3         \n333-254760
#> 4         \n333-254175
#> 5         \n333-254793
#> 6      \n333-254099-05
#> 7      \n333-254099-04
#> 8         \n333-254099
#> 9         \n333-254106
#> 10     \n333-254099-11
#> 11     \n333-254099-10
#> 12        \n333-253608
#> 13        \n333-251905
#> 14     \n333-254099-02
#> 15     \n333-254099-01
#> 16     \n333-254099-03
#> 17        \n333-255136
#> 18     \n333-254099-08
#> 19     \n333-254099-07
#> 20     \n333-254099-06
#> 21        \n333-255066
#> 22        \n333-255065
#> 23        \n333-254985
#> 24        \n333-254808
#> 25        \n333-254576
#> 26        \n333-253587
#> 27     \n333-254099-09
#> 28        \n333-254889
#> 29        \n333-255146
#> 30        \n333-253034
#> 31     \n333-253034-01
#> 32        \n333-253822
#> 33        \n333-254207
#> 34        \n333-253447
#> 35        \n333-222533
#>                                                                                                   file_href
#> 1     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-252898&owner=exclude&count=200
#> 2     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-255129&owner=exclude&count=200
#> 3     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254760&owner=exclude&count=200
#> 4     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254175&owner=exclude&count=200
#> 5     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254793&owner=exclude&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099-05&owner=exclude&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099-04&owner=exclude&count=200
#> 8     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099&owner=exclude&count=200
#> 9     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254106&owner=exclude&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099-11&owner=exclude&count=200
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099-10&owner=exclude&count=200
#> 12    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-253608&owner=exclude&count=200
#> 13    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-251905&owner=exclude&count=200
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099-02&owner=exclude&count=200
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099-01&owner=exclude&count=200
#> 16 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099-03&owner=exclude&count=200
#> 17    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-255136&owner=exclude&count=200
#> 18 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099-08&owner=exclude&count=200
#> 19 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099-07&owner=exclude&count=200
#> 20 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099-06&owner=exclude&count=200
#> 21    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-255066&owner=exclude&count=200
#> 22    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-255065&owner=exclude&count=200
#> 23    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254985&owner=exclude&count=200
#> 24    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254808&owner=exclude&count=200
#> 25    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254576&owner=exclude&count=200
#> 26    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-253587&owner=exclude&count=200
#> 27 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254099-09&owner=exclude&count=200
#> 28    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254889&owner=exclude&count=200
#> 29    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-255146&owner=exclude&count=200
#> 30    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-253034&owner=exclude&count=200
#> 31 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-253034-01&owner=exclude&count=200
#> 32    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-253822&owner=exclude&count=200
#> 33    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-254207&owner=exclude&count=200
#> 34    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-253447&owner=exclude&count=200
#> 35    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-222533&owner=exclude&count=200
#>                                        registrant
#> 1                                       AMGEN INC
#> 2                    Aerpio Pharmaceuticals, Inc.
#> 3                         Akoya Biosciences, Inc.
#> 4                      Aldeyra Therapeutics, Inc.
#> 5                             Biomea Fusion, Inc.
#> 6                             Bold Energy III LLC
#> 7                             Bold Operating, LLC
#> 8                           EARTHSTONE ENERGY INC
#> 9                           EARTHSTONE ENERGY INC
#> 10                Earthstone Energy Holdings, LLC
#> 11                      Earthstone Operating, LLC
#> 12                          FIRST BUSEY CORP /NV/
#> 13         IRSA INVESTMENTS & REPRESENTATIONS INC
#> 14          Independence Resources Land Co A, LLC
#> 15          Independence Resources Land Co B, LLC
#> 16         Independence Resources Management, LLC
#> 17                                   LENSAR, Inc.
#> 18                            Lynden Energy Corp.
#> 19                                Lynden USA Inc.
#> 20                      Lynden USA Operating, LLC
#> 21                       NORTHERN OIL & GAS, INC.
#> 22                       NORTHERN OIL & GAS, INC.
#> 23                   Oncternal Therapeutics, Inc.
#> 24                               REALNETWORKS INC
#> 25                RECURSION PHARMACEUTICALS, INC.
#> 26                           REPUBLIC OF COLOMBIA
#> 27                       Sabine River Energy, LLC
#> 28                       Troika Media Group, Inc.
#> 29                      VISIUM TECHNOLOGIES, INC.
#> 30                             Verizon ABS II LLC
#> 31                           Verizon Master Trust
#> 32                                        FT 9295
#> 33                                        FT 9329
#> 34                       MUTUAL FUND SERIES TRUST
#> 35 Jones Lang LaSalle Income Property Trust, Inc.
#>                                                                                      registrant_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000318154&owner=include&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001422142&owner=include&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001711933&owner=include&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001341235&owner=include&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001840439&owner=include&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001737571&owner=include&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001737567&owner=include&count=200
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000010254&owner=include&count=200
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000010254&owner=include&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001737623&owner=include&count=200
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001647049&owner=include&count=200
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000314489&owner=include&count=200
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000933267&owner=include&count=200
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001849865&owner=include&count=200
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001850106&owner=include&count=200
#> 16 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001850126&owner=include&count=200
#> 17 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001320350&owner=include&count=200
#> 18 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001622620&owner=include&count=200
#> 19 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001737580&owner=include&count=200
#> 20 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001737584&owner=include&count=200
#> 21 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001104485&owner=include&count=200
#> 22 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001104485&owner=include&count=200
#> 23 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001260990&owner=include&count=200
#> 24 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001046327&owner=include&count=200
#> 25 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001601830&owner=include&count=200
#> 26 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000917142&owner=include&count=200
#> 27 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001647013&owner=include&count=200
#> 28 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001021096&owner=include&count=200
#> 29 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001082733&owner=include&count=200
#> 30 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001836995&owner=include&count=200
#> 31 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001844964&owner=include&count=200
#> 32 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001843319&owner=include&count=200
#> 33 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001844192&owner=include&count=200
#> 34 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001355064&owner=include&count=200
#> 35 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001314152&owner=include&count=200
#>                 division
#> 1    Corporation Finance
#> 2    Corporation Finance
#> 3    Corporation Finance
#> 4    Corporation Finance
#> 5    Corporation Finance
#> 6    Corporation Finance
#> 7    Corporation Finance
#> 8    Corporation Finance
#> 9    Corporation Finance
#> 10   Corporation Finance
#> 11   Corporation Finance
#> 12   Corporation Finance
#> 13   Corporation Finance
#> 14   Corporation Finance
#> 15   Corporation Finance
#> 16   Corporation Finance
#> 17   Corporation Finance
#> 18   Corporation Finance
#> 19   Corporation Finance
#> 20   Corporation Finance
#> 21   Corporation Finance
#> 22   Corporation Finance
#> 23   Corporation Finance
#> 24   Corporation Finance
#> 25   Corporation Finance
#> 26   Corporation Finance
#> 27   Corporation Finance
#> 28   Corporation Finance
#> 29   Corporation Finance
#> 30   Corporation Finance
#> 31   Corporation Finance
#> 32 Investment Management
#> 33 Investment Management
#> 34 Investment Management
#> 35   Corporation Finance
#>                                                        type      effective_date
#> 1                               Securities Act Registration 2021-04-15 16:00:00
#> 2                               Securities Act Registration 2021-04-15 16:01:00
#> 3                               Securities Act Registration 2021-04-15 16:00:00
#> 4                               Securities Act Registration 2021-04-15 16:00:00
#> 5                               Securities Act Registration 2021-04-15 16:00:00
#> 6                               Securities Act Registration 2021-04-15 17:00:00
#> 7                               Securities Act Registration 2021-04-15 17:00:00
#> 8                               Securities Act Registration 2021-04-15 17:00:00
#> 9                               Securities Act Registration 2021-04-15 17:00:00
#> 10                              Securities Act Registration 2021-04-15 17:00:00
#> 11                              Securities Act Registration 2021-04-15 17:00:00
#> 12                              Securities Act Registration 2021-04-15 16:00:00
#> 13                              Securities Act Registration 2021-04-15 09:30:00
#> 14                              Securities Act Registration 2021-04-15 17:00:00
#> 15                              Securities Act Registration 2021-04-15 17:00:00
#> 16                              Securities Act Registration 2021-04-15 17:00:00
#> 17                              Securities Act Registration 2021-04-15 16:00:00
#> 18                              Securities Act Registration 2021-04-15 17:00:00
#> 19                              Securities Act Registration 2021-04-15 17:00:00
#> 20                              Securities Act Registration 2021-04-15 17:00:00
#> 21                              Securities Act Registration 2021-04-15 16:00:00
#> 22                              Securities Act Registration 2021-04-15 16:00:00
#> 23                              Securities Act Registration 2021-04-15 16:00:00
#> 24                              Securities Act Registration 2021-04-15 16:00:00
#> 25                              Securities Act Registration 2021-04-15 16:00:00
#> 26                              Securities Act Registration 2021-04-15 16:00:00
#> 27                              Securities Act Registration 2021-04-15 17:00:00
#> 28                              Securities Act Registration 2021-04-15 16:30:00
#> 29                              Securities Act Registration 2021-04-15 17:00:00
#> 30                              Securities Act Registration 2021-04-15 17:00:00
#> 31                              Securities Act Registration 2021-04-15 17:00:00
#> 32                              Securities Act Registration 2021-04-15 11:35:00
#> 33                              Securities Act Registration 2021-04-15 11:50:00
#> 34                              Securities Act Registration 2021-04-15 16:00:00
#> 35 Post-effective Amendments to Securities Act Registration 2021-04-15 00:00:00
#>    filing_date
#> 1         <NA>
#> 2         <NA>
#> 3         <NA>
#> 4         <NA>
#> 5         <NA>
#> 6         <NA>
#> 7         <NA>
#> 8         <NA>
#> 9         <NA>
#> 10        <NA>
#> 11        <NA>
#> 12        <NA>
#> 13        <NA>
#> 14        <NA>
#> 15        <NA>
#> 16        <NA>
#> 17        <NA>
#> 18        <NA>
#> 19        <NA>
#> 20        <NA>
#> 21        <NA>
#> 22        <NA>
#> 23        <NA>
#> 24        <NA>
#> 25        <NA>
#> 26        <NA>
#> 27        <NA>
#> 28        <NA>
#> 29        <NA>
#> 30        <NA>
#> 31        <NA>
#> 32        <NA>
#> 33        <NA>
#> 34        <NA>
#> 35  2021-04-05

# }