SEC Notice of Effectiveness

Returns the current Notice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

# \donttest{
effectiveness()

#>    registration_number
#> 1         \n333-248492
#> 2         \n333-248001
#> 3         \n333-239860
#> 4      \n333-239860-02
#> 5      \n333-239860-01
#> 6      \n333-239860-03
#> 7         \n333-248832
#> 8         \n333-248664
#> 9         \n333-248611
#> 10        \n333-239958
#> 11        \n333-238829
#> 12        \n333-248895
#> 13        \n333-240045
#> 14        \n333-239896
#> 15        \n333-248627
#> 16        \n333-248628
#> 17        \n333-248254
#> 18        \n333-248599
#> 19        \n333-248714
#> 20        \n333-248665
#> 21        \n333-237776
#> 22        \n333-248058
#> 23        \n333-241755
#> 24        \n333-229978
#> 25        \n333-230962
#> 26        \n333-222094
#>                                                                                                   file_href
#> 1     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248492&owner=exclude&count=200
#> 2     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248001&owner=exclude&count=200
#> 3     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-239860&owner=exclude&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-239860-02&owner=exclude&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-239860-01&owner=exclude&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-239860-03&owner=exclude&count=200
#> 7     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248832&owner=exclude&count=200
#> 8     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248664&owner=exclude&count=200
#> 9     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248611&owner=exclude&count=200
#> 10    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-239958&owner=exclude&count=200
#> 11    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-238829&owner=exclude&count=200
#> 12    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248895&owner=exclude&count=200
#> 13    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-240045&owner=exclude&count=200
#> 14    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-239896&owner=exclude&count=200
#> 15    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248627&owner=exclude&count=200
#> 16    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248628&owner=exclude&count=200
#> 17    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248254&owner=exclude&count=200
#> 18    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248599&owner=exclude&count=200
#> 19    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248714&owner=exclude&count=200
#> 20    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248665&owner=exclude&count=200
#> 21    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237776&owner=exclude&count=200
#> 22    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-248058&owner=exclude&count=200
#> 23    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-241755&owner=exclude&count=200
#> 24    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-229978&owner=exclude&count=200
#> 25    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-230962&owner=exclude&count=200
#> 26    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-222094&owner=exclude&count=200
#>                                registrant
#> 1                            AEMETIS, INC
#> 2                            Amesite Inc.
#> 3   Cheniere Corpus Christi Holdings, LLC
#> 4  Cheniere Corpus Christi Pipeline, L.P.
#> 5        Corpus Christi Liquefaction, LLC
#> 6         Corpus Christi Pipeline GP, LLC
#> 7                           Energous Corp
#> 8            FinTech Acquisition Corp. IV
#> 9                    GRAYBUG VISION, INC.
#> 10                      GeoVax Labs, Inc.
#> 11           Greenwich LifeSciences, Inc.
#> 12         Guardion Health Sciences, Inc.
#> 13 Hall of Fame Resort & Entertainment Co
#> 14            Insurance Acquisition Corp.
#> 15              PMV Pharmaceuticals, Inc.
#> 16               Prelude Therapeutics Inc
#> 17                     Siyata Mobile Inc.
#> 18                       VIA optronics AG
#> 19                   VIA optronics AG/ADR
#> 20                Vector Acquisition Corp
#> 21     BlackRock Capital Allocation Trust
#> 22                                FT 8897
#> 23  T. Rowe Price Total Return Fund, Inc.
#> 24                       IVERIC bio, Inc.
#> 25                                Majesco
#> 26                  TPT GLOBAL TECH, INC.
#>                                                                                      registrant_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000738214&owner=include&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001807166&owner=include&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001693317&owner=include&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001693288&owner=include&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001693307&owner=include&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001693287&owner=include&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001575793&owner=include&count=200
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001777835&owner=include&count=200
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001534133&owner=include&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000832489&owner=include&count=200
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001799788&owner=include&count=200
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001642375&owner=include&count=200
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001708176&owner=include&count=200
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001762322&owner=include&count=200
#> 15 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001699382&owner=include&count=200
#> 16 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001678660&owner=include&count=200
#> 17 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001649009&owner=include&count=200
#> 18 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001769116&owner=include&count=200
#> 19 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001767934&owner=include&count=200
#> 20 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001819994&owner=include&count=200
#> 21 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001809541&owner=include&count=200
#> 22 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001816914&owner=include&count=200
#> 23 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001681576&owner=include&count=200
#> 24 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001410939&owner=include&count=200
#> 25 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001626853&owner=include&count=200
#> 26 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001661039&owner=include&count=200
#>                 division
#> 1    Corporation Finance
#> 2    Corporation Finance
#> 3    Corporation Finance
#> 4    Corporation Finance
#> 5    Corporation Finance
#> 6    Corporation Finance
#> 7    Corporation Finance
#> 8    Corporation Finance
#> 9    Corporation Finance
#> 10   Corporation Finance
#> 11   Corporation Finance
#> 12   Corporation Finance
#> 13   Corporation Finance
#> 14   Corporation Finance
#> 15   Corporation Finance
#> 16   Corporation Finance
#> 17   Corporation Finance
#> 18   Corporation Finance
#> 19   Corporation Finance
#> 20   Corporation Finance
#> 21 Investment Management
#> 22 Investment Management
#> 23 Investment Management
#> 24   Corporation Finance
#> 25   Corporation Finance
#> 26   Corporation Finance
#>                                                        type      effective_date
#> 1                               Securities Act Registration 2020-09-24 16:00:00
#> 2                               Securities Act Registration 2020-09-24 17:00:00
#> 3                               Securities Act Registration 2020-09-24 16:00:00
#> 4                               Securities Act Registration 2020-09-24 16:00:00
#> 5                               Securities Act Registration 2020-09-24 16:00:00
#> 6                               Securities Act Registration 2020-09-24 16:00:00
#> 7                               Securities Act Registration 2020-09-24 16:00:00
#> 8                               Securities Act Registration 2020-09-24 16:00:00
#> 9                               Securities Act Registration 2020-09-24 16:00:00
#> 10                              Securities Act Registration 2020-09-24 17:30:00
#> 11                              Securities Act Registration 2020-09-24 17:25:00
#> 12                              Securities Act Registration 2020-09-24 16:00:00
#> 13                              Securities Act Registration 2020-09-24 09:00:00
#> 14                              Securities Act Registration 2020-09-24 15:00:00
#> 15                              Securities Act Registration 2020-09-24 16:00:00
#> 16                              Securities Act Registration 2020-09-24 16:00:00
#> 17                              Securities Act Registration 2020-09-24 17:30:00
#> 18                              Securities Act Registration 2020-09-24 16:00:00
#> 19                              Securities Act Registration 2020-09-24 16:00:00
#> 20                              Securities Act Registration 2020-09-24 16:00:00
#> 21                              Securities Act Registration 2020-09-24 09:30:00
#> 22                              Securities Act Registration 2020-09-24 11:50:00
#> 23                              Securities Act Registration 2020-09-24 09:30:00
#> 24 Post-effective Amendments to Securities Act Registration 2020-09-24 00:00:00
#> 25 Post-effective Amendments to Securities Act Registration 2020-09-24 00:00:00
#> 26 Post-effective Amendments to Securities Act Registration 2020-09-24 00:00:00
#>    filing_date
#> 1         <NA>
#> 2         <NA>
#> 3         <NA>
#> 4         <NA>
#> 5         <NA>
#> 6         <NA>
#> 7         <NA>
#> 8         <NA>
#> 9         <NA>
#> 10        <NA>
#> 11        <NA>
#> 12        <NA>
#> 13        <NA>
#> 14        <NA>
#> 15        <NA>
#> 16        <NA>
#> 17        <NA>
#> 18        <NA>
#> 19        <NA>
#> 20        <NA>
#> 21        <NA>
#> 22        <NA>
#> 23        <NA>
#> 24  2020-09-18
#> 25  2020-09-23
#> 26  2020-09-14

# }