SEC Notice of Effectiveness

Returns the current Notice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

effectiveness()

#>   registration_number
#> 1        \n333-224209
#> 2        \n333-225448
#> 3        \n333-225267
#> 4        \n333-223604
#> 5        \n333-224857
#> 6        \n333-215581
#>                                                                                               file_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224209&owner=exclude&count=200
#> 2 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-225448&owner=exclude&count=200
#> 3 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-225267&owner=exclude&count=200
#> 4 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-223604&owner=exclude&count=200
#> 5 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-224857&owner=exclude&count=200
#> 6 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-215581&owner=exclude&count=200
#>                                  registrant
#> 1                           HRG GROUP, INC.
#> 2                     REMARK HOLDINGS, INC.
#> 3                           Rise Gold Corp.
#> 4                        SAUER ENERGY, INC.
#> 5                                   FT 7424
#> 6 PIMCO CORPORATE & INCOME OPPORTUNITY FUND
#>                                                                                     registrant_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000109177&owner=include&count=200
#> 2 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001368365&owner=include&count=200
#> 3 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001424864&owner=include&count=200
#> 4 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001446152&owner=include&count=200
#> 5 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001734473&owner=include&count=200
#> 6 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001190935&owner=include&count=200
#>                division
#> 1   Corporation Finance
#> 2   Corporation Finance
#> 3   Corporation Finance
#> 4   Corporation Finance
#> 5 Investment Management
#> 6 Investment Management
#>                                                       type      effective_date
#> 1                              Securities Act Registration 2018-06-12 09:00:00
#> 2                              Securities Act Registration 2018-06-12 16:00:00
#> 3                              Securities Act Registration 2018-06-12 16:30:00
#> 4                              Securities Act Registration 2018-06-12 11:00:00
#> 5                              Securities Act Registration 2018-06-12 17:00:00
#> 6 Post-effective Amendments to Securities Act Registration 2018-06-12 00:00:00
#>   filing_date
#> 1        <NA>
#> 2        <NA>
#> 3        <NA>
#> 4        <NA>
#> 5        <NA>
#> 6  2018-06-11