SEC Notice of Effectiveness

Returns the current Notice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

effectiveness()

#>   registration_number
#> 1        \n333-234425
#> 2     \n333-234425-01
#> 3        \n333-235666
#> 4        \n333-235655
#> 5        \n333-235675
#> 6        \n333-235650
#> 7        \n333-216033
#> 8        \n333-231951
#> 9        \n333-228659
#>                                                                                                  file_href
#> 1    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-234425&owner=exclude&count=200
#> 2 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-234425-01&owner=exclude&count=200
#> 3    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235666&owner=exclude&count=200
#> 4    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235655&owner=exclude&count=200
#> 5    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235675&owner=exclude&count=200
#> 6    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-235650&owner=exclude&count=200
#> 7    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-216033&owner=exclude&count=200
#> 8    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-231951&owner=exclude&count=200
#> 9    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-228659&owner=exclude&count=200
#>                                registrant
#> 1               BANK OF AMERICA CORP /DE/
#> 2                        BofA Finance LLC
#> 3                            Cerecor Inc.
#> 4          Eiger BioPharmaceuticals, Inc.
#> 5                       FedNat Holding Co
#> 6                       Hess Midstream LP
#> 7 AlphaCentric Prime Meridian Income Fund
#> 8                Emerald Bioscience, Inc.
#> 9       FSI Low Beta Absolute Return Fund
#>                                                                                     registrant_href
#> 1 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0000070858&owner=include&count=200
#> 2 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001682472&owner=include&count=200
#> 3 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001534120&owner=include&count=200
#> 4 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001305253&owner=include&count=200
#> 5 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001069996&owner=include&count=200
#> 6 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001789832&owner=include&count=200
#> 7 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001697196&owner=include&count=200
#> 8 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001516551&owner=include&count=200
#> 9 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001527413&owner=include&count=200
#>                division
#> 1   Corporation Finance
#> 2   Corporation Finance
#> 3   Corporation Finance
#> 4   Corporation Finance
#> 5   Corporation Finance
#> 6   Corporation Finance
#> 7 Investment Management
#> 8   Corporation Finance
#> 9 Investment Management
#>                                                       type      effective_date
#> 1                              Securities Act Registration 2019-12-31 10:00:00
#> 2                              Securities Act Registration 2019-12-31 10:00:00
#> 3                              Securities Act Registration 2019-12-31 16:05:00
#> 4                              Securities Act Registration 2019-12-31 16:00:00
#> 5                              Securities Act Registration 2019-12-31 14:00:00
#> 6                              Securities Act Registration 2019-12-31 15:00:00
#> 7                              Securities Act Registration 2019-12-31 14:00:00
#> 8 Post-effective Amendments to Securities Act Registration 2019-12-31 00:00:00
#> 9 Post-effective Amendments to Securities Act Registration 2019-12-31 00:00:00
#>   filing_date
#> 1        <NA>
#> 2        <NA>
#> 3        <NA>
#> 4        <NA>
#> 5        <NA>
#> 6        <NA>
#> 7        <NA>
#> 8  2019-12-24
#> 9  2019-12-27