SEC Notice of Effectiveness

Returns the current Notice of Effectiveness from the most recently completed business day from here

effectiveness()

Value

a data.frame with each row as a submission with the following columns:

registration_number

file_href

registrant

registrant_href

filing_date

effective_date

division

type

Details

You can also see the same filings going further back by using `latest_filings()` specifying the type = "EFFECT"

Examples

# \donttest{
effectiveness()

#>    registration_number
#> 1         \n333-236974
#> 2         \n333-236977
#> 3         \n333-237693
#> 4      \n333-237693-01
#> 5         \n333-237175
#> 6         \n333-236357
#> 7         \n333-236792
#> 8         \n333-236468
#> 9         \n333-237637
#> 10        \n333-237582
#> 11        \n333-222231
#> 12        \n333-233976
#> 13        \n333-220997
#> 14        \n333-225382
#>                                                                                                   file_href
#> 1     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-236974&owner=exclude&count=200
#> 2     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-236977&owner=exclude&count=200
#> 3     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237693&owner=exclude&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237693-01&owner=exclude&count=200
#> 5     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237175&owner=exclude&count=200
#> 6     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-236357&owner=exclude&count=200
#> 7     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-236792&owner=exclude&count=200
#> 8     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-236468&owner=exclude&count=200
#> 9     https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237637&owner=exclude&count=200
#> 10    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-237582&owner=exclude&count=200
#> 11    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-222231&owner=exclude&count=200
#> 12    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-233976&owner=exclude&count=200
#> 13    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-220997&owner=exclude&count=200
#> 14    https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&filenum=333-225382&owner=exclude&count=200
#>                                 registrant
#> 1        CC Neuberger Principal Holdings I
#> 2   Chardan Healthcare Acquisition 2 Corp.
#> 3                     DEAC NV Merger Corp.
#> 4    Diamond Eagle Acquisition Corp. \\ DE
#> 5               Harpoon Therapeutics, Inc.
#> 6                                 Live Inc
#> 7               Oric Pharmaceuticals, Inc.
#> 8                                  PAE Inc
#> 9                Performance Shipping Inc.
#> 10                            TECOGEN INC.
#> 11         Nuveen Global Cities REIT, Inc.
#> 12                       Safe-T Group Ltd.
#> 13 Starwood Real Estate Income Trust, Inc.
#> 14                    Tallgrass Energy, LP
#>                                                                                      registrant_href
#> 1  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001800347&owner=include&count=200
#> 2  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001799850&owner=include&count=200
#> 3  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001797122&owner=include&count=200
#> 4  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001772757&owner=include&count=200
#> 5  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001708493&owner=include&count=200
#> 6  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001730732&owner=include&count=200
#> 7  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001796280&owner=include&count=200
#> 8  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001720821&owner=include&count=200
#> 9  https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001481241&owner=include&count=200
#> 10 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001537435&owner=include&count=200
#> 11 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001711799&owner=include&count=200
#> 12 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001725332&owner=include&count=200
#> 13 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001711929&owner=include&count=200
#> 14 https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=0001633651&owner=include&count=200
#>               division                                                     type
#> 1  Corporation Finance                              Securities Act Registration
#> 2  Corporation Finance                              Securities Act Registration
#> 3  Corporation Finance                              Securities Act Registration
#> 4  Corporation Finance                              Securities Act Registration
#> 5  Corporation Finance                              Securities Act Registration
#> 6  Corporation Finance                              Securities Act Registration
#> 7  Corporation Finance                              Securities Act Registration
#> 8  Corporation Finance                              Securities Act Registration
#> 9  Corporation Finance                              Securities Act Registration
#> 10 Corporation Finance                              Securities Act Registration
#> 11 Corporation Finance Post-effective Amendments to Securities Act Registration
#> 12 Corporation Finance Post-effective Amendments to Securities Act Registration
#> 13 Corporation Finance Post-effective Amendments to Securities Act Registration
#> 14 Corporation Finance Post-effective Amendments to Securities Act Registration
#>         effective_date filing_date
#> 1  2020-04-23 16:00:00        <NA>
#> 2  2020-04-23 16:00:00        <NA>
#> 3  2020-04-23 12:00:00        <NA>
#> 4  2020-04-23 12:00:00        <NA>
#> 5  2020-04-23 16:00:00        <NA>
#> 6  2020-04-23 16:00:00        <NA>
#> 7  2020-04-23 16:00:00        <NA>
#> 8  2020-04-23 16:00:00        <NA>
#> 9  2020-04-23 16:00:00        <NA>
#> 10 2020-04-23 16:00:00        <NA>
#> 11 2020-04-23 00:00:00  2020-04-17
#> 12 2020-04-23 00:00:00  2020-04-16
#> 13 2020-04-23 00:00:00  2020-04-16
#> 14 2020-04-23 00:00:00  2020-04-20

# }