SEC Filing Details

The SEC generates a html page as an index for every filing it receives containing all the meta-information about the filing. We extract 3 main types of information:

Filing Information

Filing date, accepted date, etc.

Documents

All the documents included in the filing

Filers

Companies included in the filing

Funds

Funds included in the filing

filing_details(x)

# S3 method for character
filing_details(x)

# S3 method for xml_node
filing_details(x)

Arguments

x
URL to a SEC filing index page

Value

A list with the following components:

information

A data.frame as returned by filing_information

documents

A data.frame as returned by filing_documents

filers

A data.frame as returned by filing_filers

funds

A data.frame as returned by filing_funds

Details

For a company, there is typically a single filer and no funds, but many filings for funds get more complicated - e.g. 400+ funds with 100's of companies

NOTE: This can get process intensive for large fund pages. If you don't need all components, try just using filing_info

Examples

# \donttest{
# Typically you'd get the URL from one of the search functions
x <- paste0("https://www.sec.gov/Archives/edgar/data/",
            "712515/000071251517000063/0000712515-17-000063-index.htm")
try(filing_details(x))

#> No encoding supplied: defaulting to UTF-8.

#> Error in check_result(res) : 
#>   EDGAR request blocked from Undeclared Automated Tool.
#> Please visit https://www.sec.gov/developer for best practices.
#> See https://mwaldstein.github.io/edgarWebR/index.html#ethical-use--fair-access for your responsibilities
#> Consider also setting the environment variable 'EDGARWEBR_USER_AGENT

# }