SEC Filing Documents

If you know you're going to want all the details of a filing, including documents funds and filers, look at `filing_details`

filing_documents(x)

Arguments

x
URL or xml_document for a SEC filing index page

Value

A dataframe with all the documents in the filing along with their meta info

Details

Information returned:

seq

description

document

href

type

size

Examples

filing_documents("https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/0000712515-17-000063-index.htm")

#>    seq                                                         description
#> 1    1                                       QUARTERLY REPORT ON FORM 10-Q
#> 2    2                                        AWARENESS LETTER OF KPMG LLP
#> 3    3 CERTIFICATION OF CHIEF EXECUTIVE OFFICER PURSUANT TO RULE 13A-14(A)
#> 4    4 CERTIFICATION OF CHIEF FINANCIAL OFFICER PURSUANT TO RULE 13A-14(A)
#> 5    5    CERTIFICATION OF CHIEF EXECUTIVE OFFICER PURSUANT TO SECTION 906
#> 6    6    CERTIFICATION OF CHIEF FINANCIAL OFFICER PURSUANT TO SECTION 906
#> 7   NA                                       Complete submission text file
#> 8    7                                              XBRL INSTANCE DOCUMENT
#> 9    8                             XBRL TAXONOMY EXTENSION SCHEMA DOCUMENT
#> 10   9               XBRL TAXONOMY EXTENSION CALCULATION LINKBASE DOCUMENT
#> 11  10                XBRL TAXONOMY EXTENSION DEFINITION LINKBASE DOCUMENT
#> 12  11                     XBRL TAXONOMY EXTENSION LABEL LINKBASE DOCUMENT
#> 13  12              XBRL TAXONOMY EXTENSION PRESENTATION LINKBASE DOCUMENT
#>                          document
#> 1      ea6302017-q1fy1810qdoc.htm
#> 2  ex-151kpmgawarenessletter6.htm
#> 3    ex-311sec302ceocert63017.htm
#> 4    ex-312sec302cfocert63017.htm
#> 5    ex-321sec906ceocert63017.htm
#> 6    ex-322sec906cfocert63017.htm
#> 7        0000712515-17-000063.txt
#> 8                 ea-20170630.xml
#> 9                 ea-20170630.xsd
#> 10            ea-20170630_cal.xml
#> 11            ea-20170630_def.xml
#> 12            ea-20170630_lab.xml
#> 13            ea-20170630_pre.xml
#>                                                                                                href
#> 1      https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea6302017-q1fy1810qdoc.htm
#> 2  https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ex-151kpmgawarenessletter6.htm
#> 3    https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ex-311sec302ceocert63017.htm
#> 4    https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ex-312sec302cfocert63017.htm
#> 5    https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ex-321sec906ceocert63017.htm
#> 6    https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ex-322sec906cfocert63017.htm
#> 7        https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/0000712515-17-000063.txt
#> 8                 https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630.xml
#> 9                 https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630.xsd
#> 10            https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630_cal.xml
#> 11            https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630_def.xml
#> 12            https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630_lab.xml
#> 13            https://www.sec.gov/Archives/edgar/data/712515/000071251517000063/ea-20170630_pre.xml
#>          type    size
#> 1        10-Q 1428163
#> 2     EX-15.1    4013
#> 3     EX-31.1   14907
#> 4     EX-31.2   15281
#> 5     EX-32.1    7059
#> 6     EX-32.2    7248
#> 7             8947076
#> 8  EX-101.INS 2088048
#> 9  EX-101.SCH   74416
#> 10 EX-101.CAL   92568
#> 11 EX-101.DEF  479648
#> 12 EX-101.LAB  888418
#> 13 EX-101.PRE  574659