SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

## This can be very slow running
full_text('intel')

#>    filing_date
#> 1   2018-08-24
#> 2   2018-08-24
#> 3   2018-08-24
#> 4   2018-08-24
#> 5   2018-08-24
#> 6   2018-08-24
#> 7   2018-08-24
#> 8   2018-08-24
#> 9   2018-08-24
#> 10  2018-08-24
#>                                                                      name
#> 1                                       N-PX for WELLS FARGO MASTER TRUST
#> 2                                       N-PX for THRIVENT SERIES FUND INC
#> 3                                          N-CSR for SIT MUTUAL FUNDS INC
#> 4                                   N-CSR for SIT MID CAP GROWTH FUND INC
#> 5                                 N-CSR for SIT LARGE CAP GROWTH FUND INC
#> 6                                          N-PX for THOMPSON IM FUNDS INC
#> 7                                               N-PX for JPMorgan Trust I
#> 8                                               N-PX for Highland Funds I
#> 9                                              N-PX for HIGHLAND FUNDS II
#> 10 INFORMATION TABLE of 13F-HR/A for SCHRODER INVESTMENT MANAGEMENT GROUP
#>                                                                                                      href
#> 1               https://www.sec.gov/Archives/edgar/data/1087961/000108140018000762/mastertrustnpx2018.txt
#> 2                       https://www.sec.gov/Archives/edgar/data/790166/000119312518256420/d502012dnpx.htm
#> 3                      https://www.sec.gov/Archives/edgar/data/877880/000119312518256703/d590614dncsr.htm
#> 4                      https://www.sec.gov/Archives/edgar/data/356787/000119312518256704/d595976dncsr.htm
#> 5                      https://www.sec.gov/Archives/edgar/data/356786/000119312518256705/d595901dncsr.htm
#> 6                    https://www.sec.gov/Archives/edgar/data/795264/000120677418002595/tim3456981-npx.htm
#> 7           https://www.sec.gov/Archives/edgar/data/1217286/000143893418000330/brdG4F_0001217286_2018.txt
#> 8                https://www.sec.gov/Archives/edgar/data/1354917/000143893418000332/brdQY8_0001354917.txt
#> 9                 https://www.sec.gov/Archives/edgar/data/891079/000143893418000334/brdQY8_0000891079.txt
#> 10 https://www.sec.gov/Archives/edgar/data/1086619/000156761918001750/xslForm13F_X01/form13fInfoTable.xml
#>                            company_name     cik  sic
#> 1              WELLS FARGO MASTER TRUST 1087961 <NA>
#> 2              THRIVENT SERIES FUND INC  790166 <NA>
#> 3                  SIT MUTUAL FUNDS INC  877880 <NA>
#> 4           SIT MID CAP GROWTH FUND INC  356787 <NA>
#> 5         SIT LARGE CAP GROWTH FUND INC  356786 <NA>
#> 6                 THOMPSON IM FUNDS INC  795264 <NA>
#> 7                      JPMorgan Trust I 1217286 <NA>
#> 8                      Highland Funds I 1354917 <NA>
#> 9                     HIGHLAND FUNDS II  891079 <NA>
#> 10 SCHRODER INVESTMENT MANAGEMENT GROUP 1086619 <NA>
#>                                                                                                                                                                                                                                                                                                        content
#> 1       UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, DC 20549 FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY Investment Company Act file number 811-09689 Wells Fargo Master Trust _ (Exact name of registrant as specified in charter) 525 Market
#> 2                          UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number: 811-4603 Thrivent Series Fund, Inc. (Exact name of registrant as specified
#> 3                                 UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number: 811-06373 Sit Mutual Funds, Inc. (Exact name of registrant as specified in
#> 4                          UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number: 811-03342 Sit Mid Cap Growth Fund, Inc. (Exact name of registrant as specified in
#> 5                           UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number: 811-03343 Sit Large Cap Growth Fund, Inc. (Exact name of registrant as specified
#> 6     UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY Investment Company Act File Number 811-4946 THOMPSON IM FUNDS, INC. (Exact name of registrant as specified in charter) 918 Deming Way
#> 7       UNITED STATES SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-21295 NAME OF REGISTRANT: JPMorgan Trust I ADDRESS OF PRINCIPAL EXECUTIVE OFFICES: 270 Park
#> 8   UNITED STATES SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-21866 NAME OF REGISTRANT: Highland Funds I ADDRESS OF PRINCIPAL EXECUTIVE OFFICES: 200 Crescent
#> 9  UNITED STATES SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-07142 NAME OF REGISTRANT: Highland Funds II ADDRESS OF PRINCIPAL EXECUTIVE OFFICES: 200 Crescent
#> 10                                                                                                                                                                                                               21VIANET GROUP INCORPORATED SPONSORED ADR (EACH REPRESENTS 6 ORDINARY SHARES)ADR90138A1032192
#>                                                                                          parent_href
#> 1                                                                                               <NA>
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5                                                                                               <NA>
#> 6                                                                                               <NA>
#> 7                                                                                               <NA>
#> 8                                                                                               <NA>
#> 9                                                                                               <NA>
#> 10 https://www.sec.gov/Archives/edgar/data/1086619/000156761918001750/xslForm13F_X01/primary_doc.xml
#>                                                                                           index_href
#> 1                                                                                               <NA>
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5                                                                                               <NA>
#> 6                                                                                               <NA>
#> 7                                                                                               <NA>
#> 8                                                                                               <NA>
#> 9                                                                                               <NA>
#> 10 https://www.sec.gov/Archives/edgar/data/1086619/000156761918001750/0001567619-18-001750-index.htm