SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(
  q = "*",
  type = "",
  reverse_order = FALSE,
  count = 10,
  page = 1,
  stemming = TRUE,
  name = "",
  cik = "",
  sic = "",
  from = "",
  to = ""
)

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

# \donttest{
## This can be very slow running
full_text('intel')

#>    filing_date
#> 1   2020-01-08
#> 2   2020-01-08
#> 3   2020-01-08
#> 4   2020-01-08
#> 5   2020-01-08
#> 6   2020-01-07
#> 7   2020-01-07
#> 8   2020-01-07
#> 9   2020-01-07
#> 10  2020-01-07
#>                                                                 name
#> 1                                    N-CSR for SUNAMERICA SERIES INC
#> 2                           N-MFP2 for T Rowe Price Summit Funds Inc
#> 3             N-MFP2 for T Rowe Price Institutional Income Funds Inc
#> 4  INFORMATION TABLE of 13F-HR for ACCURATE INVESTMENT SOLUTIONS INC
#> 5                  INFORMATION TABLE of 13F-HR for Verity Verity LLC
#> 6                            487 for Advisors Disciplined Trust 1981
#> 7                                  N-CSR for Series Portfolios Trust
#> 8                INFORMATION TABLE of 13F-HR for IBM Retirement Fund
#> 9                INFORMATION TABLE of 13F-HR for BURNS J W CO INC NY
#> 10        INFORMATION TABLE of 13F-HR for M R CAPITAL MANAGEMENT INC
#>                                                                                                          href
#> 1                      https://www.sec.gov/Archives/edgar/data/1020861/000110465920002060/a19-22918_1ncsr.htm
#> 2             https://www.sec.gov/Archives/edgar/data/912028/000114554920000459/xslN-MFP2_X01/primary_doc.xml
#> 3            https://www.sec.gov/Archives/edgar/data/1169187/000114554920000471/xslN-MFP2_X01/primary_doc.xml
#> 4             https://www.sec.gov/Archives/edgar/data/1607863/000121465920000206/xslForm13F_X01/infotable.xml
#> 5           https://www.sec.gov/Archives/edgar/data/1542108/000154210820000001/xslForm13F_X01/13f20191231.xml
#> 6                            https://www.sec.gov/Archives/edgar/data/1771775/000089109220000460/e7940_487.htm
#> 7                          https://www.sec.gov/Archives/edgar/data/1650149/000089853120000008/wabrf-ncsra.htm
#> 8       https://www.sec.gov/Archives/edgar/data/1018825/000101882520000001/xslForm13F_X01/13fibmq42019kgb.xml
#> 9    https://www.sec.gov/Archives/edgar/data/1047339/000104733920000001/xslForm13F_X01/12-31-19JWBURNS13F.xml
#> 10 https://www.sec.gov/Archives/edgar/data/1057395/000105739520000001/xslForm13F_X01/mrcapital_12_31_2019.xml
#>                                   company_name     cik  sic
#> 1                        SUNAMERICA SERIES INC 1020861 <NA>
#> 2                T Rowe Price Summit Funds Inc  912028 <NA>
#> 3  T Rowe Price Institutional Income Funds Inc 1169187 <NA>
#> 4            ACCURATE INVESTMENT SOLUTIONS INC 1607863 <NA>
#> 5                            Verity Verity LLC 1542108 <NA>
#> 6              Advisors Disciplined Trust 1981 1771775 <NA>
#> 7                      Series Portfolios Trust 1650149 <NA>
#> 8                          IBM Retirement Fund 1018825 <NA>
#> 9                          BURNS J W CO INC NY 1047339 <NA>
#> 10                  M R CAPITAL MANAGEMENT INC 1057395 <NA>
#>                                                                                                                                                                                                                                                                                                         content
#> 1                         UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number 811-07797 SunAmerica Series, Inc. (Exact name of registrant as specified in charter)
#> 2     N-MFP2 0000912028 XXXXXXXX 2019-12-31 0000912028 S000002151 1 N N 033-50319 T. Rowe Price Associates, Inc. 801-00856 PricewaterhouseCoopers LLP Baltimore MD None T. Rowe Price Services, Inc. 0000703220 084-01079 N N N Prime Y 49 93 643416119.80 662006251.10 698599891.20 761195165.80 1114117130.00
#> 3         N-MFP2 0001169187 XXXXXXXX 2019-12-31 0001169187 S000054621 1 N N 333-84634 T. Rowe Price Associates, Inc. 801-00856 PricewaterhouseCoopers LLP Baltimore MD None T. Rowe Price Services, Inc. 0000703220 084-01079 N N N Prime N 36 50 114869062.90 67080073.27 61740851.29 62505549.58 187867533.60
#> 4   ADT INC COM 00090Q103 2 310 SH SOLE 0 0 310 ABM INDS INC COM 000957100 9 226 SH SOLE 0 0 226 AGCO CORP COM 001084102 11 136 SH SOLE 0 0 136 AGNC INVT CORP COM 00123Q104 146 8285 SH SOLE 0 0 8285 ALPS ETF TR SECTR DIV DOGS 00162Q858 188 4004 SH SOLE 0 0 4004 ALPS ETF TR ALERIAN MLP 00162Q866 34 4000
#> 5      3M CO COM 88579Y101 3705 21001 SH SOLE 0 0 21001 ABBOTT LABS COM 002824100 7701 88666 SH SOLE 0 0 88666 ABBVIE INC COM 00287Y109 5589 63123 SH SOLE 0 0 63123 ACCENTURE PLC IRELAND SHS CLASS A G1151C101 7659 36371 SH SOLE 0 0 36371 AFLAC INC COM 001055102 8986 169866 SH SOLE 0 0 169866 ALASKA AIR
#> 6           1933 Act File No.: 333-234199 1940 Act File No.: 811-21056 CIK No.: 1771775 Securities and Exchange Commission Washington, D.C. 20549 Amendment No. 1 to REGISTRATION STATEMENT ON Form S-6 For Registration under the Securities Act of 1933 of Securities of Unit Investment Trusts Registered on
#> 7       UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number 811-23084 Series Portfolios Trust (Exact name of registrant as specified in charter) 615 East Michigan
#> 8      3M CO COM 88579Y101 1606 9101 SH SOLE 2853 0 6248 ABBOTT LABS COM 002824100 2437 28053 SH SOLE 8787 0 19266 ABBVIE INC COM 00287Y109 2130 24057 SH SOLE 7548 0 16509 ACCENTURE PLC IRELAND SHS CLAS COM G1151C101 2194 10417 SH SOLE 3252 0 7165 ACTIVISION BLIZZARD INC COM 00507V109 776 13055 SH SOLE
#> 9  FIRST TRUST PFD SEC & INC PFD 33739E108 1129 56250 SH SOLE 0 0 0 56250 ISHARES US PFD ETF PFD 464288687 3823 101709 SH SOLE 0 0 0 101709 NUVEEN PFD INCOME PFD 67073b106 133 12700 SH SOLE 0 0 0 12700 3M COMPANY COM 88579Y101 538 3051 SH SOLE 0 0 0 3051 ABBOTT LABS COM 002824100 3786 43587 SH SOLE 0 0
#> 10  CITIZENS FINL GROUP INC COM Stock 174610105 10266 252873 SH SOLE 252873 0 0 CHURCH & DWIGHT INC COM Stock 171340102 17720 251933 SH SOLE 251933 0 0 ISHARES CURRENCY HEDGED MSCI EUROZONE ETF ETF 46434V639 7860 247353 SH SOLE 247353 0 0 COMCAST CORP NEW CL A Stock 20030N101 9744 216589 SH SOLE 216589
#>                                                                                          parent_href
#> 1                                                                                               <NA>
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4  https://www.sec.gov/Archives/edgar/data/1047339/000104733920000001/xslForm13F_X01/primary_doc.xml
#> 5  https://www.sec.gov/Archives/edgar/data/1057395/000105739520000001/xslForm13F_X01/primary_doc.xml
#> 6                                                                                               <NA>
#> 7                                                                                               <NA>
#> 8  https://www.sec.gov/Archives/edgar/data/1018825/000101882520000001/xslForm13F_X01/primary_doc.xml
#> 9  https://www.sec.gov/Archives/edgar/data/1047339/000104733920000001/xslForm13F_X01/primary_doc.xml
#> 10 https://www.sec.gov/Archives/edgar/data/1057395/000105739520000001/xslForm13F_X01/primary_doc.xml
#>                                                                                           index_href
#> 1                                                                                               <NA>
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4  https://www.sec.gov/Archives/edgar/data/1047339/000104733920000001/0001047339-20-000001-index.htm
#> 5  https://www.sec.gov/Archives/edgar/data/1057395/000105739520000001/0001057395-20-000001-index.htm
#> 6                                                                                               <NA>
#> 7                                                                                               <NA>
#> 8  https://www.sec.gov/Archives/edgar/data/1018825/000101882520000001/0001018825-20-000001-index.htm
#> 9  https://www.sec.gov/Archives/edgar/data/1047339/000104733920000001/0001047339-20-000001-index.htm
#> 10 https://www.sec.gov/Archives/edgar/data/1057395/000105739520000001/0001057395-20-000001-index.htm

# }