SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

## This can be very slow running
full_text('intel')

#>    filing_date                                                          name
#> 1   2018-06-22                             424B2 for GOLDMAN SACHS GROUP INC
#> 2   2018-06-22                   N-CSRS for LEGG MASON PARTNERS EQUITY TRUST
#> 3   2018-06-22                              11-K for GOLDMAN SACHS GROUP INC
#> 4   2018-06-22                                        10-K for Medtronic plc
#> 5   2018-06-21                                    424B3 for Yatra Online Inc
#> 6   2018-06-21 INFORMATION TABLE of 13F-HR/A for CANYON CAPITAL ADVISORS LLC
#> 7   2018-06-21                                              424B2 for UBS AG
#> 8   2018-06-21                                              424B2 for UBS AG
#> 9   2018-06-21                                       FWP for HSBC USA INC MD
#> 10  2018-06-21                                     10-K for BARNES NOBLE INC
#>                                                                                                      href
#> 1                         https://www.sec.gov/Archives/edgar/data/886982/000114036118029503/form424b2.htm
#> 2                     https://www.sec.gov/Archives/edgar/data/880366/000119312518199725/d569204dncsrs.htm
#> 3                       https://www.sec.gov/Archives/edgar/data/886982/000119312518199799/d612714d11k.htm
#> 4                  https://www.sec.gov/Archives/edgar/data/1613103/000161310318000024/mdt-2018427x10k.htm
#> 5                   https://www.sec.gov/Archives/edgar/data/1516899/000104746918004644/a2236101z424b3.htm
#> 6  https://www.sec.gov/Archives/edgar/data/1074034/000108514618001603/xslForm13F_X01/form13fInfoTable.xml
#> 7                  https://www.sec.gov/Archives/edgar/data/1114446/000111444618007464/tc4570597i_0fwp.htm
#> 8                  https://www.sec.gov/Archives/edgar/data/1114446/000111444618007465/tc4570597f_1fwp.htm
#> 9                       https://www.sec.gov/Archives/edgar/data/83246/000114420418035070/tv496888_fwp.htm
#> 10                      https://www.sec.gov/Archives/edgar/data/890491/000119312518198716/d590891d10k.htm
#>                        company_name     cik  sic
#> 1           GOLDMAN SACHS GROUP INC  886982 6211
#> 2  LEGG MASON PARTNERS EQUITY TRUST  880366 <NA>
#> 3           GOLDMAN SACHS GROUP INC  886982 6211
#> 4                     Medtronic plc 1613103 3845
#> 5                  Yatra Online Inc 1516899 4700
#> 6       CANYON CAPITAL ADVISORS LLC 1074034 <NA>
#> 7                            UBS AG 1114446 6021
#> 8                            UBS AG 1114446 6021
#> 9                   HSBC USA INC MD   83246 6021
#> 10                 BARNES NOBLE INC  890491 5940
#>                                                                                                                                                                                                                                                                                                        content
#> 1    Filed Pursuant to Rule 424(b)(2) Registration Statement No. 333-219206 The information in this preliminary pricing supplement is not complete and may be changed. This preliminary pricing supplement is not an offer to sell nor does it seek an offer to buy these securities in any jurisdiction where
#> 2                           UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number 811-06444 Legg Mason Partners Equity Trust (Exact name of registrant as specified
#> 3                               SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 FORM 11-K ANNUAL REPORT PURSUANT TO SECTION 15(d) OF THE SECURITIES EXCHANGE ACT OF 1934 Annual report pursuant to Section 15(d) of the Securities Exchange Act of 1934 For the fiscal year ended December 31, 2017.
#> 4           UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM 10-K x Annual report pursuant to section 13 or 15(d) of the Securities Exchange Act of 1934. For the fiscal year ended April 27, 2018. o Transition report pursuant to Section 13 or 15(d) of the Securities Exchange
#> 5    Use these links to rapidly review the document Filed Pursuant to Rule 424(b)(3) Registration No. 333-224661 The information in this preliminary prospectus supplement is not complete and may be changed. This preliminary prospectus supplement and the accompanying prospectus are not an offer to sell
#> 6  ACORDA THERAPEUTICS INC NOTE 1.750% 6/1 00484MAA4 26153 28775000 PRN DFND 28775000 0 0 AETNA INC NEW COM 00817Y108 105625 625000 SH DFND 625000 0 0 ALLY FINL INC COM 02005N100 104588 3852224 SH DFND 3852224 0 0 ALTABA INC COM 021346101 3204342 43278522 SH DFND 43278522 0 0 ARCONIC INC COM 03965L100
#> 7                                                   Underlying Asset The common stock of Intel Corporation Call Feature The Securities will be called if the closing price of the underlying asset on any observation date (including the final valuation date) is equal to or greater than the initial price.
#> 8                                                   Underlying Asset The common stock of Intel Corporation Call Feature The Securities will be called if the closing price of the underlying asset on any observation date (including the final valuation date) is equal to or greater than the initial price.
#> 9                                ISSUER FREE WRITING PROSPECTUS Filed Pursuant to Rule 433 Registration Statement No. 333-223208 Dated June 21, 2018 HSBC USA Inc. Airbag Autocallable Yield Notes $ Notes Linked to the Common Stock of Intel Corporation due on or about June 28, 2019 $ Notes Linked to the
#> 10                      UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM 10-K ANNUAL REPORT PURSUANT TO SECTION 13 OR 15(d) OF THE SECURITIES EXCHANGE ACT OF 1934 For the fiscal year ended April 28, 2018 OR TRANSITION REPORT PURSUANT TO SECTION 13 OR 15(d) OF THE SECURITIES
#>                                                                                          parent_href
#> 1                                                                                               <NA>
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5                                                                                               <NA>
#> 6  https://www.sec.gov/Archives/edgar/data/1074034/000108514618001603/xslForm13F_X01/primary_doc.xml
#> 7                                                                                               <NA>
#> 8                                                                                               <NA>
#> 9                                                                                               <NA>
#> 10                                                                                              <NA>
#>                                                                                           index_href
#> 1                                                                                               <NA>
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5                                                                                               <NA>
#> 6  https://www.sec.gov/Archives/edgar/data/1074034/000108514618001603/0001085146-18-001603-index.htm
#> 7                                                                                               <NA>
#> 8                                                                                               <NA>
#> 9                                                                                               <NA>
#> 10                                                                                              <NA>