SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

## This can be very slow running
full_text('intel')

#>    filing_date                                         name
#> 1   2018-08-23                 N-PX for FIDELITY CONTRAFUND
#> 2   2018-08-23         N-PX for FIDELITY DESTINY PORTFOLIOS
#> 3   2018-08-23        N-PX for FIDELITY SCHOOL STREET TRUST
#> 4   2018-08-23            N-PX for FIDELITY SECURITIES FUND
#> 5   2018-08-23    N-PX for VARIABLE INSURANCE PRODUCTS FUND
#> 6   2018-08-23 N-PX for VARIABLE INSURANCE PRODUCTS FUND II
#> 7   2018-08-23        N-PX for FIDELITY SUMMER STREET TRUST
#> 8   2018-08-23            N-PX for FIDELITY COVINGTON TRUST
#> 9   2018-08-23           N-PX for FIDELITY ADVISOR SERIES I
#> 10  2018-08-23         N-PX for FIDELITY ADVISOR SERIES VII
#>                                                                                      href
#> 1   https://www.sec.gov/Archives/edgar/data/24238/000079542218000037/npx811-01400_114.htm
#> 2   https://www.sec.gov/Archives/edgar/data/35331/000079542218000038/npx811-01796_100.htm
#> 3  https://www.sec.gov/Archives/edgar/data/215829/000079542218000041/npx811-02676_102.htm
#> 4  https://www.sec.gov/Archives/edgar/data/754510/000079542218000042/npx811-04118_111.htm
#> 5  https://www.sec.gov/Archives/edgar/data/356494/000079542218000043/npx811-03329_103.htm
#> 6  https://www.sec.gov/Archives/edgar/data/831016/000079542218000048/npx811-05511_100.htm
#> 7  https://www.sec.gov/Archives/edgar/data/225322/000079542218000054/npx811-02737_106.htm
#> 8   https://www.sec.gov/Archives/edgar/data/945908/000079542218000058/npx811-07319_35.htm
#> 9  https://www.sec.gov/Archives/edgar/data/722574/000088019518000124/npx811-03785_117.htm
#> 10  https://www.sec.gov/Archives/edgar/data/315700/000088019518000125/npx811-03010_98.htm
#>                           company_name    cik  sic
#> 1                  FIDELITY CONTRAFUND  24238 <NA>
#> 2          FIDELITY DESTINY PORTFOLIOS  35331 <NA>
#> 3         FIDELITY SCHOOL STREET TRUST 215829 <NA>
#> 4             FIDELITY SECURITIES FUND 754510 <NA>
#> 5     VARIABLE INSURANCE PRODUCTS FUND 356494 <NA>
#> 6  VARIABLE INSURANCE PRODUCTS FUND II 831016 <NA>
#> 7         FIDELITY SUMMER STREET TRUST 225322 <NA>
#> 8             FIDELITY COVINGTON TRUST 945908 <NA>
#> 9            FIDELITY ADVISOR SERIES I 722574 <NA>
#> 10         FIDELITY ADVISOR SERIES VII 315700 <NA>
#>                                                                                                                                                                                                                                                                                                         content
#> 1    FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-01400 EXACT NAME OF REGISTRANT AS SPECIFIED IN CHARTER: Fidelity Contrafund 245 SUMMER STREET, BOSTON, MA 02210 (ADDRESS OF PRINCIPAL EXECUTIVE OFFICES)(ZIP CODE) MARC
#> 2       FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-01796 EXACT NAME OF REGISTRANT AS SPECIFIED IN CHARTER: Fidelity Destiny Portfolios 245 SUMMER STREET, BOSTON, MA 02210 (ADDRESS OF PRINCIPAL EXECUTIVE OFFICES)(ZIP
#> 3      FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-02676 EXACT NAME OF REGISTRANT AS SPECIFIED IN CHARTER: Fidelity School Street Trust 245 SUMMER STREET, BOSTON, MA 02210 (ADDRESS OF PRINCIPAL EXECUTIVE OFFICES)(ZIP
#> 4    FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-04118 EXACT NAME OF REGISTRANT AS SPECIFIED IN CHARTER: Fidelity Securities Fund 245 SUMMER STREET, BOSTON, MA 02210 (ADDRESS OF PRINCIPAL EXECUTIVE OFFICES)(ZIP CODE)
#> 5  FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-03329 EXACT NAME OF REGISTRANT AS SPECIFIED IN CHARTER: Variable Insurance Products Fund 245 SUMMER STREET, BOSTON, MA 02210 (ADDRESS OF PRINCIPAL EXECUTIVE OFFICES)(ZIP
#> 6            FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-05511 EXACT NAME OF REGISTRANT AS SPECIFIED IN CHARTER: Variable Insurance Products Fund II 245 SUMMER STREET, BOSTON, MA 02210 (ADDRESS OF PRINCIPAL EXECUTIVE
#> 7      FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-02737 EXACT NAME OF REGISTRANT AS SPECIFIED IN CHARTER: Fidelity Summer Street Trust 245 SUMMER STREET, BOSTON, MA 02210 (ADDRESS OF PRINCIPAL EXECUTIVE OFFICES)(ZIP
#> 8    FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-07319 EXACT NAME OF REGISTRANT AS SPECIFIED IN CHARTER: Fidelity Covington Trust 245 SUMMER STREET, BOSTON, MA 02210 (ADDRESS OF PRINCIPAL EXECUTIVE OFFICES)(ZIP CODE)
#> 9   FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-03785 EXACT NAME OF REGISTRANT AS SPECIFIED IN CHARTER: Fidelity Advisor Series I 245 SUMMER STREET, BOSTON, MA 02210 (ADDRESS OF PRINCIPAL EXECUTIVE OFFICES)(ZIP CODE)
#> 10      FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY INVESTMENT COMPANY ACT FILE NUMBER: 811-03010 EXACT NAME OF REGISTRANT AS SPECIFIED IN CHARTER: Fidelity Advisor Series VII 245 SUMMER STREET, BOSTON, MA 02210 (ADDRESS OF PRINCIPAL EXECUTIVE OFFICES)(ZIP
#>    parent_href index_href
#> 1         <NA>       <NA>
#> 2         <NA>       <NA>
#> 3         <NA>       <NA>
#> 4         <NA>       <NA>
#> 5         <NA>       <NA>
#> 6         <NA>       <NA>
#> 7         <NA>       <NA>
#> 8         <NA>       <NA>
#> 9         <NA>       <NA>
#> 10        <NA>       <NA>