SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(
  q = "*",
  type = "",
  reverse_order = FALSE,
  count = 10,
  page = 1,
  stemming = TRUE,
  name = "",
  cik = "",
  sic = "",
  from = "",
  to = ""
)

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

# \donttest{
## This can be very slow running
full_text('intel')

#>    filing_date                                                        name
#> 1   2020-04-24 INFORMATION TABLE of 13F-HR for ARGUS INVESTORS COUNSEL INC
#> 2   2020-04-24                                         10-Q for INTEL CORP
#> 3   2020-04-24                              EX-31.1 of 10-Q for INTEL CORP
#> 4   2020-04-24                              EX-31.2 of 10-Q for INTEL CORP
#> 5   2020-04-24                              EX-32.1 of 10-Q for INTEL CORP
#> 6   2020-04-24                              EX-10.1 of 10-Q for INTEL CORP
#> 7   2020-04-24                              EX-10.3 of 10-Q for INTEL CORP
#> 8   2020-04-24                                  XML of 10-Q for INTEL CORP
#> 9   2020-04-24                                  XML of 10-Q for INTEL CORP
#> 10  2020-04-24                                  XML of 10-Q for INTEL CORP
#>                                                                                           href
#> 1  https://www.sec.gov/Archives/edgar/data/7195/000000719520000002/xslForm13F_X01/inftable.xml
#> 2  https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a0328202010qdocument-u.htm
#> 3      https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a03282020ex311-use.htm
#> 4      https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a03282020ex312-use.htm
#> 5      https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a03282020ex321-use.htm
#> 6       https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a03282020q1-ex101.htm
#> 7       https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a03282020q1-ex103.htm
#> 8           https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/FilingSummary.xml
#> 9                      https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/R1.htm
#> 10                    https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/R10.htm
#>                   company_name   cik  sic
#> 1  ARGUS INVESTORS COUNSEL INC  7195 <NA>
#> 2                   INTEL CORP 50863 3674
#> 3                   INTEL CORP 50863 3674
#> 4                   INTEL CORP 50863 3674
#> 5                   INTEL CORP 50863 3674
#> 6                   INTEL CORP 50863 3674
#> 7                   INTEL CORP 50863 3674
#> 8                   INTEL CORP 50863 3674
#> 9                   INTEL CORP 50863 3674
#> 10                  INTEL CORP 50863 3674
#>                                                                                                                                                                                                                                                                                                       content
#> 1  3M Co COM 88579Y101 3145 23040 SH SOLE 17140 0 5900 Adobe Inc COM 00724F101 4215 13245 SH SOLE 9195 0 4050 Air Prods & Chems Inc COM 009158106 3457 17320 SH SOLE 13250 0 4070 Alphabet Inc COM 02079K305 4128 3553 SH SOLE 2420 0 1133 Amazon.com Inc COM 023135106 4751 2437 SH SOLE 1617 0 820 American
#> 2   Document false -12-26 Q1 2020 2020-03-28 0000050863 73321000000 75686000000 0.63 0.66 4290000000 4234000000 4290000000 4234000000 0.0035 0.0008 0.0270 0.0240 0.0373 0.0245 0.0325 0.0410 0.0480 0.0400 0.0310 0.0270 0.0288 0.0410 0.0315 0.0325 0.0245 0.0330 0.0490 0.0370 0.0270 0.0235 0.0170 0.0410
#> 3                                                                                                                                                                                                                                 I have reviewed this quarterly report on Form 10-Q of Intel Corporation; 2.
#> 4                                                                                                                                                                                                                                 I have reviewed this quarterly report on Form 10-Q of Intel Corporation; 2.
#> 5        This written statement is being furnished to the Securities and Exchange Commission as an exhibit to such Form 10-Q. A signed original of this statement has been provided to Intel and will be retained by Intel and furnished to the Securities and Exchange Commission or its staff upon request.
#> 6                                                                                                                                                                                           Those components, as explained below in this Section 4 are: (i) the Intel Relative TSR and (ii) Intel EPS Growth.
#> 7                                                         Exhibit 10.3 INTEL CORPORATION SHELTERED EMPLOYEE RETIREMENT PLAN PLUS (As Amended and Restated Effective January 1, 2020) INTEL CORPORATION SHELTERED EMPLOYEE RETIREMENT PLAN PLUS (As Amended and Restated Effective January 1, 2020) SECTION 1.
#> 8                          3.20.1 html 381 319 1 false 104 0 false 5 false false R1.htm 0001000 - Document - Cover Page Cover Page Sheet http:/www.intel.com/role/CoverPageCoverPage Cover Page Cover Page Cover 1 false false R2.htm 1001000 - Statement - Consolidated Condensed Statements of Income Sheet
#> 9       v3.20.1 Cover Page Cover Page shares in Millions 3 Months Ended Mar. 28, 2020 shares 10-Q true false Mar. 28, 2020 000-06217 INTEL CORPORATION 0000050863 -12-26 2020 Q1 false DE 94-1672743 2200 Mission College Boulevard, Santa Clara, CA 95054-1549 (408) 765-8080 Common stock, $0.001 par value
#> 10                                                                                                                                                                                          Except for these differences, the accounting policies for segment reporting are the same as for Intel as a whole.
#>                                                                                       parent_href
#> 1  https://www.sec.gov/Archives/edgar/data/7195/000000719520000002/xslForm13F_X01/primary_doc.xml
#> 2                                                                                            <NA>
#> 3     https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a0328202010qdocument-u.htm
#> 4     https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a0328202010qdocument-u.htm
#> 5     https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a0328202010qdocument-u.htm
#> 6     https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a0328202010qdocument-u.htm
#> 7     https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a0328202010qdocument-u.htm
#> 8     https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a0328202010qdocument-u.htm
#> 9     https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/a0328202010qdocument-u.htm
#> 10 https://www.sec.gov/Archives/edgar/data/7195/000000719520000002/xslForm13F_X01/primary_doc.xml
#>                                                                                         index_href
#> 1   https://www.sec.gov/Archives/edgar/data/7195/000000719520000002/0000007195-20-000002-index.htm
#> 2                                                                                             <NA>
#> 3  https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/0000050863-20-000017-index.htm
#> 4  https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/0000050863-20-000017-index.htm
#> 5  https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/0000050863-20-000017-index.htm
#> 6  https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/0000050863-20-000017-index.htm
#> 7  https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/0000050863-20-000017-index.htm
#> 8  https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/0000050863-20-000017-index.htm
#> 9  https://www.sec.gov/Archives/edgar/data/50863/000005086320000017/0000050863-20-000017-index.htm
#> 10  https://www.sec.gov/Archives/edgar/data/7195/000000719520000002/0000007195-20-000002-index.htm

# }