SEC Full-Text Search
Source: R/full_text.R

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

## This can be very slow running
full_text('intel')

#>    filing_date                                                             name
#> 1   2018-04-23               INFORMATION TABLE of 13F-HR for MEYER HANDELMAN CO
#> 2   2018-04-23                          40-24B2 for SIT MID CAP GROWTH FUND INC
#> 3   2018-04-23   INFORMATION TABLE of 13F-HR for PERKINS CAPITAL MANAGEMENT INC
#> 4   2018-04-23 INFORMATION TABLE of 13F-HR for CAPITAL ONE NATIONAL ASSOCIATION
#> 5   2018-04-23        INFORMATION TABLE of 13F-HR for EXCALIBUR MANAGEMENT CORP
#> 6   2018-04-23                               DEF 14A for COLUMBIA SPORTSWEAR CO
#> 7   2018-04-23    INFORMATION TABLE of 13F-HR for Brookstone Capital Management
#> 8   2018-04-23          INFORMATION TABLE of 13F-HR for Selway Asset Management
#> 9   2018-04-23          INFORMATION TABLE of 13F-HR for Beech Hill Advisors Inc
#> 10  2018-04-23                INFORMATION TABLE of 13F-HR for St Louis Trust Co
#>                                                                                                      href
#> 1            https://www.sec.gov/Archives/edgar/data/45319/000004531918000002/xslForm13F_X01/inftable.xml
#> 2             https://www.sec.gov/Archives/edgar/data/356787/000035678718000022/SitMFFundPages4232018.pdf
#> 3      https://www.sec.gov/Archives/edgar/data/884300/000088430018000006/xslForm13F_X01/2018-13F-1qtr.xml
#> 4    https://www.sec.gov/Archives/edgar/data/47288/000095012318003721/xslForm13F_X01/form13fInfoTable.xml
#> 5    https://www.sec.gov/Archives/edgar/data/1009003/000100900318000002/xslForm13F_X01/Excal13F033118.xml
#> 6                   https://www.sec.gov/Archives/edgar/data/1050797/000105079718000019/colm2017def14a.htm
#> 7  https://www.sec.gov/Archives/edgar/data/1599584/000106299318001772/xslForm13F_X01/form13fInfoTable.xml
#> 8          https://www.sec.gov/Archives/edgar/data/1080523/000108052318000007/xslForm13F_X01/inftable.xml
#> 9  https://www.sec.gov/Archives/edgar/data/1356202/000108514618001126/xslForm13F_X01/form13fInfoTable.xml
#> 10 https://www.sec.gov/Archives/edgar/data/1661144/000108514618001128/xslForm13F_X01/form13fInfoTable.xml
#>                        company_name     cik  sic
#> 1                MEYER HANDELMAN CO   45319 <NA>
#> 2       SIT MID CAP GROWTH FUND INC  356787 <NA>
#> 3    PERKINS CAPITAL MANAGEMENT INC  884300 <NA>
#> 4  CAPITAL ONE NATIONAL ASSOCIATION   47288 6021
#> 5         EXCALIBUR MANAGEMENT CORP 1009003 <NA>
#> 6            COLUMBIA SPORTSWEAR CO 1050797 2300
#> 7     Brookstone Capital Management 1599584 <NA>
#> 8           Selway Asset Management 1080523 <NA>
#> 9           Beech Hill Advisors Inc 1356202 <NA>
#> 10                St Louis Trust Co 1661144 <NA>
#>                                                                                                                                                                                                                                                                                                         content
#> 1     3M COMPANY COM 88579Y101 45462 207099 SH SOLE 203899 3200 0 3M COMPANY COM 88579Y101 702 3200 SH DFND 02 3200 3200 0 3M COMPANY COM 88579Y101 66 300 SH OTR 300 0 0 3M COMPANY COM 88579Y101 702 3200 SH OTR 02 3200 3200 0 ABBOTT LABS INC COM COM 002824100 5728 95593 SH SOLE 94516 1077 0 ABBOTT LABS
#> 2     [iJ Sit Mutual Funds Mutu al Funds Account Management Retirement and Education Plans Sit Balanced Fund Forms and Publications Fi rm Information One Step Diversification FASTTRACK. Daily Pnces I Account Access I Mutual Fund Performance I Contact Us SIT BALANCED FUND Ticker: SI BAX Overview Returns
#> 3         3M CO COM 88579Y101 565 2576 SH SOLE 2576 0 0 ABBOTT LABS COM 002824100 4851 80950 SH SOLE 34679 0 46271 ADVANCED MICRO DEVICES INC COM 007903107 160 15900 SH SOLE 12900 0 3000 AFLAC INC COM 001055102 1593 36400 SH SOLE 24900 0 11500 AIR INDS GROUP COM NEW 00912N205 25 16000 SH SOLE 0 0 16000
#> 4                                                                                                                                                                                                                                                                                       AAR CORPCOM000361105753
#> 5    3M Co COM 88579y101 684 3114 SH SOLE 3114 0 0 ABBOTT LABS COM 002824100 1343 22410 SH SOLE 22410 0 0 ALPHABET CL A COM 02079K305 1026 989 SH SOLE 989 0 0 ALTRIA GROUP COM 02209s103 795 12762 SH SOLE 12762 0 0 AMAZON.COM INC COM 023135106 414 286 SH SOLE 286 0 0 AMDOCS LTD COM G02602103 347 5200 SH
#> 6                                                                                               UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 SCHEDULE 14A INFORMATION Proxy Statement Pursuant to Section 14(a) of the Securities Exchange Act of 1934 (Amendment No. ) Filed by the
#> 7  AFLAC INC COM 001055102 277 6329 SH SOLE no 0 0 6329 AGNC INVT CORP COM 00123Q104 418 22070 SH SOLE no 0 0 22070 AES CORP COM 00130H105 5795 509697 SH SOLE no 0 0 509697 AMC ENTMT HLDGS INC CL A COM 00165C104 403 28680 SH SOLE no 0 0 28680 AT&T INC COM 00206R102 8711 244348 SH SOLE no 0 0 244348 ARK
#> 8  2X BDC Index COM 90267b765 9853 669340 SH SOLE 669340 0 0 Abbvie COM 00287y109 1900 20075 SH SOLE 20075 0 0 Advance Auto Parts COM 00751Y106 3811 32150 SH SOLE 32150 0 0 American Express COM 025816109 1362 14600 SH SOLE 14600 0 0 AmerisourceBergen COM 03073e105 2501 29015 SH SOLE 29015 0 0 Amgen Inc
#> 9    ABBOTT LABS COM 002824100 790 13189 SH SOLE 0 0 13189 ABBVIE INC COM 00287Y109 5261 55582 SH SOLE 5500 0 50082 ABIOMED INC COM 003654100 291 1000 SH SOLE 0 0 1000 ALBEMARLE CORP COM 012653101 2277 24555 SH SOLE 3250 0 21305 ALPHABET INC CAP STK CL C 02079K107 360 349 SH SOLE 100 0 249 ALPHABET INC
#> 10    3M CO COM 88579Y101 275 1251 SH SOLE 1251 0 0 3M CO COM 88579Y101 648 2950 SH OTR 2950 0 0 ABBOTT LABS COM 002824100 836 13955 SH OTR 13955 0 0 ABBOTT LABS COM 002824100 90 1500 SH SOLE 1500 0 0 ABBVIE INC COM 00287Y109 1132 11964 SH OTR 11964 0 0 ABBVIE INC COM 00287Y109 95 1000 SH SOLE 1000 0 0
#>                                                                                          parent_href
#> 1    https://www.sec.gov/Archives/edgar/data/45319/000004531918000002/xslForm13F_X01/primary_doc.xml
#> 2                                                                                               <NA>
#> 3    https://www.sec.gov/Archives/edgar/data/47288/000095012318003721/xslForm13F_X01/primary_doc.xml
#> 4  https://www.sec.gov/Archives/edgar/data/1009003/000100900318000002/xslForm13F_X01/primary_doc.xml
#> 5  https://www.sec.gov/Archives/edgar/data/1599584/000106299318001772/xslForm13F_X01/primary_doc.xml
#> 6                                                                                               <NA>
#> 7  https://www.sec.gov/Archives/edgar/data/1356202/000108514618001126/xslForm13F_X01/primary_doc.xml
#> 8  https://www.sec.gov/Archives/edgar/data/1661144/000108514618001128/xslForm13F_X01/primary_doc.xml
#> 9    https://www.sec.gov/Archives/edgar/data/45319/000004531918000002/xslForm13F_X01/primary_doc.xml
#> 10  https://www.sec.gov/Archives/edgar/data/884300/000088430018000006/xslForm13F_X01/primary_doc.xml
#>                                                                                           index_href
#> 1    https://www.sec.gov/Archives/edgar/data/45319/000004531918000002/0000045319-18-000002-index.htm
#> 2                                                                                               <NA>
#> 3    https://www.sec.gov/Archives/edgar/data/47288/000095012318003721/0000950123-18-003721-index.htm
#> 4  https://www.sec.gov/Archives/edgar/data/1009003/000100900318000002/0001009003-18-000002-index.htm
#> 5  https://www.sec.gov/Archives/edgar/data/1599584/000106299318001772/0001062993-18-001772-index.htm
#> 6                                                                                               <NA>
#> 7  https://www.sec.gov/Archives/edgar/data/1356202/000108514618001126/0001085146-18-001126-index.htm
#> 8  https://www.sec.gov/Archives/edgar/data/1661144/000108514618001128/0001085146-18-001128-index.htm
#> 9    https://www.sec.gov/Archives/edgar/data/45319/000004531918000002/0000045319-18-000002-index.htm
#> 10  https://www.sec.gov/Archives/edgar/data/884300/000088430018000006/0000884300-18-000006-index.htm