SEC Full-Text Search
Source: R/full_text.R

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

## This can be very slow running
full_text('intel')

#>    filing_date
#> 1   2018-04-23
#> 2   2018-04-23
#> 3   2018-04-23
#> 4   2018-04-23
#> 5   2018-04-23
#> 6   2018-04-23
#> 7   2018-04-23
#> 8   2018-04-23
#> 9   2018-04-23
#> 10  2018-04-23
#>                                                                                        name
#> 1                            INFORMATION TABLE of 13F-HR for PERKINS CAPITAL MANAGEMENT INC
#> 2                          INFORMATION TABLE of 13F-HR for CAPITAL ONE NATIONAL ASSOCIATION
#> 3                                 INFORMATION TABLE of 13F-HR for EXCALIBUR MANAGEMENT CORP
#> 4                                   INFORMATION TABLE of 13F-HR for Beech Hill Advisors Inc
#> 5                                         INFORMATION TABLE of 13F-HR for St Louis Trust Co
#> 6                                              EX-99.1 of 8-K for Mellanox Technologies Ltd
#> 7                                          EX-99.1 of DEFA14A for Mellanox Technologies Ltd
#> 8                                                              S-4/A for SALESFORCE COM INC
#> 9  N-CSRS for MFS INSTITUTIONAL MONEY MARKET PORTFOLIO A SERIES OF THE MFS SERIES TRUST XIV
#> 10                                INFORMATION TABLE of 13F-HR for BALASA DINVERNO FOLTZ LLC
#>                                                                                                      href
#> 1      https://www.sec.gov/Archives/edgar/data/884300/000088430018000006/xslForm13F_X01/2018-13F-1qtr.xml
#> 2    https://www.sec.gov/Archives/edgar/data/47288/000095012318003721/xslForm13F_X01/form13fInfoTable.xml
#> 3    https://www.sec.gov/Archives/edgar/data/1009003/000100900318000002/xslForm13F_X01/Excal13F033118.xml
#> 4  https://www.sec.gov/Archives/edgar/data/1356202/000108514618001126/xslForm13F_X01/form13fInfoTable.xml
#> 5  https://www.sec.gov/Archives/edgar/data/1661144/000108514618001128/xslForm13F_X01/form13fInfoTable.xml
#> 6                    https://www.sec.gov/Archives/edgar/data/1356104/000119312518125775/d572926dex991.htm
#> 7                    https://www.sec.gov/Archives/edgar/data/1356104/000119312518125776/d572926dex991.htm
#> 8                      https://www.sec.gov/Archives/edgar/data/1108524/000119312518125938/d552163ds4a.htm
#> 9                    https://www.sec.gov/Archives/edgar/data/1392417/000119312518126260/d523671dncsrs.htm
#> 10          https://www.sec.gov/Archives/edgar/data/1308757/000130875718000004/xslForm13F_X01/Mar2018.xml
#>                                                                     company_name
#> 1                                                 PERKINS CAPITAL MANAGEMENT INC
#> 2                                               CAPITAL ONE NATIONAL ASSOCIATION
#> 3                                                      EXCALIBUR MANAGEMENT CORP
#> 4                                                        Beech Hill Advisors Inc
#> 5                                                              St Louis Trust Co
#> 6                                                      Mellanox Technologies Ltd
#> 7                                                      Mellanox Technologies Ltd
#> 8                                                             SALESFORCE COM INC
#> 9  MFS INSTITUTIONAL MONEY MARKET PORTFOLIO A SERIES OF THE MFS SERIES TRUST XIV
#> 10                                                     BALASA DINVERNO FOLTZ LLC
#>        cik  sic
#> 1   884300 <NA>
#> 2    47288 6021
#> 3  1009003 <NA>
#> 4  1356202 <NA>
#> 5  1661144 <NA>
#> 6  1356104 3674
#> 7  1356104 3674
#> 8  1108524 7372
#> 9  1392417 <NA>
#> 10 1308757 <NA>
#>                                                                                                                                                                                                                                                                                                         content
#> 1         3M CO COM 88579Y101 565 2576 SH SOLE 2576 0 0 ABBOTT LABS COM 002824100 4851 80950 SH SOLE 34679 0 46271 ADVANCED MICRO DEVICES INC COM 007903107 160 15900 SH SOLE 12900 0 3000 AFLAC INC COM 001055102 1593 36400 SH SOLE 24900 0 11500 AIR INDS GROUP COM NEW 00912N205 25 16000 SH SOLE 0 0 16000
#> 2                                                                                                                                                                                                                                                                                       AAR CORPCOM000361105753
#> 3    3M Co COM 88579y101 684 3114 SH SOLE 3114 0 0 ABBOTT LABS COM 002824100 1343 22410 SH SOLE 22410 0 0 ALPHABET CL A COM 02079K305 1026 989 SH SOLE 989 0 0 ALTRIA GROUP COM 02209s103 795 12762 SH SOLE 12762 0 0 AMAZON.COM INC COM 023135106 414 286 SH SOLE 286 0 0 AMDOCS LTD COM G02602103 347 5200 SH
#> 4    ABBOTT LABS COM 002824100 790 13189 SH SOLE 0 0 13189 ABBVIE INC COM 00287Y109 5261 55582 SH SOLE 5500 0 50082 ABIOMED INC COM 003654100 291 1000 SH SOLE 0 0 1000 ALBEMARLE CORP COM 012653101 2277 24555 SH SOLE 3250 0 21305 ALPHABET INC CAP STK CL C 02079K107 360 349 SH SOLE 100 0 249 ALPHABET INC
#> 5     3M CO COM 88579Y101 275 1251 SH SOLE 1251 0 0 3M CO COM 88579Y101 648 2950 SH OTR 2950 0 0 ABBOTT LABS COM 002824100 836 13955 SH OTR 13955 0 0 ABBOTT LABS COM 002824100 90 1500 SH SOLE 1500 0 0 ABBVIE INC COM 00287Y109 1132 11964 SH OTR 11964 0 0 ABBVIE INC COM 00287Y109 95 1000 SH SOLE 1000 0 0
#> 6                                              Exhibit 99.1 Corporate Update April 23, 2018 1 Disclaimers These slides and the accompanying oral presentation contain forward-looking statements and information. The use of words such as may , might , will , should , expect , plan , anticipate , believe ,
#> 7                                              Exhibit 99.1 Corporate Update April 23, 2018 1 Disclaimers These slides and the accompanying oral presentation contain forward-looking statements and information. The use of words such as may , might , will , should , expect , plan , anticipate , believe ,
#> 8                           As filed with the Securities and Exchange Commission on April 23, 2018 Registration No. 333-224067 UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 Amendment No. 1 To FORM S-4 REGISTRATION STATEMENT UNDER THE SECURITIES ACT OF 1933 SALESFORCE.COM, INC.
#> 9     UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number 811-22033 MFS SERIES TRUST XIV (Exact name of registrant as specified in charter) 111 Huntington Avenue,
#> 10 Guggenheim S&P 500 Pure Growth GUG S&P500 PU GR 78355W403 967 8807 SH SOLE 0 0 8807 0 Guggenheim S&P Small Cap 600 Pure Growth GUG S&P SC600 PG 78355W809 266 2313 SH SOLE 0 0 2313 0 Guggenheim S&P Small Cap 600 Pure Value GUG S&P SC600 PV 78355W700 278 3957 SH SOLE 0 0 3957 0 Health Fund Sector SPDR
#>                                                                                          parent_href
#> 1   https://www.sec.gov/Archives/edgar/data/884300/000088430018000006/xslForm13F_X01/primary_doc.xml
#> 2    https://www.sec.gov/Archives/edgar/data/47288/000095012318003721/xslForm13F_X01/primary_doc.xml
#> 3  https://www.sec.gov/Archives/edgar/data/1009003/000100900318000002/xslForm13F_X01/primary_doc.xml
#> 4  https://www.sec.gov/Archives/edgar/data/1356202/000108514618001126/xslForm13F_X01/primary_doc.xml
#> 5  https://www.sec.gov/Archives/edgar/data/1661144/000108514618001128/xslForm13F_X01/primary_doc.xml
#> 6                  https://www.sec.gov/Archives/edgar/data/1356104/000119312518125775/d572926d8k.htm
#> 7                  https://www.sec.gov/Archives/edgar/data/1356104/000119312518125776/d572926d8k.htm
#> 8                                                                                               <NA>
#> 9                                                                                               <NA>
#> 10   https://www.sec.gov/Archives/edgar/data/47288/000095012318003721/xslForm13F_X01/primary_doc.xml
#>                                                                                           index_href
#> 1   https://www.sec.gov/Archives/edgar/data/884300/000088430018000006/0000884300-18-000006-index.htm
#> 2    https://www.sec.gov/Archives/edgar/data/47288/000095012318003721/0000950123-18-003721-index.htm
#> 3  https://www.sec.gov/Archives/edgar/data/1009003/000100900318000002/0001009003-18-000002-index.htm
#> 4  https://www.sec.gov/Archives/edgar/data/1356202/000108514618001126/0001085146-18-001126-index.htm
#> 5  https://www.sec.gov/Archives/edgar/data/1661144/000108514618001128/0001085146-18-001128-index.htm
#> 6  https://www.sec.gov/Archives/edgar/data/1356104/000119312518125775/0001193125-18-125775-index.htm
#> 7  https://www.sec.gov/Archives/edgar/data/1356104/000119312518125776/0001193125-18-125776-index.htm
#> 8                                                                                               <NA>
#> 9                                                                                               <NA>
#> 10   https://www.sec.gov/Archives/edgar/data/47288/000095012318003721/0000950123-18-003721-index.htm