SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

## This can be very slow running
full_text('intel')

#>    filing_date                                             name
#> 1   2018-05-22 INFORMATION TABLE of 13F-HR for AMALGAMATED BANK
#> 2   2018-05-22                          424B2 for CITIGROUP INC
#> 3   2018-05-22                             10-Q for NVIDIA CORP
#> 4   2018-05-22 INFORMATION TABLE of 13F-HR for McIlrath Eck LLC
#> 5   2018-05-22                     N-Q for Securian Funds Trust
#> 6   2018-05-22  INFORMATION TABLE of 13F-HR for NBT BANK N A NY
#> 7   2018-05-22                        424B2 for HSBC USA INC MD
#> 8   2018-05-22                                6-K for TIM S p A
#> 9   2018-05-22                   N-Q for Transamerica ETF Trust
#> 10  2018-05-22                       N-Q for TRANSAMERICA FUNDS
#>                                                                                                           href
#> 1  https://www.sec.gov/Archives/edgar/data/919192/000091919218000004/xslForm13F_X01/form_f13_data_20180331.xml
#> 2                https://www.sec.gov/Archives/edgar/data/831001/000095010318006395/dp91206_424b2-us1849743.htm
#> 3                         https://www.sec.gov/Archives/edgar/data/1045810/000104581018000080/nvda2019q110q.htm
#> 4           https://www.sec.gov/Archives/edgar/data/1701714/000109544918000034/xslForm13F_X01/table2018-03.xml
#> 5                           https://www.sec.gov/Archives/edgar/data/766351/000110465918035009/a18-3927_1nq.htm
#> 6        https://www.sec.gov/Archives/edgar/data/898413/000114036118025075/xslForm13F_X01/form13fInfoTable.xml
#> 7                          https://www.sec.gov/Archives/edgar/data/83246/000114420418030249/tv494783_424b2.htm
#> 8                             https://www.sec.gov/Archives/edgar/data/948642/000119312518170089/d577854d6k.htm
#> 9                            https://www.sec.gov/Archives/edgar/data/1673996/000119312518170330/d551530dnq.htm
#> 10                            https://www.sec.gov/Archives/edgar/data/787623/000119312518170335/d551532dnq.htm
#>              company_name     cik  sic
#> 1        AMALGAMATED BANK  919192 <NA>
#> 2           CITIGROUP INC  831001 6021
#> 3             NVIDIA CORP 1045810 3674
#> 4        McIlrath Eck LLC 1701714 <NA>
#> 5    Securian Funds Trust  766351 <NA>
#> 6         NBT BANK N A NY  898413 <NA>
#> 7         HSBC USA INC MD   83246 6021
#> 8               TIM S p A  948642 4899
#> 9  Transamerica ETF Trust 1673996 <NA>
#> 10     TRANSAMERICA FUNDS  787623 <NA>
#>                                                                                                                                                                                                                                                                                                       content
#> 1  ABBOTT LABORATORIES COM 002824100 12666 211381 SH SOLE 211381 0 0 ADVANCED MICRO DEVICES INC COM 007903107 1938 192859 SH SOLE 192859 0 0 AETNA INC COM 00817Y108 7107 42054 SH SOLE 42054 0 0 AIR PRODUCTS CHEMICALS INC COM 009158106 4243 26683 SH SOLE 26683 0 0 AMERICAN ELECTRIC POWER COM 025537101
#> 2                                                                   Each of these risks will depend on the performance of the worst performing of the shares of common stock of Intel Corporation and the shares of common stock of PayPal Holdings, Inc. (each, the underlying shares ), as described below.
#> 3      UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM 10-Q [x] QUARTERLY REPORT PURSUANT TO SECTION 13 OR 15(d) OF THE SECURITIES EXCHANGE ACT OF 1934 For the quarterly period ended April 29, 2018 OR [_] TRANSITION REPORT PURSUANT TO SECTION 13 OR 15(d) OF THE SECURITIES
#> 4         ACCO BRANDS CORP COM 00081T108 8 650 SH SOLE 650 0 0 AGNC INVT CORP COM 00123Q104 5 256 SH SOLE 256 0 0 ALPS ETF TR C&S GLOBL ETF 00162Q106 25 590 SH SOLE 590 0 0 ALPS ETF TR ALERIAN MLP 00162Q866 477 50903 SH SOLE 50903 0 0 AT&T INC COM 00206R102 125 3501 SH SOLE 3501 0 0 AU OPTRONICS CORP
#> 5                   UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-Q QUARTERLY SCHEDULE OF PORTFOLIO HOLDINGS OF REGISTERED MANAGEMENT INVESTMENT COMPANY Investment Company Act file number 811-4279 Securian Funds Trust (Exact name of registrant as specified in charter)
#> 6                                                                                                                                                                                                                                                                             AT&T IncCom00206R102313287854SH
#> 7                             Calculation of Registration Fee Title of Each Class of Securities Offered Maximum Aggregate Offering Price Amount of Registration Fee (1) Debt Securities (INTC) $2, 424, 000 $301.79 Debt Securities (MGM) $2, 956, 000 $368.02 Debt Securities (WDC) $1, 295, 000 $161.23 (1)
#> 8                          UNITED STATES SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 FORM 6-K REPORT OF FOREIGN PRIVATE ISSUER PURSUANT TO RULE 13a-16 OR 15D-16 UNDER THE SECURITIES EXCHANGE ACT OF 1934 FOR THE MONTH OF MAY 2018 TIM S.p.A. (Translation of registrant s name into English)
#> 9                        UNITED STATES SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 FORM N-Q QUARTERLY SCHEDULE OF PORTFOLIO HOLDINGS OF REGISTERED MANAGEMENT INVESTMENT COMPANY Investment Company Act file number: 811-23237 TRANSAMERICA ETF TRUST (Exact Name of Registrant as Specified in
#> 10                                SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 FORM N-Q QUARTERLY SCHEDULE OF PORTFOLIO HOLDINGS OF REGISTERED MANAGEMENT INVESTMENT COMPANY Investment Company Act file number: 811-04556 TRANSAMERICA FUNDS (Exact Name of Registrant as Specified in Charter)
#>                                                                                         parent_href
#> 1  https://www.sec.gov/Archives/edgar/data/919192/000091919218000004/xslForm13F_X01/primary_doc.xml
#> 2                                                                                              <NA>
#> 3                                                                                              <NA>
#> 4  https://www.sec.gov/Archives/edgar/data/919192/000091919218000004/xslForm13F_X01/primary_doc.xml
#> 5                                                                                              <NA>
#> 6  https://www.sec.gov/Archives/edgar/data/898413/000114036118025075/xslForm13F_X01/primary_doc.xml
#> 7                                                                                              <NA>
#> 8                                                                                              <NA>
#> 9                                                                                              <NA>
#> 10                                                                                             <NA>
#>                                                                                          index_href
#> 1  https://www.sec.gov/Archives/edgar/data/919192/000091919218000004/0000919192-18-000004-index.htm
#> 2                                                                                              <NA>
#> 3                                                                                              <NA>
#> 4  https://www.sec.gov/Archives/edgar/data/919192/000091919218000004/0000919192-18-000004-index.htm
#> 5                                                                                              <NA>
#> 6  https://www.sec.gov/Archives/edgar/data/898413/000114036118025075/0001140361-18-025075-index.htm
#> 7                                                                                              <NA>
#> 8                                                                                              <NA>
#> 9                                                                                              <NA>
#> 10                                                                                             <NA>