SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

full_text('intel')

#>    filing_date
#> 1   2017-12-27
#> 2   2017-12-27
#> 3   2017-12-27
#>                                                                                      name
#> 1  EX-102 of ABS-EE for Bank of America Merrill Lynch Commercial Mortgage Trust 2017 BNK3
#> 2                                                                        424B2 for UBS AG
#> 3                                                                        424B2 for UBS AG
#>                                                                                      href
#> 1          https://www.sec.gov/Archives/edgar/data/1694649/000105640417005963/exh_102.xml
#> 2  https://www.sec.gov/Archives/edgar/data/1114446/000111444617007888/tc4214518i_0fwp.htm
#> 3  https://www.sec.gov/Archives/edgar/data/1114446/000111444617007891/tc4214518f_1fwp.htm
#>                                                         company_name     cik
#> 1  Bank of America Merrill Lynch Commercial Mortgage Trust 2017 BNK3 1694649
#> 2                                                             UBS AG 1114446
#> 3                                                             UBS AG 1114446
#>     sic
#> 1  6189
#> 2  6021
#> 3  6021
#>                                                                                                                                                                                                                                                                                                         content
#> 1      Prospectus Loan ID 1 11-14-2017 12-11-2017 BANA 12-20-2016 73325000.00 120 01-01-2027 0.04762 0.04762 3 1 120 02-01-2017 true 1 PP 3 300677.31 73325000.00 1 1 1 0 true true false false false 06-30-2026 THE SUMMIT BIRMINGHAM 214 SUMMIT BOULEVARD Birmingham AL 35243 Jefferson RT 681245 681245 1997
#> 2                                   Underlying Equity The common stock of Intel Corporation Contingent Coupon If the closing price of the underlying equity is equal to or greater than the coupon barrier on any observation date, UBS will pay you the contingent coupon applicable to such observation date.
#> 3                                   Underlying Equity The common stock of Intel Corporation Contingent Coupon If the closing price of the underlying equity is equal to or greater than the coupon barrier on any observation date, UBS will pay you the contingent coupon applicable to such observation date.
#>                                                                                     parent_href
#> 1  https://www.sec.gov/Archives/edgar/data/1694649/000105640417005963/bal17bk3_absee-201712.htm
#> 2                                                                                          <NA>
#> 3                                                                                          <NA>
#>                                                                                           index_href
#> 1  https://www.sec.gov/Archives/edgar/data/1694649/000105640417005963/0001056404-17-005963-index.htm
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#>  [ reached getOption("max.print") -- omitted 7 rows ]