SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(
  q = "*",
  type = "",
  reverse_order = FALSE,
  count = 10,
  page = 1,
  stemming = TRUE,
  name = "",
  cik = "",
  sic = "",
  from = "",
  to = ""
)

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

# \donttest{
## This can be very slow running
full_text('intel')

#>    filing_date
#> 1   2020-08-31
#> 2   2020-08-31
#> 3   2020-08-31
#> 4   2020-08-31
#> 5   2020-08-31
#> 6   2020-08-31
#> 7   2020-08-31
#> 8   2020-08-31
#> 9   2020-08-31
#> 10  2020-08-31
#>                                                                        name
#> 1                                         N-CSRS for AMERICAN BALANCED FUND
#> 2                           N-CSRS for AMERICAN FUNDS FUNDAMENTAL INVESTORS
#> 3                                       N-CSRS for INVESTMENT CO OF AMERICA
#> 4                                         N-PX for AMANA MUTUAL FUNDS TRUST
#> 5                                         N-PX for SATURNA INVESTMENT TRUST
#> 6                                                    N-CSRS for TORRAY FUND
#> 7              EX-102 of ABS-EE for CSAIL 2017 C8 Commercial Mortgage Trust
#> 8    EX-102 of 10-D for JPMCC Commercial Mortgage Securities Trust 2017 JP7
#> 9  EX-102 of ABS-EE for JPMCC Commercial Mortgage Securities Trust 2017 JP7
#> 10                                          N-CSRS for Securian Funds Trust
#>                                                                                       href
#> 1          https://www.sec.gov/Archives/edgar/data/4568/000005193120000747/ambal_ncsrs.htm
#> 2            https://www.sec.gov/Archives/edgar/data/39473/000005193120000751/fi_ncsrs.htm
#> 3           https://www.sec.gov/Archives/edgar/data/51931/000005193120000752/ica_ncsrs.htm
#> 4  https://www.sec.gov/Archives/edgar/data/766285/000076628520000016/amanaformnpx19-20.htm
#> 5    https://www.sec.gov/Archives/edgar/data/811860/000081186020000043/sitformnpx19-20.htm
#> 6          https://www.sec.gov/Archives/edgar/data/862696/000089853120000358/ttf-ncsrs.htm
#> 7           https://www.sec.gov/Archives/edgar/data/1708131/000105640420009721/exh_102.xml
#> 8           https://www.sec.gov/Archives/edgar/data/1709967/000105640420009734/exh_102.xml
#> 9           https://www.sec.gov/Archives/edgar/data/1709967/000105640420009735/exh_102.xml
#> 10 https://www.sec.gov/Archives/edgar/data/766351/000110465920100749/tm2025137d1_ncsrs.htm
#>                                           company_name     cik  sic
#> 1                               AMERICAN BALANCED FUND    4568 <NA>
#> 2                 AMERICAN FUNDS FUNDAMENTAL INVESTORS   39473 <NA>
#> 3                             INVESTMENT CO OF AMERICA   51931 <NA>
#> 4                             AMANA MUTUAL FUNDS TRUST  766285 <NA>
#> 5                             SATURNA INVESTMENT TRUST  811860 <NA>
#> 6                                          TORRAY FUND  862696 <NA>
#> 7              CSAIL 2017 C8 Commercial Mortgage Trust 1708131 6189
#> 8  JPMCC Commercial Mortgage Securities Trust 2017 JP7 1709967 6189
#> 9  JPMCC Commercial Mortgage Securities Trust 2017 JP7 1709967 6189
#> 10                                Securian Funds Trust  766351 <NA>
#>                                                                                                                                                                                                                                                                                                       content
#> 1                  UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR Certified Shareholder Report of Registered Management Investment Companies Investment Company Act File Number: 811-00066 American Balanced Fund (Exact Name of Registrant as Specified in Charter) 6455
#> 2                  UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR Certified Shareholder Report of Registered Management Investment Companies Investment Company Act File Number: 811-00032 American Funds Fundamental Investors (Exact Name of Registrant as Specified in
#> 3            UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR Certified Shareholder Report of Registered Management Investment Companies Investment Company Act File Number: 811-00116 The Investment Company of America (Exact Name of Registrant as Specified in Charter)
#> 4    UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, DC 20549 FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY Investment Company Act file number: 811-04276 AMANA MUTUAL FUNDS TRUST (Exact Name of Registrant as Specified in Charter) 1300 N. State
#> 5    UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, DC 20549 FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY Investment Company Act file number: 811-05071 SATURNA INVESTMENT TRUST (Exact Name of Registrant as Specified in Charter) 1300 N. State
#> 6  UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number 811-06096 The Torray Fund (Exact name of registrant as specified in charter) 7501 Wisconsin Avenue, Suite
#> 7   Prospectus Loan ID 1 07-14-2020 08-11-2020 Natixis Real Estate Capital LLC 05-24-2017 70000000.00 120 06-05-2027 0.0341253 0.0341253 3 1 120 07-05-2017 true 1 PP 3 259494.47 70000000.00 1 1 1 0 true true true false false 06-04-2018 12-04-2026 12-04-2026 85 BROAD STREET 85 BROAD STREET New York NY
#> 8     Prospectus Loan ID 1 07-14-2020 08-11-2020 JPMorgan Chase Bank, National Association/Natixis Real Estate Capital LLC/Soci t G n rale/Deutsche Bank AG, New York Branch/Barclays Bank PLC 05-05-2017 75000000.00 120 06-01-2027 0.036694 0.036694 3 1 120 07-01-2017 true 1 A1 3 229337.50 75000000.00 1
#> 9     Prospectus Loan ID 1 07-14-2020 08-11-2020 JPMorgan Chase Bank, National Association/Natixis Real Estate Capital LLC/Soci t G n rale/Deutsche Bank AG, New York Branch/Barclays Bank PLC 05-05-2017 75000000.00 120 06-01-2027 0.036694 0.036694 3 1 120 07-01-2017 true 1 A1 3 229337.50 75000000.00 1
#> 10                      UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number 811-4279 Securian Funds Trust (Exact name of registrant as specified in charter) 400
#>                                                                                     parent_href
#> 1                                                                                          <NA>
#> 2                                                                                          <NA>
#> 3                                                                                          <NA>
#> 4                                                                                          <NA>
#> 5                                                                                          <NA>
#> 6                                                                                          <NA>
#> 7  https://www.sec.gov/Archives/edgar/data/1708131/000105640420009721/csc17sa8_absee-202008.htm
#> 8    https://www.sec.gov/Archives/edgar/data/1709967/000105640420009734/jpc17jp7_10d-202008.htm
#> 9  https://www.sec.gov/Archives/edgar/data/1709967/000105640420009735/jpc17jp7_absee-202008.htm
#> 10                                                                                         <NA>
#>                                                                                           index_href
#> 1                                                                                               <NA>
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5                                                                                               <NA>
#> 6                                                                                               <NA>
#> 7  https://www.sec.gov/Archives/edgar/data/1708131/000105640420009721/0001056404-20-009721-index.htm
#> 8  https://www.sec.gov/Archives/edgar/data/1709967/000105640420009734/0001056404-20-009734-index.htm
#> 9  https://www.sec.gov/Archives/edgar/data/1709967/000105640420009735/0001056404-20-009735-index.htm
#> 10                                                                                              <NA>

# }