SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

full_text('intel')

#>    filing_date
#> 1   2018-01-18
#> 2   2018-01-18
#> 3   2018-01-18
#>                                                                        name
#> 1                      INFORMATION TABLE of 13F-HR for ANDERSON HOAGLAND CO
#> 2  INFORMATION TABLE of 13F-HR for LOCKHEED MARTIN INVESTMENT MANAGEMENT CO
#> 3          INFORMATION TABLE of 13F-HR for Carroll Financial Associates Inc
#>                                                                                                      href
#> 1   https://www.sec.gov/Archives/edgar/data/791191/000079119118000001/xslForm13F_X01/form13fInfoTable.xml
#> 2    https://www.sec.gov/Archives/edgar/data/897599/000089759918000002/xslForm13F_X01/inftable-2017q4.xml
#> 3           https://www.sec.gov/Archives/edgar/data/1567784/000090901218000005/xslForm13F_X01/aci_13f.xml
#>                                company_name     cik  sic
#> 1                      ANDERSON HOAGLAND CO  791191 <NA>
#> 2  LOCKHEED MARTIN INVESTMENT MANAGEMENT CO  897599 3670
#> 3          Carroll Financial Associates Inc 1567784 <NA>
#>                                                                                                                                                                                                                                                                                                        content
#> 1   APPLE INC COMMON STOCK 037833100 3325 19650 SH SOLE 0 19650 0 0 ABB LTD ADR COMMON STOCK 000375204 817 30470 SH SOLE 0 30470 0 0 ABBVIE INC COMMON STOCK 00287Y109 687 7105 SH SOLE 0 7105 0 0 AUTOLIV INC COMMON STOCK 052800109 1250 9835 SH SOLE 0 9835 0 0 AMETEK, INC COMMON STOCK 031100100 679 9375
#> 2   ACTIVISION BLIZZARD INC COM 00507V109 681 10750 SH SOLE 10750 0 0 ADIENT PLC COM G0084W101 677 8600 SH SOLE 8600 0 0 ADOBE SYSTEMS INC COM 00724F101 2348 13400 SH SOLE 13400 0 0 ALEXION PHARMACEUTICALS, INC. COM 015351109 239 2000 SH SOLE 2000 0 0 ALIBABA GROUP HLDS LTD COM 01609W102 4462 25878 SH
#> 3     AMAZON COM INC COM Stock 023135106 1199 1026 SH SOLE 0 1026 0 0 ISHARES MORNINGSTAR LARGE-CAP GROWTH ETF ETF 464287119 24 155 SH SOLE 0 155 0 0 HSN INC COM Stock 404303109 0 11 SH SOLE 0 11 0 0 ILG INC COM Stock 44967H101 0 18 SH SOLE 0 18 0 0 LENDINGTREE INC NEW COM Stock 52603B107 1396 4101 SH
#>                                                                                          parent_href
#> 1   https://www.sec.gov/Archives/edgar/data/791191/000079119118000001/xslForm13F_X01/primary_doc.xml
#> 2   https://www.sec.gov/Archives/edgar/data/897599/000089759918000002/xslForm13F_X01/primary_doc.xml
#> 3  https://www.sec.gov/Archives/edgar/data/1567784/000090901218000005/xslForm13F_X01/primary_doc.xml
#>                                                                                           index_href
#> 1   https://www.sec.gov/Archives/edgar/data/791191/000079119118000001/0000791191-18-000001-index.htm
#> 2   https://www.sec.gov/Archives/edgar/data/897599/000089759918000002/0000897599-18-000002-index.htm
#> 3  https://www.sec.gov/Archives/edgar/data/1567784/000090901218000005/0000909012-18-000005-index.htm
#>  [ reached getOption("max.print") -- omitted 7 rows ]