SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

full_text('intel')

#>    filing_date
#> 1   2017-12-22
#> 2   2017-12-22
#> 3   2017-12-21
#>                                                                  name
#> 1                       N-CSR for PRUDENTIAL INVESTMENT PORTFOLIOS 16
#> 2                PART II AND III of 1-A/A for GLOBALTECH HOLDINGS INC
#> 3                                    N-Q for LORD ABBETT EQUITY TRUST
#>                                                                                      href
#> 1     https://www.sec.gov/Archives/edgar/data/1067442/000119312517376750/d304597dncsr.htm
#> 2    https://www.sec.gov/Archives/edgar/data/943770/000168316817003377/globaltech_poc.htm
#> 3        https://www.sec.gov/Archives/edgar/data/1139819/000093041317004073/c89894_nq.htm
#>                                              company_name     cik  sic
#> 1                     PRUDENTIAL INVESTMENT PORTFOLIOS 16 1067442 <NA>
#> 2                                 GLOBALTECH HOLDINGS INC  943770 7372
#> 3                                LORD ABBETT EQUITY TRUST 1139819 <NA>
#>                                                                                                                                                                                                                                                                                                         content
#> 1                UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number: 811-08915 Exact name of registrant as specified in charter: Prudential Investment Portfolios
#> 2         Preliminary Offering Circular dated December _, 2017 An Offering Statement pursuant to Regulation A relating to these securities has been filed with the Securities and Exchange Commission. Information contained in this Preliminary Offering Circular is subject to completion or amendment. These
#> 3                        UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-Q QUARTERLY SCHEDULE OF PORTFOLIO HOLDINGS OF REGISTERED MANAGEMENT INVESTMENT COMPANY Investment Company Act File Number: 811-10371 LORD ABBETT EQUITY TRUST (Exact name of Registrant as specified in
#>                                                                                     parent_href
#> 1                                                                                          <NA>
#> 2  https://www.sec.gov/Archives/edgar/data/943770/000168316817003377/xsl1-A_X01/primary_doc.xml
#> 3                                                                                          <NA>
#>                                                                                          index_href
#> 1                                                                                              <NA>
#> 2  https://www.sec.gov/Archives/edgar/data/943770/000168316817003377/0001683168-17-003377-index.htm
#> 3                                                                                              <NA>
#>  [ reached getOption("max.print") -- omitted 7 rows ]