SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

## This can be very slow running
full_text('intel')

#>    filing_date                                                     name
#> 1   2018-07-06 INFORMATION TABLE of 13F-HR for FIRST AMERICAN TRUST FSB
#> 2   2018-07-06        INFORMATION TABLE of 13F-HR for ROCKLAND TRUST CO
#> 3   2018-07-06           N-2/A for Hedge Fund Guided Portfolio Solution
#> 4   2018-07-06                          N-CSRS for HENNESSY FUNDS TRUST
#> 5   2018-07-06  INFORMATION TABLE of 13F-HR for DnB Asset Management AS
#> 6   2018-07-06  INFORMATION TABLE of 13F-HR for ABNER HERRMAN BROCK LLC
#> 7   2018-07-06     INFORMATION TABLE of 13F-HR for West Oak Capital LLC
#> 8   2018-07-06                        N-PX for Reality Shares ETF Trust
#> 9   2018-07-06          N-CSR for OPPENHEIMER DIVIDEND OPPORTUNITY FUND
#> 10  2018-07-06                N-CSRS for OPPENHEIMER EQUITY INCOME FUND
#>                                                                                                      href
#> 1    https://www.sec.gov/Archives/edgar/data/36066/000003606618000011/xslForm13F_X01/form13fInfoTable.xml
#> 2           https://www.sec.gov/Archives/edgar/data/84616/000008461618000002/xslForm13F_X01/13fr-0630.xml
#> 3                              https://www.sec.gov/Archives/edgar/data/1738008/000089843218000774/n2a.htm
#> 4                      https://www.sec.gov/Archives/edgar/data/891944/000089853118000390/hft_hf-ncsrs.htm
#> 5  https://www.sec.gov/Archives/edgar/data/1456670/000095012318006606/xslForm13F_X01/form13fInfoTable.xml
#> 6    https://www.sec.gov/Archives/edgar/data/1038661/000103866118000003/xslForm13F_X01/inftable070618.xml
#> 7  https://www.sec.gov/Archives/edgar/data/1387399/000114036118031695/xslForm13F_X01/form13fInfoTable.xml
#> 8                    https://www.sec.gov/Archives/edgar/data/1573496/000114420418037444/tv497973_n-px.htm
#> 9                     https://www.sec.gov/Archives/edgar/data/1191290/000119312518213229/d504917dncsr.htm
#> 10                    https://www.sec.gov/Archives/edgar/data/799029/000119312518213248/d600714dncsrs.htm
#>                             company_name     cik  sic
#> 1               FIRST AMERICAN TRUST FSB   36066 <NA>
#> 2                      ROCKLAND TRUST CO   84616 <NA>
#> 3   Hedge Fund Guided Portfolio Solution 1738008 <NA>
#> 4                   HENNESSY FUNDS TRUST  891944 <NA>
#> 5                DnB Asset Management AS 1456670 <NA>
#> 6                ABNER HERRMAN BROCK LLC 1038661 <NA>
#> 7                   West Oak Capital LLC 1387399 <NA>
#> 8               Reality Shares ETF Trust 1573496 <NA>
#> 9  OPPENHEIMER DIVIDEND OPPORTUNITY FUND 1191290 <NA>
#> 10        OPPENHEIMER EQUITY INCOME FUND  799029 <NA>
#>                                                                                                                                                                                                                                                                                                        content
#> 1                                                                                                                                                                                                                                                                                                             
#> 2     ABBOTT LABS COMMON STOCK 002824100 4388 71945 SH DFND 0 71945 0 0 ABBVIE INC COMMON STOCK 00287Y109 1055 11387 SH DFND 0 11247 0 140 ADOBE SYS INC COMMON STOCK 00724F101 700 2871 SH DFND 0 2871 0 0 AFLAC INC COMMON STOCK 001055102 772 17934 SH DFND 0 17934 0 0 AIR PRODUCTS & CHEMICALS INC COMMON
#> 3                         Securities Act File No. 333-224465 Investment Company Act File No. 811-23347 UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 Form N-2 REGISTRATION STATEMENT UNDER THE SECURITIES ACT OF 1933 Pre-Effective Amendment No. 1 Post-Effective Amendment No. and
#> 4             As filed with the Securities and Exchange Commission on July 6, 2018 UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number (811-07168) Hennessy Funds
#> 5                                                                                                                                                                                                                                                                      Noble Energy Inccommon65504410527463504
#> 6  3M CO COM 88579Y101 1113 5657 SH SOLE 5657 0 0 ABBOTT LABORATORIES COM 002824100 13142 215477 SH SOLE 215477 0 0 ABBVIE INC COM 00287Y109 604 6518 SH SOLE 6518 0 0 ACCENTURE PLC IRELAND SHS CL A COM g1151c101 502 3067 SH SOLE 3067 0 0 ACTIVISION BLIZZARD INC COM 00507V109 8063 105645 SH SOLE 105645
#> 7                                                                                                                                                                                                                                                                           3M CompanyCOM88579Y1013590141825SH
#> 8                    UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, DC 20549 FORM N-PX ANNUAL REPORT OF PROXY VOTING RECORD OF REGISTERED MANAGEMENT INVESTMENT COMPANY Investment Company Act file number 811-22991 Reality Shares ETF Trust (Exact name of registrant as specified in charter)
#> 9     UNITED STATES SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number 811-21208 Oppenheimer Dividend Opportunity Fund (Exact name of registrant as specified in charter) 6803
#> 10     UNITED STATES SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number 811-04797 Oppenheimer Equity Income Fund (Exact name of registrant as specified in charter) 6803 South
#>                                                                                          parent_href
#> 1    https://www.sec.gov/Archives/edgar/data/36066/000003606618000011/xslForm13F_X01/primary_doc.xml
#> 2    https://www.sec.gov/Archives/edgar/data/84616/000008461618000002/xslForm13F_X01/primary_doc.xml
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5  https://www.sec.gov/Archives/edgar/data/1387399/000114036118031695/xslForm13F_X01/primary_doc.xml
#> 6    https://www.sec.gov/Archives/edgar/data/36066/000003606618000011/xslForm13F_X01/primary_doc.xml
#> 7    https://www.sec.gov/Archives/edgar/data/84616/000008461618000002/xslForm13F_X01/primary_doc.xml
#> 8                                                                                               <NA>
#> 9                                                                                               <NA>
#> 10                                                                                              <NA>
#>                                                                                           index_href
#> 1    https://www.sec.gov/Archives/edgar/data/36066/000003606618000011/0000036066-18-000011-index.htm
#> 2    https://www.sec.gov/Archives/edgar/data/84616/000008461618000002/0000084616-18-000002-index.htm
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5  https://www.sec.gov/Archives/edgar/data/1387399/000114036118031695/0001140361-18-031695-index.htm
#> 6    https://www.sec.gov/Archives/edgar/data/36066/000003606618000011/0000036066-18-000011-index.htm
#> 7    https://www.sec.gov/Archives/edgar/data/84616/000008461618000002/0000084616-18-000002-index.htm
#> 8                                                                                               <NA>
#> 9                                                                                               <NA>
#> 10                                                                                              <NA>