SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(
  q = "*",
  type = "",
  reverse_order = FALSE,
  count = 10,
  page = 1,
  stemming = TRUE,
  name = "",
  cik = "",
  sic = "",
  from = "",
  to = ""
)

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

# \donttest{
## This can be very slow running
full_text('intel')

#>    filing_date                                                     name
#> 1   2020-09-25                                         424B2 for UBS AG
#> 2   2020-09-25                                         424B2 for UBS AG
#> 3   2020-09-25                            N-CSR/A for SSGA Active Trust
#> 4   2020-09-25                          S-1 for Aligos Therapeutics Inc
#> 5   2020-09-25                        424B2 for GOLDMAN SACHS GROUP INC
#> 6   2020-09-25 INFORMATION TABLE of 13F-HR/A for Your Advocates Ltd LLP
#> 7   2020-09-25                     DEF 14A for NIOCORP DEVELOPMENTS LTD
#> 8   2020-09-24                                         424B2 for UBS AG
#> 9   2020-09-24                                  10-Q for BLACKBERRY Ltd
#> 10  2020-09-24                             424B4 for Corsair Gaming Inc
#>                                                                                                  href
#> 1              https://www.sec.gov/Archives/edgar/data/1114446/000111444620005166/tc6218048i_0fwp.htm
#> 2              https://www.sec.gov/Archives/edgar/data/1114446/000111444620005169/tc6218048f_1fwp.htm
#> 3                 https://www.sec.gov/Archives/edgar/data/1516212/000119312520254021/d12026dncsra.htm
#> 4                    https://www.sec.gov/Archives/edgar/data/1799448/000119312520254760/d31402ds1.htm
#> 5                      https://www.sec.gov/Archives/edgar/data/886982/000156459020044652/gs-424b2.htm
#> 6  https://www.sec.gov/Archives/edgar/data/1774437/000177443720000005/xslForm13F_X01/advocatesamt.xml
#> 7               https://www.sec.gov/Archives/edgar/data/1512228/000178598220000899/sf0030h_def14a.htm
#> 8             https://www.sec.gov/Archives/edgar/data/1114446/000091412120003403/ub55320920-424b2.htm
#> 9                https://www.sec.gov/Archives/edgar/data/1070235/000107023520000101/bbry-20200831.htm
#> 10               https://www.sec.gov/Archives/edgar/data/1743759/000119312520252781/d507744d424b4.htm
#>                company_name     cik  sic
#> 1                    UBS AG 1114446 6021
#> 2                    UBS AG 1114446 6021
#> 3         SSGA Active Trust 1516212 <NA>
#> 4   Aligos Therapeutics Inc 1799448 2836
#> 5   GOLDMAN SACHS GROUP INC  886982 6211
#> 6    Your Advocates Ltd LLP 1774437 <NA>
#> 7  NIOCORP DEVELOPMENTS LTD 1512228 1000
#> 8                    UBS AG 1114446 6021
#> 9            BLACKBERRY Ltd 1070235 7372
#> 10       Corsair Gaming Inc 1743759 3577
#>                                                                                                                                                                                                                                                                                                       content
#> 1                                   Underlying Asset The common stock of Intel Corporation Contingent Coupon If the closing price of the underlying asset is equal to or greater than the coupon barrier on any observation date, UBS will pay you the contingent coupon applicable to such observation date.
#> 2                                   Underlying Asset The common stock of Intel Corporation Contingent Coupon If the closing price of the underlying asset is equal to or greater than the coupon barrier on any observation date, UBS will pay you the contingent coupon applicable to such observation date.
#> 3                            UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number: 811-22542 SSGA ACTIVE TRUST (Exact name of registrant as specified in charter)
#> 4                                                           As filed with the Securities and Exchange Commission on September 25, 2020. Registration No. 333- UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM S-1 REGISTRATION STATEMENT Under The Securities Act of 1933 Aligos
#> 5        Filed Pursuant to Rule 424(b)(2) Registration Statement No. 333-239610 Underlier Supplement No. 12 to the Prospectus and the Prospectus Supplement, each as may be amended from time to time, that form a part of Registration Statement No. 333-239610 GS Finance Corp. Medium-Term Notes, Series E
#> 6  Advanced Micro Devices COM 007903107 11 206 SH SOLE 0 0 206 Air Prod & Chemicals Inc COM 009158106 29 100 SH SOLE 0 0 100 Albemarle Corp COM 012653101 1 12 SH SOLE 0 0 12 Allegiance Bank COM 01748H107 32 1414 SH SOLE 0 0 1414 Allstate Corporation COM 020002101 36 400 SH SOLE 0 0 400 Amazon Com Inc
#> 7                              UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, DC 20549 SCHEDULE 14A Proxy Statement Pursuant to Section 14(a) of the Securities Exchange Act of 1934 (Amendment No. _) Filed by the Registrant Filed by a Party other than the Registrant Check the appropriate
#> 8               PRICING SUPPLEMENT Dated September 23, 2020 Filed Pursuant to Rule 424(b)(2) Registration Statement No. 333-225551 (To Prospectus dated October 31, 2018 and Product Supplement dated October 31, 2018) UBS AG Airbag Autocallable Yield Notes UBS AG $5, 000, 000 Notes Linked to the Common
#> 9                             bbry-20200831 8/31/2020 False 2021 Q2 0001070235 -02-28 556, 467, 825 554, 199, 016 556, 467, 825 554, 199, 016 14 9 1 2 1 1 P1M P5Y 605 3.75 60.5 10 November 13, 2020 7.75 115 35 50 0001070235 2020-03-01 2020-08-31 0001070235 exch:XNYS 2020-03-01 2020-08-31 xbrli:shares
#> 10                                                    Filed pursuant to Rule 424(b)(4) Registration No. 333-248247 14, 000, 000 Shares Corsair Gaming, Inc. Common Stock This is an initial public offering of common stock of Corsair Gaming, Inc. We are selling 7, 500, 00 shares of our common stock. The
#>                                                                                          parent_href
#> 1                                                                                               <NA>
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5                                                                                               <NA>
#> 6  https://www.sec.gov/Archives/edgar/data/1774437/000177443720000005/xslForm13F_X01/primary_doc.xml
#> 7                                                                                               <NA>
#> 8                                                                                               <NA>
#> 9                                                                                               <NA>
#> 10                                                                                              <NA>
#>                                                                                           index_href
#> 1                                                                                               <NA>
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5                                                                                               <NA>
#> 6  https://www.sec.gov/Archives/edgar/data/1774437/000177443720000005/0001774437-20-000005-index.htm
#> 7                                                                                               <NA>
#> 8                                                                                               <NA>
#> 9                                                                                               <NA>
#> 10                                                                                              <NA>

# }