SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

## This can be very slow running
full_text('intel')

#>    filing_date
#> 1   2018-05-21
#> 2   2018-05-21
#> 3   2018-05-21
#> 4   2018-05-21
#> 5   2018-05-21
#> 6   2018-05-21
#> 7   2018-05-21
#> 8   2018-05-21
#> 9   2018-05-21
#> 10  2018-05-21
#>                                                                 name
#> 1                                         497 for PUTNAM FUNDS TRUST
#> 2        INFORMATION TABLE of 13F-HR for MASSMUTUAL TRUST CO FSB ADV
#> 3                            487 for Advisors Disciplined Trust 1867
#> 4                                       N-Q for DIVIDEND INCOME FUND
#> 5                                10-K/A for X Rail Entertainment Inc
#> 6           INFORMATION TABLE of 13F-HR for NAVELLIER ASSOCIATES INC
#> 7                                              20-F for SuperCom Ltd
#> 8  EX-99.(B)(7)(A) of 485BPOS for FARMERS ANNUITY SEPARATE ACCOUNT A
#> 9                                EX-99.1 of 10-12B/A for Veoneer Inc
#> 10                                    N-CSRS for MFS SERIES TRUST XI
#>                                                                                                                href
#> 1                                   https://www.sec.gov/Archives/edgar/data/1005942/000092881618001080/a_mcc497.htm
#> 2            https://www.sec.gov/Archives/edgar/data/1103653/000095012318005905/xslForm13F_X01/form13fInfoTable.xml
#> 3                                https://www.sec.gov/Archives/edgar/data/1728581/000101322818000566/adt1867-487.txt
#> 4                                      https://www.sec.gov/Archives/edgar/data/1059213/000105921318000012/dninq.htm
#> 5                                      https://www.sec.gov/Archives/edgar/data/1697935/000109690618000271/xrail.htm
#> 6  https://www.sec.gov/Archives/edgar/data/872163/000110465918034722/xslForm13F_X01/a18-13858_1informationtable.xml
#> 7                               https://www.sec.gov/Archives/edgar/data/1291855/000114420418029944/tv493829_20f.htm
#> 8                            https://www.sec.gov/Archives/edgar/data/1089912/000119312518168311/d542975dex99b7a.htm
#> 9                              https://www.sec.gov/Archives/edgar/data/1733186/000119312518168532/d550335dex991.htm
#> 10                              https://www.sec.gov/Archives/edgar/data/911637/000119312518168812/d495134dncsrs.htm
#>                          company_name     cik  sic
#> 1                  PUTNAM FUNDS TRUST 1005942 <NA>
#> 2         MASSMUTUAL TRUST CO FSB ADV 1103653 <NA>
#> 3     Advisors Disciplined Trust 1867 1728581 <NA>
#> 4                DIVIDEND INCOME FUND 1059213 <NA>
#> 5            X Rail Entertainment Inc 1697935 4700
#> 6            NAVELLIER ASSOCIATES INC  872163 <NA>
#> 7                        SuperCom Ltd 1291855 3674
#> 8  FARMERS ANNUITY SEPARATE ACCOUNT A 1089912 <NA>
#> 9                         Veoneer Inc 1733186 3714
#> 10                MFS SERIES TRUST XI  911637 <NA>
#>                                                                                                                                                                                                                                                                                                     content
#> 1    May 25, 2018 Dear Fellow Shareholder: The Board of Trustees of Putnam Investors Fund is pleased to announce the upcoming merger of Putnam Investors Fund with and into Putnam Multi-Cap Core Fund. In this merger, shares of Putnam Investors Fund would, in effect, be exchanged for shares of Putnam
#> 2                                                                                                                                                                                                                                                      1st Source CorpCommon / Ordinary Stock33690110360744
#> 3  1933 ACT FILE NO.: 333-222794 1940 ACT FILE NO.: 811-21056 CIK NO.: 1728581 SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 AMENDMENT NO. 1 TO REGISTRATION STATEMENT ON FORM S-6 FOR REGISTRATION UNDER THE SECURITIES ACT OF 1933 OF SECURITIES OF UNIT INVESTMENT TRUSTS REGISTERED ON FORM
#> 4                    UNITED STATES SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 FORM N-Q QUARTERLY SCHEDULE OF PORTFOLIO HOLDINGS OF REGISTERED MANAGEMENT INVESTMENT COMPANY Investment Company Act file number: 811-08747 DIVIDEND AND INCOME FUND (Exact name of registrant as specified in
#> 5          UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM 10-K/A (MARK ONE) [X] ANNUAL REPORT PURSUANT TO SECTION 13 OR 15(d) OF THE SECURITIES EXCHANGE ACT OF 1934 For the fiscal year ended December 31, 2017 Commission file number 333-218746 X RAIL ENTERTAINMENT, INC.
#> 6                                                                                                                                                                                                                                                                                51Job Inc.COM3168271041413
#> 7                 UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM 20-F REGISTRATION STATEMENT PURSUANT TO SECTION 12(b) OR (g) OF THE SECURITIES EXCHANGE ACT OF 1934 OR x ANNUAL REPORT PURSUANT TO SECTION 13 OR 15(d) OF THE SECURITIES EXCHANGE ACT OF 1934 For the fiscal
#> 8                                                  Execution Copy REINSURANCE AGREEMENT Between FARMERS NEW WORLD LIFE INSURANCE COMPANY (referred to as the Ceding Company) and RGA REINSURANCE COMPANY (referred to as the Reinsurer) TABLE OF CONTENTS ARTICLE I DEFINITIONS 1 Section 1.1 Definitions 1
#> 9                                       Exhibit 99.1 , 2018 Dear Autoliv, Inc. Stockholder: On December 12, 2017, we announced that, following the conclusion of a strategic review by our Board of Directors, we intend to spin-off our Electronics business segment, creating a new, independent publicly
#> 10 UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number 811-07992 MFS SERIES TRUST XI (Exact name of registrant as specified in charter) 111 Huntington Avenue,
#>                                                                                          parent_href
#> 1                                                                                               <NA>
#> 2   https://www.sec.gov/Archives/edgar/data/872163/000110465918034722/xslForm13F_X01/primary_doc.xml
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5                                                                                               <NA>
#> 6   https://www.sec.gov/Archives/edgar/data/872163/000110465918034722/xslForm13F_X01/primary_doc.xml
#> 7                                                                                               <NA>
#> 8              https://www.sec.gov/Archives/edgar/data/1733186/000119312518168532/d550335d1012ba.htm
#> 9  https://www.sec.gov/Archives/edgar/data/1103653/000095012318005905/xslForm13F_X01/primary_doc.xml
#> 10                                                                                              <NA>
#>                                                                                           index_href
#> 1                                                                                               <NA>
#> 2   https://www.sec.gov/Archives/edgar/data/872163/000110465918034722/0001104659-18-034722-index.htm
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5                                                                                               <NA>
#> 6   https://www.sec.gov/Archives/edgar/data/872163/000110465918034722/0001104659-18-034722-index.htm
#> 7                                                                                               <NA>
#> 8  https://www.sec.gov/Archives/edgar/data/1733186/000119312518168532/0001193125-18-168532-index.htm
#> 9  https://www.sec.gov/Archives/edgar/data/1103653/000095012318005905/0000950123-18-005905-index.htm
#> 10                                                                                              <NA>