SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

## This can be very slow running
full_text('intel')

#>    filing_date                                         name
#> 1   2018-05-24            10-Q for APPLIED MATERIALS INC DE
#> 2   2018-05-24     XML of 10-Q for APPLIED MATERIALS INC DE
#> 3   2018-05-24     XML of 10-Q for APPLIED MATERIALS INC DE
#> 4   2018-05-24      N-Q for AMERICAN FUNDS INSURANCE SERIES
#> 5   2018-05-24               N-Q for AMERICAN BALANCED FUND
#> 6   2018-05-24 N-Q for AMERICAN FUNDS FUNDAMENTAL INVESTORS
#> 7   2018-05-24             N-Q for INVESTMENT CO OF AMERICA
#> 8   2018-05-24                       10-K for MCKESSON CORP
#> 9   2018-05-24    N-CSRS for VANGUARD WELLESLEY INCOME FUND
#> 10  2018-05-24             N-CSRS for VANGUARD FENWAY FUNDS
#>                                                                                           href
#> 1   https://www.sec.gov/Archives/edgar/data/6951/000000695118000017/amat0429201810-qq22018.htm
#> 2                      https://www.sec.gov/Archives/edgar/data/6951/000000695118000017/R20.htm
#> 3                      https://www.sec.gov/Archives/edgar/data/6951/000000695118000017/R31.htm
#> 4                https://www.sec.gov/Archives/edgar/data/729528/000005193118000518/afis_nq.htm
#> 5                 https://www.sec.gov/Archives/edgar/data/4568/000005193118000519/ambal_nq.htm
#> 6                   https://www.sec.gov/Archives/edgar/data/39473/000005193118000523/fi_nq.htm
#> 7                  https://www.sec.gov/Archives/edgar/data/51931/000005193118000524/ica-nq.htm
#> 8        https://www.sec.gov/Archives/edgar/data/927653/000092765318000010/mck_10kx3312018.htm
#> 9  https://www.sec.gov/Archives/edgar/data/105544/000093247118006083/wellesleyincome_final.htm
#> 10          https://www.sec.gov/Archives/edgar/data/826473/000093247118006084/fenway_final.htm
#>                            company_name    cik  sic
#> 1              APPLIED MATERIALS INC DE   6951 3674
#> 2              APPLIED MATERIALS INC DE   6951 3674
#> 3              APPLIED MATERIALS INC DE   6951 3674
#> 4       AMERICAN FUNDS INSURANCE SERIES 729528 <NA>
#> 5                AMERICAN BALANCED FUND   4568 <NA>
#> 6  AMERICAN FUNDS FUNDAMENTAL INVESTORS  39473 <NA>
#> 7              INVESTMENT CO OF AMERICA  51931 <NA>
#> 8                         MCKESSON CORP 927653 5122
#> 9        VANGUARD WELLESLEY INCOME FUND 105544 <NA>
#> 10                VANGUARD FENWAY FUNDS 826473 <NA>
#>                                                                                                                                                                                                                                                                                           content
#> 1  UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 Form 10-Q (Mark One) QUARTERLY REPORT PURSUANT TO SECTION 13 OR 15(d) OF THE SECURITIES EXCHANGE ACT OF 1934 For the quarterly period ended April 29, 2018 or TRANSITION REPORT PURSUANT TO SECTION 13 OR 15(d) OF THE
#> 2                                                                                                                              Percentage of Net Sales Samsung Electronics Co., Ltd. 18 % Taiwan Semiconductor Manufacturing Company Limited 12 % Intel Corporation 12 % No definition available.
#> 3                           v3.8.0.1 Industry Segment Operations (Tables) 6 Months Ended Apr. 29, 2018 Net sales and operating income (loss) for each reportable segment were as follows: Three Months Ended Six Months Ended Net Sales Operating Income (Loss) Net Sales Operating Income (Loss)
#> 4                UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-Q Quarterly Schedule of Portfolio Holdings of Registered Management Investment Company Investment Company Act File Number: 811-03857 American Funds Insurance Series (Exact Name of Registrant as
#> 5            UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-Q Quarterly Schedule of Portfolio Holdings of Registered Management Investment Company Investment Company Act File Number: 811-00066 American Balanced Fund (Exact Name of Registrant as Specified in
#> 6              UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-Q Quarterly Schedule of Portfolio Holdings of Registered Management Investment Company Investment Company Act File Number: 811-00032 American Funds Fundamental Investors (Exact Name of Registrant
#> 7              UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-Q Quarterly Schedule of Portfolio Holdings of Registered Management Investment Company Investment Company Act File Number: 811-00116 The Investment Company of America (Exact Name of Registrant as
#> 8        UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM 10-K x ANNUAL REPORT PURSUANT TO SECTION 13 OR 15(d) OF THE SECURITIES EXCHANGE ACT OF 1934 For the fiscal year ended March 31, 2018 OR TRANSITION REPORT PURSUANT TO SECTION 13 OR 15(d) OF THE SECURITIES
#> 9  UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number: 811-01766 Name of Registrant: Vanguard Wellesley Income Fund Address of Registrant: P.O. Box
#> 10     UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number: 811-05445 Name of Registrant: Vanguard Fenway Funds Address of Registrant: P.O. Box 2600
#>                                                                                   parent_href
#> 1                                                                                        <NA>
#> 2  https://www.sec.gov/Archives/edgar/data/6951/000000695118000017/amat0429201810-qq22018.htm
#> 3  https://www.sec.gov/Archives/edgar/data/6951/000000695118000017/amat0429201810-qq22018.htm
#> 4                                                                                        <NA>
#> 5                                                                                        <NA>
#> 6                                                                                        <NA>
#> 7                                                                                        <NA>
#> 8                                                                                        <NA>
#> 9                                                                                        <NA>
#> 10                                                                                       <NA>
#>                                                                                        index_href
#> 1                                                                                            <NA>
#> 2  https://www.sec.gov/Archives/edgar/data/6951/000000695118000017/0000006951-18-000017-index.htm
#> 3  https://www.sec.gov/Archives/edgar/data/6951/000000695118000017/0000006951-18-000017-index.htm
#> 4                                                                                            <NA>
#> 5                                                                                            <NA>
#> 6                                                                                            <NA>
#> 7                                                                                            <NA>
#> 8                                                                                            <NA>
#> 9                                                                                            <NA>
#> 10                                                                                           <NA>