SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(
  q = "*",
  type = "",
  reverse_order = FALSE,
  count = 10,
  page = 1,
  stemming = TRUE,
  name = "",
  cik = "",
  sic = "",
  from = "",
  to = ""
)

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

# \donttest{
## This can be very slow running
full_text('intel')

#>    filing_date
#> 1   2020-01-02
#> 2   2020-01-02
#> 3   2020-01-02
#> 4   2020-01-02
#> 5   2020-01-02
#> 6   2020-01-02
#> 7   2020-01-02
#> 8   2020-01-02
#> 9   2020-01-02
#> 10  2020-01-02
#>                                                                  name
#> 1  INFORMATION TABLE of 13F-HR for STONEBRIDGE CAPITAL MANAGEMENT INC
#> 2                           N-CSR for DAVIS NEW YORK VENTURE FUND INC
#> 3                        485APOS for AMERICAN CENTURY MUNICIPAL TRUST
#> 4                        N-CSR for BNY Mellon Investment Funds IV Inc
#> 5                                                    424B2 for UBS AG
#> 6          INFORMATION TABLE of 13F-HR for CACTI ASSET MANAGEMENT LLC
#> 7               INFORMATION TABLE of 13F-HR for PARK NATIONAL CORP OH
#> 8                                        FWP for ROYAL BANK OF CANADA
#> 9                                        N-CSR for SCHWAB INVESTMENTS
#> 10                                   N-CSR for Goldman Sachs Trust II
#>                                                                                                     href
#> 1  https://www.sec.gov/Archives/edgar/data/51812/000005181220000001/xslForm13F_X01/SEC13f123119.xlsx.xml
#> 2                 https://www.sec.gov/Archives/edgar/data/71701/000007170120000002/dgf_dif_ncsr_1019.htm
#> 3                 https://www.sec.gov/Archives/edgar/data/746458/000074645820000002/acmt2020mash485a.htm
#> 4                         https://www.sec.gov/Archives/edgar/data/819940/000081994020000002/lp1-0310.htm
#> 5                https://www.sec.gov/Archives/edgar/data/1114446/000091412120000016/ub54548307-424b2.htm
#> 6        https://www.sec.gov/Archives/edgar/data/1452208/000110465920000145/xslForm13F_X01/infotable.xml
#> 7         https://www.sec.gov/Archives/edgar/data/805676/000110465920000356/xslForm13F_X01/infotable.xml
#> 8                         https://www.sec.gov/Archives/edgar/data/1000275/000114036120000021/formfwp.htm
#> 9                     https://www.sec.gov/Archives/edgar/data/869365/000119312520000455/d732533dncsr.htm
#> 10                   https://www.sec.gov/Archives/edgar/data/1557156/000119312520000470/d787992dncsr.htm
#>                          company_name     cik  sic
#> 1  STONEBRIDGE CAPITAL MANAGEMENT INC   51812 <NA>
#> 2     DAVIS NEW YORK VENTURE FUND INC   71701 <NA>
#> 3    AMERICAN CENTURY MUNICIPAL TRUST  746458 <NA>
#> 4  BNY Mellon Investment Funds IV Inc  819940 <NA>
#> 5                              UBS AG 1114446 6021
#> 6          CACTI ASSET MANAGEMENT LLC 1452208 <NA>
#> 7               PARK NATIONAL CORP OH  805676 6021
#> 8                ROYAL BANK OF CANADA 1000275 6029
#> 9                  SCHWAB INVESTMENTS  869365 <NA>
#> 10             Goldman Sachs Trust II 1557156 <NA>
#>                                                                                                                                                                                                                                                                                                         content
#> 1   3M COMPANY COM COM 88579Y101 1576 8932 SH SOLE 0 0 8932 3M COMPANY COM COM 88579Y101 159 900 SH OTR 0 0 900 ABB LTD SPONSORED ADR FOREIGN COMMON 000375204 982 40750 SH SOLE 0 0 40750 ABB LTD SPONSORED ADR FOREIGN COMMON 000375204 314 13050 SH OTR 0 0 13050 ABBOTT LABS COM 002824100 356 4094 SH SOLE
#> 2         UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, DC 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANY Investment Company Act file number 811-01701 DAVIS NEW YORK VENTURE FUND, INC. (Exact name of registrant as specified in charter) 2949 East
#> 3                            As Filed with the U.S. Securities and Exchange Commission on January 2, 2020 1933 Act File No. 002-91229 1940 Act File No. 811-04025 UNITED STATES SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 _ FORM N-1A _ REGISTRATION STATEMENT UNDER THE SECURITIES ACT OF 1933
#> 4                                  UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number 811-05202 BNY Mellon Investment Funds IV, Inc. (Exact name of Registrant as
#> 5             The information in this preliminary pricing supplement is not complete and may be changed. We may not sell these Notes until the pricing supplement, the accompanying product supplement and the accompanying prospectus (collectively, the Offering Documents ) are delivered in final form. The
#> 6  ADOBE SYS INC COM 37045V100 1025 3100 SH SOLE 3100 0 0 ALLIANCEBERNSTEIN HLDG LP UNIT LTD PARTN 01881G106 1852 61850 SH SOLE 61850 0 0 AMERICAN EXPRESS CO COM 025816109 69042 551499 SH SOLE 551499 0 0 APPOLLO GLOBAL MGMT INC CL A COM 03768E105 1346 28000 SH SOLE 28000 0 0 APPLE INC COM COM 037833100
#> 7             Aflac Inc Com 001055102 426 8052 SH SOLE 8052 0 0 Aflac Inc Com 001055102 68 1286 SH DFND 1286 0 0 AIA Group Ltd Sponsored ADR Com 001317205 453 10796 SH SOLE 5920 0 4876 AT&T Inc Com 00206R102 27382 700655 SH SOLE 693384 0 7271 AT&T Inc Com 00206R102 1167 29874 SH DFND 29554 0 320 Abbott
#> 8           RBC Capital Markets Filed Pursuant to Rule 433 Registration Statement No. 333-227001 The information in this preliminary terms supplement is not complete and may be changed. Preliminary Terms Supplement Subject to Completion: Dated January 2, 2020 Pricing Supplement Dated January _, 2020 to
#> 9                             UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number: 811-6200 Schwab Investments Schwab 1000 Index Fund (Exact name of registrant as
#> 10                     UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number 811-22781 Goldman Sachs Trust II (Exact name of registrant as specified in charter) 200
#>                                                                                         parent_href
#> 1   https://www.sec.gov/Archives/edgar/data/51812/000005181220000001/xslForm13F_X01/primary_doc.xml
#> 2                                                                                              <NA>
#> 3                                                                                              <NA>
#> 4                                                                                              <NA>
#> 5                                                                                              <NA>
#> 6  https://www.sec.gov/Archives/edgar/data/805676/000110465920000356/xslForm13F_X01/primary_doc.xml
#> 7   https://www.sec.gov/Archives/edgar/data/51812/000005181220000001/xslForm13F_X01/primary_doc.xml
#> 8                                                                                              <NA>
#> 9                                                                                              <NA>
#> 10                                                                                             <NA>
#>                                                                                          index_href
#> 1   https://www.sec.gov/Archives/edgar/data/51812/000005181220000001/0000051812-20-000001-index.htm
#> 2                                                                                              <NA>
#> 3                                                                                              <NA>
#> 4                                                                                              <NA>
#> 5                                                                                              <NA>
#> 6  https://www.sec.gov/Archives/edgar/data/805676/000110465920000356/0001104659-20-000356-index.htm
#> 7   https://www.sec.gov/Archives/edgar/data/51812/000005181220000001/0000051812-20-000001-index.htm
#> 8                                                                                              <NA>
#> 9                                                                                              <NA>
#> 10                                                                                             <NA>

# }