SEC Full-Text Search
Source: R/full_text.R

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

## This can be very slow running
full_text('intel')

#>    filing_date
#> 1   2018-04-19
#> 2   2018-04-19
#> 3   2018-04-19
#> 4   2018-04-19
#> 5   2018-04-19
#> 6   2018-04-19
#> 7   2018-04-19
#> 8   2018-04-19
#> 9   2018-04-19
#> 10  2018-04-19
#>                                                                     name
#> 1           INFORMATION TABLE of 13F-HR for VALLEY NATIONAL ADVISERS INC
#> 2         INFORMATION TABLE of 13F-HR for MOUNT VERNON ASSOCIATES INC MD
#> 3            INFORMATION TABLE of 13F-HR for Fisher Asset Management LLC
#> 4           INFORMATION TABLE of 13F-HR for GILLESPIE ROBINSON GRIMM INC
#> 5        INFORMATION TABLE of 13F-HR for Endurance Wealth Management Inc
#> 6                                487 for Advisors Disciplined Trust 1864
#> 7                                487 for Advisors Disciplined Trust 1863
#> 8  INFORMATION TABLE of 13F-HR for ARCADIA INVESTMENT MANAGEMENT CORP MI
#> 9                         INFORMATION TABLE of 13F-HR for WOODSTOCK CORP
#> 10          INFORMATION TABLE of 13F-HR for UNIVEST CORP OF PENNSYLVANIA
#>                                                                                                      href
#> 1             https://www.sec.gov/Archives/edgar/data/773411/000077341118000002/xslForm13F_X01/13F-HR.xml
#> 2          https://www.sec.gov/Archives/edgar/data/799003/000079900318000002/xslForm13F_X01/13f033118.xml
#> 3               https://www.sec.gov/Archives/edgar/data/850529/000085052918000006/xslForm13F_X01/F13F.xml
#> 4    https://www.sec.gov/Archives/edgar/data/926834/000092683418000004/xslForm13F_X01/Edgar13F_033118.xml
#> 5  https://www.sec.gov/Archives/edgar/data/1476225/000095012318003639/xslForm13F_X01/form13fInfoTable.xml
#> 6                      https://www.sec.gov/Archives/edgar/data/1724656/000101322818000489/adt1864-487.txt
#> 7                      https://www.sec.gov/Archives/edgar/data/1724655/000101322818000492/adt1863-487.txt
#> 8          https://www.sec.gov/Archives/edgar/data/1016972/000101697218000002/xslForm13F_X01/inftable.xml
#> 9  https://www.sec.gov/Archives/edgar/data/1008877/000106299318001722/xslForm13F_X01/form13fInfoTable.xml
#> 10  https://www.sec.gov/Archives/edgar/data/102212/000108514618001109/xslForm13F_X01/form13fInfoTable.xml
#>                             company_name     cik  sic
#> 1           VALLEY NATIONAL ADVISERS INC  773411 <NA>
#> 2         MOUNT VERNON ASSOCIATES INC MD  799003 <NA>
#> 3            Fisher Asset Management LLC  850529 <NA>
#> 4           GILLESPIE ROBINSON GRIMM INC  926834 <NA>
#> 5        Endurance Wealth Management Inc 1476225 <NA>
#> 6        Advisors Disciplined Trust 1864 1724656 <NA>
#> 7        Advisors Disciplined Trust 1863 1724655 <NA>
#> 8  ARCADIA INVESTMENT MANAGEMENT CORP MI 1016972 <NA>
#> 9                         WOODSTOCK CORP 1008877 <NA>
#> 10          UNIVEST CORP OF PENNSYLVANIA  102212 6022
#>                                                                                                                                                                                                                                                                                                         content
#> 1       AMERICAN INTL GROUP INC *W EXP 01/19/202 026874156 1 68 PRN SOLE 0 0 68 LIBERTY PPTY TR SH BEN INT 531172104 57 1434 SH SOLE 0 0 1434 ADVISORSHARES TR NEW TECH MEDIA 00768Y552 2 104 SH SOLE 0 0 104 GENERAL MTRS CO *W EXP 07/10/201 37045V126 0 11 PRN SOLE 0 0 11 ROYAL DUTCH SHELL PLC SPONS ADR A
#> 2      AT&T Inc. com 00206R102 927 25991 SH SOLE 5125 0 20866 Alphabet Inc. Cl C com 02079K107 2236 2167 SH SOLE 235 0 1932 Amazon.com Inc. com 023135106 4132 2855 SH SOLE 500 0 2355 American Express Co. com 025816109 1705 18275 SH SOLE 2950 0 15325 American Intl Group, Inc. com 026874784 1081 19861 SH
#> 3   AAR CORP COM 000361105 56032 1270271 SH DFND 438496 0 831775 ABB LTD ADR COM 000375204 231209 9739214 SH DFND 9462251 0 276963 AFLAC INC COM 001055102 5998 137060 SH DFND 3998 0 133062 AGL ENERGY LTD ADR COM 001201102 626 37652 SH DFND 3330 0 34322 AGNC INVT CORP COM COM 00123Q104 256 13530 SH DFND
#> 4      APPLE INC. COM 037833100 33518 199772 SH SOLE 0 196234 0 3538 ABBVIE INC COM 00287Y109 737 7783 SH SOLE 0 7783 0 0 ABBOTT LABORATORIES COM 002824100 783 13074 SH SOLE 0 12774 0 300 ACCENTURE LTD - CL A COM G1151C101 25255 164530 SH SOLE 0 160940 0 3590 AUTOMATIC DATA PROCESSING COM 053015103 318
#> 5                                                                                                                                                                                                                                                                         3D SYSTEMS CORPORATIONStock88554D2050
#> 6      1933 ACT FILE NO.: 333-222250 1940 ACT FILE NO.: 811-21056 CIK NO.: 1724656 SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 AMENDMENT NO. 1 TO REGISTRATION STATEMENT ON FORM S-6 FOR REGISTRATION UNDER THE SECURITIES ACT OF 1933 OF SECURITIES OF UNIT INVESTMENT TRUSTS REGISTERED ON FORM
#> 7      1933 ACT FILE NO.: 333-222621 1940 ACT FILE NO.: 811-21056 CIK NO.: 1724655 SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 AMENDMENT NO. 2 TO REGISTRATION STATEMENT ON FORM S-6 FOR REGISTRATION UNDER THE SECURITIES ACT OF 1933 OF SECURITIES OF UNIT INVESTMENT TRUSTS REGISTERED ON FORM
#> 8       BERKSHIRE HATHAWAY INC DEL CL B NEW 084670702 237 1186 SH SOLE 1186 0 0 BERKSHIRE HATHAWAY INC DEL CL B NEW 084670702 70 350 SH SOLE 100 0 250 CENTRAL FD CDA LTD CL A 153501101 1455 110030 PRN SOLE 110030 0 0 STERLING BANCORP DEL COM 85917A100 14 606 SH SOLE 606 0 0 SPDR SER TR PORTFOLIO SH TSR
#> 9  3M Co COMMON STOCK 88579Y101 10030 45692 SH SOLE 44092 0 1600 Abbott Laboratories COMMON STOCK 002824100 7010 116982 SH SOLE 116702 0 280 Abbvie Inc COMMON STOCK 00287Y109 8668 91581 SH SOLE 89901 0 1680 Air Prods & Chems Inc COMMON STOCK 009158106 5185 32604 SH SOLE 30124 0 2480 Akamai Technologies
#> 10    3M CO COM 88579Y101 2804 12773 SH OTR 12738 0 35 ABBVIE INC COM 00287Y109 3761 39736 SH OTR 39496 0 240 AETNA INC NEW COM 00817Y108 3747 22172 SH OTR 22018 0 154 AIR PRODS & CHEMS INC COM 009158106 2997 18845 SH OTR 18796 0 49 ALLSTATE CORP COM 020002101 4307 45432 SH OTR 45304 0 128 ALPHABET INC
#>                                                                                          parent_href
#> 1   https://www.sec.gov/Archives/edgar/data/773411/000077341118000002/xslForm13F_X01/primary_doc.xml
#> 2   https://www.sec.gov/Archives/edgar/data/799003/000079900318000002/xslForm13F_X01/primary_doc.xml
#> 3   https://www.sec.gov/Archives/edgar/data/850529/000085052918000006/xslForm13F_X01/primary_doc.xml
#> 4   https://www.sec.gov/Archives/edgar/data/926834/000092683418000004/xslForm13F_X01/primary_doc.xml
#> 5  https://www.sec.gov/Archives/edgar/data/1476225/000095012318003639/xslForm13F_X01/primary_doc.xml
#> 6                                                                                               <NA>
#> 7                                                                                               <NA>
#> 8   https://www.sec.gov/Archives/edgar/data/102212/000108514618001109/xslForm13F_X01/primary_doc.xml
#> 9   https://www.sec.gov/Archives/edgar/data/773411/000077341118000002/xslForm13F_X01/primary_doc.xml
#> 10  https://www.sec.gov/Archives/edgar/data/799003/000079900318000002/xslForm13F_X01/primary_doc.xml
#>                                                                                           index_href
#> 1   https://www.sec.gov/Archives/edgar/data/773411/000077341118000002/0000773411-18-000002-index.htm
#> 2   https://www.sec.gov/Archives/edgar/data/799003/000079900318000002/0000799003-18-000002-index.htm
#> 3   https://www.sec.gov/Archives/edgar/data/850529/000085052918000006/0000850529-18-000006-index.htm
#> 4   https://www.sec.gov/Archives/edgar/data/926834/000092683418000004/0000926834-18-000004-index.htm
#> 5  https://www.sec.gov/Archives/edgar/data/1476225/000095012318003639/0000950123-18-003639-index.htm
#> 6                                                                                               <NA>
#> 7                                                                                               <NA>
#> 8   https://www.sec.gov/Archives/edgar/data/102212/000108514618001109/0001085146-18-001109-index.htm
#> 9   https://www.sec.gov/Archives/edgar/data/773411/000077341118000002/0000773411-18-000002-index.htm
#> 10  https://www.sec.gov/Archives/edgar/data/799003/000079900318000002/0000799003-18-000002-index.htm