SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(
  q = "*",
  type = "",
  reverse_order = FALSE,
  count = 10,
  page = 1,
  stemming = TRUE,
  name = "",
  cik = "",
  sic = "",
  from = "",
  to = ""
)

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

# \donttest{
## This can be very slow running
full_text('intel')

#>    filing_date
#> 1   2020-01-17
#> 2   2020-01-17
#> 3   2020-01-17
#> 4   2020-01-17
#> 5   2020-01-17
#> 6   2020-01-17
#> 7   2020-01-17
#> 8   2020-01-17
#> 9   2020-01-17
#> 10  2020-01-17
#>                                                                  name
#> 1         INFORMATION TABLE of 13F-HR for ARGUS INVESTORS COUNSEL INC
#> 2        INFORMATION TABLE of 13F-HR for FIRST NATIONAL BANK OF OMAHA
#> 3            INFORMATION TABLE of 13F-HR for HUNTINGTON NATIONAL BANK
#> 4  INFORMATION TABLE of 13F-HR for CANANDAIGUA NATIONAL BANK TRUST CO
#> 5           INFORMATION TABLE of 13F-HR for HC FINANCIAL ADVISORS INC
#> 6  INFORMATION TABLE of 13F-HR for CREATIVE FINANCIAL DESIGNS INC ADV
#> 7               INFORMATION TABLE of 13F-HR for STEWART PATTEN CO LLC
#> 8        INFORMATION TABLE of 13F-HR for WENDELL DAVID ASSOCIATES INC
#> 9          INFORMATION TABLE of 13F-HR for BTR CAPITAL MANAGEMENT INC
#> 10    INFORMATION TABLE of 13F-HR for American Research Management Co
#>                                                                                                               href
#> 1                      https://www.sec.gov/Archives/edgar/data/7195/000000719520000001/xslForm13F_X01/inftable.xml
#> 2                       https://www.sec.gov/Archives/edgar/data/36644/000003664420000001/xslForm13F_X01/PTR289.xml
#> 3              https://www.sec.gov/Archives/edgar/data/49205/000004920520000001/xslForm13F_X01/QTR42019_13F_HR.xml
#> 4  https://www.sec.gov/Archives/edgar/data/1074266/000075945820000001/xslForm13F_X01/CNB_13F_Submission_201912.xml
#> 5           https://www.sec.gov/Archives/edgar/data/784045/000078404520000001/xslForm13F_X01/hc13fexport123119.xml
#> 6               https://www.sec.gov/Archives/edgar/data/819864/000081986420000002/xslForm13F_X01/cfd13freport4.xml
#> 7                    https://www.sec.gov/Archives/edgar/data/860580/000086058020000002/xslForm13F_X01/inftable.xml
#> 8           https://www.sec.gov/Archives/edgar/data/878228/000087822820000002/xslForm13F_X01/dwasec13f12312019.xml
#> 9                        https://www.sec.gov/Archives/edgar/data/883790/000088379020000002/xslForm13F_X01/F13F.xml
#> 10                   https://www.sec.gov/Archives/edgar/data/1506796/000090901220000005/xslForm13F_X01/aci_13f.xml
#>                          company_name     cik  sic
#> 1         ARGUS INVESTORS COUNSEL INC    7195 <NA>
#> 2        FIRST NATIONAL BANK OF OMAHA   36644 <NA>
#> 3            HUNTINGTON NATIONAL BANK   49205 <NA>
#> 4  CANANDAIGUA NATIONAL BANK TRUST CO 1074266 <NA>
#> 5           HC FINANCIAL ADVISORS INC  784045 <NA>
#> 6  CREATIVE FINANCIAL DESIGNS INC ADV  819864 <NA>
#> 7               STEWART PATTEN CO LLC  860580 <NA>
#> 8        WENDELL DAVID ASSOCIATES INC  878228 <NA>
#> 9          BTR CAPITAL MANAGEMENT INC  883790 <NA>
#> 10    American Research Management Co 1506796 <NA>
#>                                                                                                                                                                                                                                                                                                         content
#> 1   3M Co COM 88579Y101 3947 22370 SH SOLE 17220 0 5150 Adobe Inc COM 00724F101 5047 15304 SH SOLE 10810 0 4494 Air Prods & Chems Inc COM 009158106 4188 17820 SH SOLE 13750 0 4070 Alphabet Inc COM 02079K305 4914 3669 SH SOLE 2540 0 1129 Amazon.com Inc COM 023135106 4958 2683 SH SOLE 1858 0 825 American
#> 2  AMBARELLA INC SHS G037AX101 1652 27278 SH SOLE 27278 0 0 AMBARELLA INC SHS G037AX101 32 525 SH OTR 525 0 0 AON PLC SHS CL A G0408V102 5830 27988 SH SOLE 25075 0 2913 AON PLC SHS CL A G0408V102 158 760 SH OTR 585 175 0 ARGO GROUP INTL HLDGS LTD COM G0464B107 1159 17632 SH SOLE 17632 0 0 ACCENTURE PLC
#> 3                                                                                                                                                                                                                                                                                                              
#> 4                                                                                                                                                                                                                                                                               Aflac, Inc. Com 001055102 1811 
#> 5   ENTERPRISE PRODS PARTNERS L COM 293792107 244 8671 SH SOLE 0 0 8671 AT&T INC COM 00206R102 261 6668 SH SOLE 0 0 6668 INOGEN INC COM 45780L104 279 4077 SH SOLE 0 0 4077 INVESCO EXCHNG TRADED FD TR FTSE RAFI DEV 46138E743 280 6588 SH SOLE 0 0 6588 MONDELEZ INTL INC CL A 609207105 364 6600 SH SOLE 0 0
#> 6       AMAZON COM INC COM Stock 023135106 1312 691 SH SOLE 0 0 691 ISHARES MORNINGSTAR MID-CAP GROWTH ETF ETF 464288307 890 3348 SH SOLE 0 0 3348 CERNER CORP COM Stock 156782104 45 615 SH SOLE 0 0 614 ISHARES NORTH AMERICAN NATURAL RESOURCES ETF 464287374 6 185 SH SOLE 0 0 184 VIACOMCBS INC CL B Stock
#> 7  3M Company COM STK 88579y101 20129 114099 SH SOLE 3043 0 111056 3M Company COM STK 88579y101 24 135 SH DFND 0 0 135 AFLAC, Inc. COM STK 001055102 801 15148 SH SOLE 0 0 15148 AFLAC, Inc. COM STK 001055102 8 150 SH DFND 0 0 150 AT & T, Inc. COM STK 00206r102 17743 454023 SH SOLE 13675 0 440348 AT & T,
#> 8          1st Trust Pfd Sec & Inc ETF DEBT 33739E108 261 13000 SH OTR 0 0 13000 3M Company COMM 88579Y101 1613 9144 SH OTR 0 0 9144 Abbott Labs COMM 002824100 16452 189413 SH OTR 0 0 189413 AbbVie COMM 00287Y109 9278 104788 SH OTR 0 0 104788 Accenture 'A' COMM G1151C101 7666 36404 SH OTR 0 0 36404 Air
#> 9    ALPS ALERIAN MLP ETF COM 00162Q866 98 11575 SH DFND 0 0 11575 AT&T INC COM 00206R102 3119 79816 SH DFND 7277 0 72539 ABBOTT LABORATORIES COM 002824100 8866 102072 SH DFND 2850 0 99222 ABBVIE INC COM 00287Y109 1792 20243 SH DFND 150 0 20093 ADOBE INC. COM 00724F101 536 1625 SH DFND 0 0 1625 AGILENT
#> 10 3M Company com 88579Y101 123 700 SH SOLE 01 02 700 0 0 Abbott Laboratories com 002824100 455 5237 SH SOLE 01 02 5237 0 0 AbbVie Inc com 00287Y109 304 3437 SH SOLE 01 02 3437 0 0 Acadia Pharmaceuticals, Inc com 004225108 4 100 SH SOLE 01 02 100 0 0 Accenture Plc com G1150G111 26 124 SH SOLE 01 02 124
#>                                                                                          parent_href
#> 1     https://www.sec.gov/Archives/edgar/data/7195/000000719520000001/xslForm13F_X01/primary_doc.xml
#> 2    https://www.sec.gov/Archives/edgar/data/36644/000003664420000001/xslForm13F_X01/primary_doc.xml
#> 3    https://www.sec.gov/Archives/edgar/data/49205/000004920520000001/xslForm13F_X01/primary_doc.xml
#> 4  https://www.sec.gov/Archives/edgar/data/1074266/000075945820000001/xslForm13F_X01/primary_doc.xml
#> 5   https://www.sec.gov/Archives/edgar/data/784045/000078404520000001/xslForm13F_X01/primary_doc.xml
#> 6   https://www.sec.gov/Archives/edgar/data/819864/000081986420000002/xslForm13F_X01/primary_doc.xml
#> 7   https://www.sec.gov/Archives/edgar/data/860580/000086058020000002/xslForm13F_X01/primary_doc.xml
#> 8   https://www.sec.gov/Archives/edgar/data/878228/000087822820000002/xslForm13F_X01/primary_doc.xml
#> 9   https://www.sec.gov/Archives/edgar/data/883790/000088379020000002/xslForm13F_X01/primary_doc.xml
#> 10 https://www.sec.gov/Archives/edgar/data/1506796/000090901220000005/xslForm13F_X01/primary_doc.xml
#>                                                                                           index_href
#> 1     https://www.sec.gov/Archives/edgar/data/7195/000000719520000001/0000007195-20-000001-index.htm
#> 2    https://www.sec.gov/Archives/edgar/data/36644/000003664420000001/0000036644-20-000001-index.htm
#> 3    https://www.sec.gov/Archives/edgar/data/49205/000004920520000001/0000049205-20-000001-index.htm
#> 4  https://www.sec.gov/Archives/edgar/data/1074266/000075945820000001/0000759458-20-000001-index.htm
#> 5   https://www.sec.gov/Archives/edgar/data/784045/000078404520000001/0000784045-20-000001-index.htm
#> 6   https://www.sec.gov/Archives/edgar/data/819864/000081986420000002/0000819864-20-000002-index.htm
#> 7   https://www.sec.gov/Archives/edgar/data/860580/000086058020000002/0000860580-20-000002-index.htm
#> 8   https://www.sec.gov/Archives/edgar/data/878228/000087822820000002/0000878228-20-000002-index.htm
#> 9   https://www.sec.gov/Archives/edgar/data/883790/000088379020000002/0000883790-20-000002-index.htm
#> 10 https://www.sec.gov/Archives/edgar/data/1506796/000090901220000005/0000909012-20-000005-index.htm

# }