SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

full_text('intel')

#>    filing_date
#> 1   2018-01-12
#> 2   2018-01-12
#> 3   2018-01-12
#>                                                                       name
#> 1                                             N-CSRS for World Funds Trust
#> 2             INFORMATION TABLE of 13F-HR for NISA INVESTMENT ADVISORS LLC
#> 3                 INFORMATION TABLE of 13F-HR for APPLETON PARTNERS INC MA
#>                                                                                                      href
#> 1                    https://www.sec.gov/Archives/edgar/data/1396092/000089180418000025/wft72578-ncsr.htm
#> 2        https://www.sec.gov/Archives/edgar/data/937615/000093761518000003/xslForm13F_X01/13F12312017.xml
#> 3  https://www.sec.gov/Archives/edgar/data/1055290/000095012318000184/xslForm13F_X01/form13fInfoTable.xml
#>                               company_name     cik  sic
#> 1                        World Funds Trust 1396092 <NA>
#> 2             NISA INVESTMENT ADVISORS LLC  937615 <NA>
#> 3                 APPLETON PARTNERS INC MA 1055290 <NA>
#>                                                                                                                                                                                                                                                                                                         content
#> 1                    UNITED STATES SECURITIES AND EXCHANGE COMMISSION WASHINGTON, D.C. 20549 FORM N-CSR CERTIFIED SHAREHOLDER REPORT OF REGISTERED MANAGEMENT INVESTMENT COMPANIES Investment Company Act file number: 811-22172 Exact name of registrant as specified in charter: World Funds Trust Address of
#> 2        AAC HOLDINGS INC COM 000307108 89 9900 SH SOLE 9900 0 0 AAON INC COM 000360206 15 420 SH SOLE 420 0 0 AAR CORP COM 000361105 201 5120 SH SOLE 5120 0 0 ACCO BRANDS CORP COM 00081T108 245 20120 SH SOLE 20120 0 0 ABM INDS INC COM 000957100 390 10349 SH SOLE 10349 0 0 AFLAC INC COM 001055102 14693
#> 3                                                                                                                                                                                                                                                                                     3M CO COMCOM88579Y1013839
#>                                                                                          parent_href
#> 1                                                                                               <NA>
#> 2  https://www.sec.gov/Archives/edgar/data/1055290/000095012318000184/xslForm13F_X01/primary_doc.xml
#> 3  https://www.sec.gov/Archives/edgar/data/1569452/000106299318000139/xslForm13F_X01/primary_doc.xml
#>                                                                                           index_href
#> 1                                                                                               <NA>
#> 2  https://www.sec.gov/Archives/edgar/data/1055290/000095012318000184/0000950123-18-000184-index.htm
#> 3  https://www.sec.gov/Archives/edgar/data/1569452/000106299318000139/0001062993-18-000139-index.htm
#>  [ reached getOption("max.print") -- omitted 7 rows ]