SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

full_text('intel')

#>    filing_date                                             name
#> 1   2017-12-20                           6-K for BLACKBERRY Ltd
#> 2   2017-12-19                                 424B2 for UBS AG
#> 3   2017-12-19                      POS AM for Yatra Online Inc
#>                                                                                       href
#> 1  https://www.sec.gov/Archives/edgar/data/1070235/000107023517000048/q318pressrelease.htm
#> 2  https://www.sec.gov/Archives/edgar/data/1114446/000091412117001843/ub39902383-424b2.htm
#> 3  https://www.sec.gov/Archives/edgar/data/1516899/000110465917074055/a17-28625_1posam.htm
#>                                company_name     cik  sic
#> 1                            BLACKBERRY Ltd 1070235 7372
#> 2                                    UBS AG 1114446 6021
#> 3                          Yatra Online Inc 1516899 4700
#>                                                                                                                                                                                                                                                                                                     content
#> 1                                                                                                                            Intel is collaborating in the development of this product Q3 Results Non-GAAP revenue for the third quarter of fiscal 2018 was $235 million with GAAP revenue of $226 million.
#> 2    PRICING SUPPLEMENT Dated December 15, 2017 Filed Pursuant to Rule 424(b)(2) Registration Statement No. 333-204908 (To Prospectus dated April 29, 2016 and Product Supplement dated September 30, 2016) UBS AG $5, 145, 000 Step Down Trigger Autocallable Notes Linked to the least performing between
#> 3                   As filed with the United States Securities and Exchange Commission on December 19, 2017. Registration No. 333-215653 UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 POST-EFFECTIVE AMENDMENT NO. 4 TO FORM F-1 REGISTRATION STATEMENT UNDER THE SECURITIES ACT
#>    parent_href index_href
#> 1         <NA>       <NA>
#> 2         <NA>       <NA>
#> 3         <NA>       <NA>
#>  [ reached getOption("max.print") -- omitted 7 rows ]