SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(
  q = "*",
  type = "",
  reverse_order = FALSE,
  count = 100,
  page = 1,
  stemming = TRUE,
  name = "",
  cik = "",
  sic = "",
  from = "",
  to = "",
  location = "",
  incorporated_location = FALSE
)

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'. Can also be a list of types - e.g. c("10-K", "10-Q")

reverse_order
[DEP] If true, order by oldest first instead of newest first

count
[DEP] Number of results to return - will always try to return 100

page
Which page of results to return

stemming
[DEP] Search by base words(default) or exactly as entered

name
Company name OR individual's name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
[DEP] Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

location
Filter based on company's location

incorporated_location
boolean to use location of incorporation rather than location of HQ

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

# \donttest{
try(full_text('intel'))

#>    filing_date                    name
#> 1   2007-08-06            10-Q EX-10.3
#> 2   2008-05-02            10-Q EX-10.1
#> 3   2007-08-03             6-K EX-99.1
#> 4   2021-03-23             8-K EX-99.1
#> 5   2001-03-26                   424B2
#> 6   2008-05-02            10-Q EX-10.2
#> 7   2019-05-09             8-K EX-99.3
#> 8   2003-04-24                   424B2
#> 9   2003-05-12                   424B3
#> 10  2004-03-31                 DEF 14A
#> 11  2013-04-26            10-Q EX-10.2
#> 12  2000-10-27            S-1 EX-10.21
#> 13  2003-05-29                   424B3
#> 14  2019-05-09             8-K EX-99.4
#> 15  2016-08-19               DRS EX-10
#> 16  2000-08-11              10-Q EX-10
#> 17  2007-11-27            S-1 EX-10.23
#> 18  2009-03-26           10-K EX-10.27
#> 19  2005-04-19             8-K EX-99.1
#> 20  2000-11-13            S-1 EX-10.21
#> 21  2008-08-22            S-4 EX-10.60
#> 22  2015-06-01             8-K EX-99.1
#> 23  2017-01-11            DRS EX-10.12
#> 24  2012-02-23            10-K EX-21.1
#> 25  2009-06-16            S-1 EX-10.16
#> 26  2011-10-18             8-K EX-99.1
#> 27  2007-03-27                   424B2
#> 28  2017-05-16            DRS EX-10.12
#> 29  2010-05-12             8-K EX-99.2
#> 30  2004-03-30             8-K EX-99.1
#> 31  2011-01-31             8-K EX-99.2
#> 32  2011-07-20             8-K EX-99.1
#> 33  2002-07-30                   424B3
#> 34  2004-01-14               8-K EX-99
#> 35  2011-01-31             8-K EX-99.1
#> 36  2015-11-20             8-K EX-99.1
#> 37  2001-03-08       SC TO-T SC TO-T/A
#> 38  2007-04-10          10-Q EX-10.164
#> 39  2015-01-27           10-K EX-10.13
#> 40  2009-11-17             8-K EX-10.1
#> 41  2009-09-22 SC TO-I EX-99.(A)(1)(G)
#> 42  2011-01-24             8-K EX-99.1
#> 43  2002-08-15                   424B3
#> 44  2007-08-06            10-Q EX-10.4
#> 45  2002-03-13                    10-K
#> 46  2008-03-17           10-K EX-10.38
#> 47  2011-03-14             S-1 EX-10.8
#> 48  2019-04-03             8-K EX-10.1
#> 49  2008-03-04               8-K EX-99
#> 50  2001-12-04        10-K405 EX-10.15
#> 51  2020-10-20                     8-K
#> 52  2002-04-30             8-K EX-10.1
#> 53  2004-02-23            10-K EX-21.1
#> 54  2019-01-31             8-K EX-99.1
#> 55  2016-11-14               DRS EX-10
#> 56  2001-01-16                 SC TO-C
#> 57  2007-01-08            S-1 EX-10.58
#> 58  2007-02-20                   424B2
#> 59  2007-11-15               8-K EX-99
#> 60  2012-11-20          10-Q EX-10.113
#> 61  2004-02-17          8-A12B EX-99.B
#> 62  2009-11-12             8-K EX-10.1
#> 63  2017-06-22       SC TO-T SC TO-T/A
#> 64  2009-06-26            S-1 EX-10.16
#> 65  2021-01-14             8-K EX-10.1
#> 66  2016-02-12            10-K EX-21.1
#> 67  2007-12-11            S-1 EX-10.23
#> 68  2004-07-16            SB-2 EX-10.9
#> 69  2015-02-13            10-K EX-21.1
#> 70  2015-07-27            10-Q EX-10.4
#> 71  2017-03-31            S-1 EX-10.12
#> 72  2008-11-17           10-Q EX-10.42
#> 73  2007-01-08            S-1 EX-10.52
#> 74  2019-01-31             8-K EX-10.1
#> 75  2006-02-27            10-K EX-21.1
#> 76  2001-08-10           10-Q EX-10.58
#> 77  2004-10-12             8-K EX-99.1
#> 78  2015-11-20             8-K EX-99.3
#> 79  2015-01-15             8-K EX-99.1
#> 80  2017-10-06            S-1 EX-10.12
#> 81  2007-02-20                     FWP
#> 82  2010-05-12             8-K EX-99.1
#> 83  2020-08-14             8-K EX-10.1
#> 84  2009-06-04                 CORRESP
#> 85  2016-02-12           10-K EX-10.14
#> 86  2012-07-09             8-K EX-99.1
#> 87  2005-02-22            10-K EX-21.1
#> 88  2015-07-02             8-K EX-99.1
#> 89  2021-01-14                     8-K
#> 90  2018-10-25             8-K EX-99.1
#> 91  2015-10-27           10-K EX-10.51
#> 92  2003-02-13         SC 13G SC 13G/A
#> 93  2007-12-06            S-1 EX-10.23
#> 94  2007-08-03            S-1 EX-10.23
#> 95  2007-06-29          SC 13D EX-99.A
#> 96  2017-08-08       SC TO-T SC TO-T/A
#> 97  2017-03-31             S-1 EX-4.04
#> 98  2010-08-05            10-Q EX-10.1
#> 99  2011-05-17             8-K EX-99.1
#>                                                                                                href
#> 1                https://www.sec.gov/Archives/edgar/data/50863/000095013407016808/f32127exv10w3.htm
#> 2                https://www.sec.gov/Archives/edgar/data/50863/000095013408008205/f40150exv10w1.htm
#> 3               https://www.sec.gov/Archives/edgar/data/932787/000095012307010792/y01797exv99w1.htm
#> 4                https://www.sec.gov/Archives/edgar/data/50863/000119312521091374/d153275dex991.htm
#> 5              https://www.sec.gov/Archives/edgar/data/200245/000095012301002657/e45834b2e424b2.txt
#> 6                https://www.sec.gov/Archives/edgar/data/50863/000095013408008205/f40150exv10w2.htm
#> 7                https://www.sec.gov/Archives/edgar/data/50863/000119312519142117/d741646dex993.htm
#> 8              https://www.sec.gov/Archives/edgar/data/200245/000095012303004687/y85661b2e424b2.txt
#> 9               https://www.sec.gov/Archives/edgar/data/895421/000095010303001288/may0603_424b3.txt
#> 10                     https://www.sec.gov/Archives/edgar/data/50863/000119312504054199/ddef14a.htm
#> 11    https://www.sec.gov/Archives/edgar/data/768251/000076825113000023/altera_2013q1x10qxex102.htm
#> 12           https://www.sec.gov/Archives/edgar/data/1022652/000109581100004115/a62761a6ex10-21.txt
#> 13         https://www.sec.gov/Archives/edgar/data/895421/000095010303001359/may2303_424b3-ps48.txt
#> 14               https://www.sec.gov/Archives/edgar/data/50863/000119312519142117/d741646dex994.htm
#> 15                 https://www.sec.gov/Archives/edgar/data/1316016/000095012316020279/filename3.htm
#> 16                      https://www.sec.gov/Archives/edgar/data/932290/000091205700036562/ex-10.htm
#> 17                   https://www.sec.gov/Archives/edgar/data/1120105/000119312507253790/dex1023.htm
#> 18            https://www.sec.gov/Archives/edgar/data/1442505/000095013409006072/v51173exv10w27.htm
#> 19               https://www.sec.gov/Archives/edgar/data/50863/000115752305003483/a4867865ex991.txt
#> 20           https://www.sec.gov/Archives/edgar/data/1022652/000095015000000898/v62761b4ex10-21.txt
#> 21            https://www.sec.gov/Archives/edgar/data/1442505/000095012308009987/y63518exv10w60.htm
#> 22               https://www.sec.gov/Archives/edgar/data/50863/000119312515209231/d935756dex991.htm
#> 23                 https://www.sec.gov/Archives/edgar/data/1316016/000095012317004142/filename2.htm
#> 24               https://www.sec.gov/Archives/edgar/data/50863/000119312512075534/d302695dex211.htm
#> 25          https://www.sec.gov/Archives/edgar/data/1420302/000095012309014140/b75316a8exv10w16.htm
#> 26             https://www.sec.gov/Archives/edgar/data/50863/000115752311005759/a50033968-ex991.htm
#> 27                    https://www.sec.gov/Archives/edgar/data/1318281/000119312507064810/d424b2.htm
#> 28                 https://www.sec.gov/Archives/edgar/data/1316016/000095012317005440/filename2.htm
#> 29                      https://www.sec.gov/Archives/edgar/data/50863/000005086310000050/exh992.htm
#> 30       https://www.sec.gov/Archives/edgar/data/351145/000035114504000009/intel-dell_agreement.pdf
#> 31                      https://www.sec.gov/Archives/edgar/data/50863/000005086311000020/exh992.htm
#> 32              https://www.sec.gov/Archives/edgar/data/50863/000115752311004100/a6800886ex99_1.htm
#> 33              https://www.sec.gov/Archives/edgar/data/895421/000095010302000756/jul2602_424b3.txt
#> 34                https://www.sec.gov/Archives/edgar/data/50863/000115752304000218/a4550018ex99.txt
#> 35                      https://www.sec.gov/Archives/edgar/data/50863/000005086311000020/exh991.htm
#> 36                https://www.sec.gov/Archives/edgar/data/50863/000119312515382423/d90133dex991.htm
#> 37             https://www.sec.gov/Archives/edgar/data/50863/000089161801500166/f68556a6scto-ta.txt
#> 38         https://www.sec.gov/Archives/edgar/data/723125/000072312507000028/q2-07exhibit10_164.htm
#> 39                   https://www.sec.gov/Archives/edgar/data/1171012/000118518515000158/ex10-13.htm
#> 40                       https://www.sec.gov/Archives/edgar/data/2488/000119312509236705/dex101.htm
#> 41       https://www.sec.gov/Archives/edgar/data/50863/000095012309044666/f53534exv99wxayx1yxgy.htm
#> 42                      https://www.sec.gov/Archives/edgar/data/50863/000005086311000013/exh991.htm
#> 43           https://www.sec.gov/Archives/edgar/data/895421/000095010302000842/aug1202_424b3ps7.txt
#> 44               https://www.sec.gov/Archives/edgar/data/50863/000095013407016808/f32127exv10w4.htm
#> 45               https://www.sec.gov/Archives/edgar/data/50863/000091205702009698/a2072777z10-k.htm
#> 46            https://www.sec.gov/Archives/edgar/data/1001193/000095013408004919/f38968exv10w38.htm
#> 47                    https://www.sec.gov/Archives/edgar/data/1280363/000119312511065181/dex108.htm
#> 48               https://www.sec.gov/Archives/edgar/data/50863/000119312519096036/d642676dex101.htm
#> 49                       https://www.sec.gov/Archives/edgar/data/50863/000005086308000089/exh99.htm
#> 50                    https://www.sec.gov/Archives/edgar/data/917273/000101287001503032/dex1015.txt
#> 51                   https://www.sec.gov/Archives/edgar/data/50863/000119312520272580/d76122d8k.htm
#> 52         https://www.sec.gov/Archives/edgar/data/351145/000035114502000021/redactedsettlement.txt
#> 53                      https://www.sec.gov/Archives/edgar/data/50863/000119312504027430/dex211.htm
#> 54               https://www.sec.gov/Archives/edgar/data/50863/000119312519023301/d702987dex991.htm
#> 55                 https://www.sec.gov/Archives/edgar/data/1316016/000095012317004138/filename2.htm
#> 56              https://www.sec.gov/Archives/edgar/data/50863/000089161801000052/f68556c1scto-c.txt
#> 57          https://www.sec.gov/Archives/edgar/data/1285551/000089102007000003/v25599a1exv10w58.txt
#> 58                    https://www.sec.gov/Archives/edgar/data/1318281/000119312507034882/d424b2.htm
#> 59                      https://www.sec.gov/Archives/edgar/data/50863/000005086307000402/exh991.htm
#> 60        https://www.sec.gov/Archives/edgar/data/723125/000072312512000163/a2012q310qaex10-113.htm
#> 61                      https://www.sec.gov/Archives/edgar/data/65100/000119312504025255/dex99b.htm
#> 62                      https://www.sec.gov/Archives/edgar/data/50863/000005086309000213/exh101.htm
#> 63              https://www.sec.gov/Archives/edgar/data/50863/000119312517209888/d413660dsctota.htm
#> 64          https://www.sec.gov/Archives/edgar/data/1420302/000095012309017935/b75316a9exv10w16.htm
#> 65                https://www.sec.gov/Archives/edgar/data/50863/000119312521009142/d82915dex101.htm
#> 66             https://www.sec.gov/Archives/edgar/data/50863/000005086316000105/ex21112262015q4.htm
#> 67                   https://www.sec.gov/Archives/edgar/data/1120105/000119312507262676/dex1023.htm
#> 68             https://www.sec.gov/Archives/edgar/data/1283073/000093041304003280/c32316_ex10-9.txt
#> 69           https://www.sec.gov/Archives/edgar/data/50863/000005086315000015/a10kex21112272014.htm
#> 70         https://www.sec.gov/Archives/edgar/data/50863/000005086315000055/intcex10410q6272015.htm
#> 71            https://www.sec.gov/Archives/edgar/data/1535379/000162828017003221/s-1exhibit1012.htm
#> 72            https://www.sec.gov/Archives/edgar/data/1001193/000089161808000487/f50377exv10w42.htm
#> 73          https://www.sec.gov/Archives/edgar/data/1285551/000089102007000003/v25599a1exv10w52.txt
#> 74               https://www.sec.gov/Archives/edgar/data/50863/000119312519023301/d702987dex101.htm
#> 75               https://www.sec.gov/Archives/edgar/data/50863/000089161806000089/f12963exv21w1.htm
#> 76                      https://www.sec.gov/Archives/edgar/data/2488/000101287001501532/dex1058.txt
#> 77               https://www.sec.gov/Archives/edgar/data/50863/000115752304009185/a4738419ex991.txt
#> 78                https://www.sec.gov/Archives/edgar/data/50863/000119312515382423/d90133dex993.htm
#> 79       https://www.sec.gov/Archives/edgar/data/50863/000005086315000009/earningsreleaseq42014.htm
#> 80            https://www.sec.gov/Archives/edgar/data/1316016/000119312517306042/d106686dex1012.htm
#> 81                      https://www.sec.gov/Archives/edgar/data/1318281/000119312507034886/dfwp.htm
#> 82                      https://www.sec.gov/Archives/edgar/data/50863/000005086310000050/exh991.htm
#> 83               https://www.sec.gov/Archives/edgar/data/50863/000119312520220333/d229439dex101.htm
#> 84                   https://www.sec.gov/Archives/edgar/data/50863/000005086309000116/filename1.htm
#> 85            https://www.sec.gov/Archives/edgar/data/50863/000005086316000105/ex101412262015q4.htm
#> 86               https://www.sec.gov/Archives/edgar/data/50863/000119312512297577/d372356dex991.htm
#> 87                      https://www.sec.gov/Archives/edgar/data/50863/000119312505033162/dex211.htm
#> 88                      https://www.sec.gov/Archives/edgar/data/50863/000005086315000044/exh991.htm
#> 89                   https://www.sec.gov/Archives/edgar/data/50863/000119312521009142/d82915d8k.htm
#> 90     https://www.sec.gov/Archives/edgar/data/50863/000005086318000027/q3-2018_earningsrelease.htm
#> 91 https://www.sec.gov/Archives/edgar/data/723125/000072312515000112/a2015q4ex10-51supplyagreem.htm
#> 92                    https://www.sec.gov/Archives/edgar/data/50863/000005086303000053/pccwtest.txt
#> 93                   https://www.sec.gov/Archives/edgar/data/1120105/000119312507260104/dex1023.htm
#> 94                   https://www.sec.gov/Archives/edgar/data/1120105/000119312507170681/dex1023.htm
#> 95             https://www.sec.gov/Archives/edgar/data/1388447/000089102007000191/v31483exv99wa.txt
#> 96              https://www.sec.gov/Archives/edgar/data/50863/000119312517251345/d439645dsctota.htm
#> 97             https://www.sec.gov/Archives/edgar/data/1535379/000162828017003221/s-1exhibit404.htm
#> 98             https://www.sec.gov/Archives/edgar/data/1442505/000095012310073483/v55909exv10w1.htm
#> 99                      https://www.sec.gov/Archives/edgar/data/50863/000005086311000094/exh991.htm
#>                                                             company_name
#> 1                                                     INTEL CORP  (INTC)
#> 2                                                     INTEL CORP  (INTC)
#> 3                                           STMICROELECTRONICS NV  (STM)
#> 4                                                     INTEL CORP  (INTC)
#> 5                                      SALOMON SMITH BARNEY HOLDINGS INC
#> 6                                                     INTEL CORP  (INTC)
#> 7                                                     INTEL CORP  (INTC)
#> 8                                  CITIGROUP GLOBAL MARKETS HOLDINGS INC
#> 9  MORGAN STANLEY  (MS, CNY, INR, URR, MLPY, MS-PA, MS-PI, MS-PK, MS-PL)
#> 10                                                    INTEL CORP  (INTC)
#> 11                                                           ALTERA CORP
#> 12                                          NOVATEL WIRELESS INC  (INSG)
#> 13 MORGAN STANLEY  (MS, CNY, INR, URR, MLPY, MS-PA, MS-PI, MS-PK, MS-PL)
#> 14                                                    INTEL CORP  (INTC)
#> 15                                                         AQUANTIA CORP
#> 16                                        CENTERSPAN COMMUNICATIONS CORP
#> 17                                                         INTELLON CORP
#> 18                                                    Clearwire Corp /DE
#> 19                                                    INTEL CORP  (INTC)
#> 20                                          NOVATEL WIRELESS INC  (INSG)
#> 21                                                    New Clearwire CORP
#> 22                                                    INTEL CORP  (INTC)
#> 23                                                         AQUANTIA CORP
#> 24                                                    INTEL CORP  (INTC)
#> 25                                                 LogMeIn, Inc.  (LOGM)
#> 26                                                    INTEL CORP  (INTC)
#> 27                                                Citigroup Funding Inc.
#> 28                                                         AQUANTIA CORP
#> 29                                                    INTEL CORP  (INTC)
#> 30                                                       INTERGRAPH CORP
#> 31                                                    INTEL CORP  (INTC)
#> 32                                                    INTEL CORP  (INTC)
#> 33 MORGAN STANLEY  (MS, CNY, INR, URR, MLPY, MS-PA, MS-PI, MS-PK, MS-PL)
#> 34                                                    INTEL CORP  (INTC)
#> 35                                                    INTEL CORP  (INTC)
#> 36                                                    INTEL CORP  (INTC)
#> 37                                                    INTEL CORP  (INTC)
#> 38                                           MICRON TECHNOLOGY INC  (MU)
#> 39                                                     Uni-Pixel  (UNXL)
#> 40                                     ADVANCED MICRO DEVICES INC  (AMD)
#> 41                                                    INTEL CORP  (INTC)
#> 42                                                    INTEL CORP  (INTC)
#> 43 MORGAN STANLEY  (MS, CNY, INR, URR, MLPY, MS-PA, MS-PI, MS-PK, MS-PL)
#> 44                                                    INTEL CORP  (INTC)
#> 45                                                    INTEL CORP  (INTC)
#> 46                                                        TRANSMETA CORP
#> 47                                           APACHE DESIGN SOLUTIONS INC
#> 48                                                    INTEL CORP  (INTC)
#> 49                                                    INTEL CORP  (INTC)
#> 50                                                    RAMBUS INC  (RMBS)
#> 51                                                    INTEL CORP  (INTC)
#> 52                                                       INTERGRAPH CORP
#> 53                                                    INTEL CORP  (INTC)
#> 54                                                    INTEL CORP  (INTC)
#> 55                                                         AQUANTIA CORP
#> 56                                                    INTEL CORP  (INTC)
#> 57                                                        CLEARWIRE CORP
#> 58                                                Citigroup Funding Inc.
#> 59                                                    INTEL CORP  (INTC)
#> 60                                           MICRON TECHNOLOGY INC  (MU)
#> 61                                                MERRILL LYNCH & CO INC
#> 62                                                    INTEL CORP  (INTC)
#> 63                                                    INTEL CORP  (INTC)
#> 64                                                 LogMeIn, Inc.  (LOGM)
#> 65                                                    INTEL CORP  (INTC)
#> 66                                                    INTEL CORP  (INTC)
#> 67                                                         INTELLON CORP
#> 68                                                          GURUNET CORP
#> 69                                                    INTEL CORP  (INTC)
#> 70                                                    INTEL CORP  (INTC)
#> 71                                                Cloudera, Inc.  (CLDR)
#> 72                                                        TRANSMETA CORP
#> 73                                                        CLEARWIRE CORP
#> 74                                                    INTEL CORP  (INTC)
#> 75                                                    INTEL CORP  (INTC)
#> 76                                     ADVANCED MICRO DEVICES INC  (AMD)
#> 77                                                    INTEL CORP  (INTC)
#> 78                                                    INTEL CORP  (INTC)
#> 79                                                    INTEL CORP  (INTC)
#> 80                                                         AQUANTIA CORP
#> 81                                                Citigroup Funding Inc.
#> 82                                                    INTEL CORP  (INTC)
#> 83                                                    INTEL CORP  (INTC)
#> 84                                                    INTEL CORP  (INTC)
#> 85                                                    INTEL CORP  (INTC)
#> 86                                                    INTEL CORP  (INTC)
#> 87                                                    INTEL CORP  (INTC)
#> 88                                                    INTEL CORP  (INTC)
#> 89                                                    INTEL CORP  (INTC)
#> 90                                                    INTEL CORP  (INTC)
#> 91                                           MICRON TECHNOLOGY INC  (MU)
#> 92                                                    INTEL CORP  (INTC)
#> 93                                                         INTELLON CORP
#> 94                                                         INTELLON CORP
#> 95                                             Eagle River Holdings, LLC
#> 96                                                    INTEL CORP  (INTC)
#> 97                                                Cloudera, Inc.  (CLDR)
#> 98                                                    Clearwire Corp /DE
#> 99                                                    INTEL CORP  (INTC)
#>           cik  sic
#> 1  0000050863 3674
#> 2  0000050863 3674
#> 3  0000932787 3674
#> 4  0000050863 3674
#> 5  0000200245 6211
#> 6  0000050863 3674
#> 7  0000050863 3674
#> 8  0000200245 6211
#> 9  0000895421 6199
#> 10 0000050863 3674
#> 11 0000768251 3674
#> 12 0001022652 7370
#> 13 0000895421 6199
#> 14 0000050863 3674
#> 15 0001316016 3674
#> 16 0000932290 3577
#> 17 0001120105 3674
#> 18 0001442505 4899
#> 19 0000050863 3674
#> 20 0001022652 7370
#> 21 0001442505 <NA>
#> 22 0000050863 3674
#> 23 0001316016 3674
#> 24 0000050863 3674
#> 25 0001420302 7372
#> 26 0000050863 3674
#> 27 0001318281 6199
#> 28 0001316016 3674
#> 29 0000050863 3674
#> 30 0000351145 7373
#> 31 0000050863 3674
#> 32 0000050863 3674
#> 33 0000895421 6199
#> 34 0000050863 3674
#> 35 0000050863 3674
#> 36 0000050863 3674
#> 37 0000050863 3674
#> 38 0000723125 3674
#> 39 0001171012 3679
#> 40 0000002488 3674
#> 41 0000050863 3674
#> 42 0000050863 3674
#> 43 0000895421 6199
#> 44 0000050863 3674
#> 45 0000050863 3674
#> 46 0001001193 3674
#> 47 0001280363 <NA>
#> 48 0000050863 3674
#> 49 0000050863 3674
#> 50 0000917273 3674
#> 51 0000050863 3674
#> 52 0000351145 7373
#> 53 0000050863 3674
#> 54 0000050863 3674
#> 55 0001316016 3674
#> 56 0000050863 3674
#> 57 0001285551 4899
#> 58 0001318281 6199
#> 59 0000050863 3674
#> 60 0000723125 3674
#> 61 0000065100 6211
#> 62 0000050863 3674
#> 63 0000050863 3674
#> 64 0001420302 7372
#> 65 0000050863 3674
#> 66 0000050863 3674
#> 67 0001120105 3674
#> 68 0001283073 7389
#> 69 0000050863 3674
#> 70 0000050863 3674
#> 71 0001535379 7372
#> 72 0001001193 3674
#> 73 0001285551 4899
#> 74 0000050863 3674
#> 75 0000050863 3674
#> 76 0000002488 3674
#> 77 0000050863 3674
#> 78 0000050863 3674
#> 79 0000050863 3674
#> 80 0001316016 3674
#> 81 0001318281 6199
#> 82 0000050863 3674
#> 83 0000050863 3674
#> 84 0000050863 3674
#> 85 0000050863 3674
#> 86 0000050863 3674
#> 87 0000050863 3674
#> 88 0000050863 3674
#> 89 0000050863 3674
#> 90 0000050863 3674
#> 91 0000723125 3674
#> 92 0000050863 3674
#> 93 0001120105 3674
#> 94 0001120105 3674
#> 95 0001388447 4899
#> 96 0000050863 3674
#> 97 0001535379 7372
#> 98 0001442505 4899
#> 99 0000050863 3674
#>                                                                                 content
#> 1                                                                          EXHIBIT 10.3
#> 2                                                                          EXHIBIT 10.1
#> 3                                                                          EXHIBIT 99.1
#> 4                                                                               EX-99.1
#> 5                                                    SALOMON SMITH BARNEY HOLDINGS INC.
#> 6                                                                          EXHIBIT 10.2
#> 7                                                                               EX-99.3
#> 8                                                 CITIGROUP GLOBAL MARKETS HOLDINGS INC
#> 9                                                                                  <NA>
#> 10                                                           DEFINITIVE PROXY MATERIALS
#> 11           INTEL CORP AMENDED AND RESTATED FOUNDRY AND MANUFACTURING CUSTOMER AGREEME
#> 12                                                                        EXHIBIT 10.21
#> 13                                                                                 <NA>
#> 14                                                                              EX-99.4
#> 15                                                                                 <NA>
#> 16                                                                           EXHIBIT 10
#> 17                                                                        EXHIBIT 10.23
#> 18                                                                             EX-10.27
#> 19                                                       INTEL CORPORATION EXHIBIT 99.1
#> 20                                                                        EXHIBIT 10.21
#> 21                                                                        EXHIBIT 10.60
#> 22                                                                              EX-99.1
#> 23                                                                                 <NA>
#> 24                                                       INTEL CORPORATION SUBSIDIARIES
#> 25                              EX-10.16 - CONNECTIVITY SERVICE AND MARKETING AGREEMENT
#> 26                                                                         EXHIBIT 99.1
#> 27                                                             FINAL PRICING SUPPLEMENT
#> 28                                                                                 <NA>
#> 29                                                                         EXHIBIT 99.2
#> 30               UNOFFICIAL COPY OF SETTLEMENT AGREEMENT WITH INTEL AND LICENSE TO DELL
#> 31                                                                         EXHIBIT 99.2
#> 32                                                                         EXHIBIT 99.1
#> 33                                                                                 <NA>
#> 34                                                           EXHIBIT 99.1 PRESS RELEASE
#> 35                                                                         EXHIBIT 99.1
#> 36                                                                              EX-99.1
#> 37                                                     AMENDMENT NO. 6 TO SCHEDULE TO-T
#> 38                                                                       EXHIBIT 10.164
#> 39                                                                             EX-10.13
#> 40                                                                 SETTLEMENT AGREEMENT
#> 41                                                                      EX-99.(A)(1)(G)
#> 42                                                                         EXHIBIT 99.1
#> 43                                                                                 <NA>
#> 44                                                                         EXHIBIT 10.4
#> 45                                                                                 10-K
#> 46                                                                        EXHIBIT 10.38
#> 47                                           DEVELOPMENT, MARKETING & LICENSE AGREEMENT
#> 48                                                                              EX-10.1
#> 49                                                                         EXHIBIT 99.1
#> 50                                                             PATENT LICENSE AGREEMENT
#> 51                                                                                  8-K
#> 52                                                                 SETTLEMENT AGREEMENT
#> 53                                                                   INTEL SUBSIDIARIES
#> 54                                                                              EX-99.1
#> 55                                                                                 <NA>
#> 56                                                                        SCHEDULE TO-C
#> 57                                                                SIDE LETTER AGREEMENT
#> 58                                                       PRELIMINARY PRICING SUPPLEMENT
#> 59                                                                         EXHIBIT 99.1
#> 60                                                                       EXHIBIT 10.113
#> 61                                                                         FORM OF NOTE
#> 62                                                                         EXHIBIT 10.1
#> 63                                                                            SC TO-T/A
#> 64 EX-10.16 CONNECTIVITY SERVICE AND MARKETING AGREEMENT, DATED AS OF DECEMBER 26, 2007
#> 65                                                                              EX-10.1
#> 66                                                                         EXHIBIT 21.1
#> 67                                                                        EXHIBIT 10.23
#> 68                                                                                 <NA>
#> 69                                                                         EXHIBIT 21.1
#> 70                                                                         EXHIBIT 10.4
#> 71                                                                        EXHIBIT 10.12
#> 72                                                                        EXHIBIT 10.42
#> 73                                         MOBILE WIMAX NETWORK COLLABORATION AGREEMENT
#> 74                                                                              EX-10.1
#> 75                                               EX-21.1 INTEL CORPORATION SUBSIDIARIES
#> 76                                                       PATENT CROSS-LICENSE AGREEMENT
#> 77                                                                        PRESS RELEASE
#> 78                                                                              EX-99.3
#> 79                                                                              EXHIBIT
#> 80                                                                             EX-10.12
#> 81                                                                     OFFERING SUMMARY
#> 82                                                                         EXHIBIT 99.1
#> 83                                                                              EX-10.1
#> 84                                                                                 <NA>
#> 85                                                                        EXHIBIT 10.14
#> 86                                                                        PRESS RELEASE
#> 87                                                                   INTEL SUBSIDIARIES
#> 88                                                                         EXHIBIT 99.1
#> 89                                                                                  8-K
#> 90                                                                         EXHIBIT 99.1
#> 91                                                                        EXHIBIT 10.51
#> 92                                                                                 <NA>
#> 93                                                                        EXHIBIT 10.23
#> 94                                                                        EXHIBIT 10.23
#> 95                                                                         EXHIBIT 99.A
#> 96                                                                            SC TO-T/A
#> 97                                                                         EXHIBIT 4.04
#> 98                                                                              EX-10.1
#> 99                                                                         EXHIBIT 99.1
#>                                                                                          parent_href
#> 1    https://www.sec.gov/Archives/edgar/data/50863/000095013407016808/0000950134-07-016808-index.htm
#> 2    https://www.sec.gov/Archives/edgar/data/50863/000095013408008205/0000950134-08-008205-index.htm
#> 3   https://www.sec.gov/Archives/edgar/data/932787/000095012307010792/0000950123-07-010792-index.htm
#> 4    https://www.sec.gov/Archives/edgar/data/50863/000119312521091374/0001193125-21-091374-index.htm
#> 5   https://www.sec.gov/Archives/edgar/data/200245/000095012301002657/0000950123-01-002657-index.htm
#> 6    https://www.sec.gov/Archives/edgar/data/50863/000095013408008205/0000950134-08-008205-index.htm
#> 7    https://www.sec.gov/Archives/edgar/data/50863/000119312519142117/0001193125-19-142117-index.htm
#> 8   https://www.sec.gov/Archives/edgar/data/200245/000095012303004687/0000950123-03-004687-index.htm
#> 9   https://www.sec.gov/Archives/edgar/data/895421/000095010303001288/0000950103-03-001288-index.htm
#> 10   https://www.sec.gov/Archives/edgar/data/50863/000119312504054199/0001193125-04-054199-index.htm
#> 11  https://www.sec.gov/Archives/edgar/data/768251/000076825113000023/0000768251-13-000023-index.htm
#> 12 https://www.sec.gov/Archives/edgar/data/1022652/000109581100004115/0001095811-00-004115-index.htm
#> 13  https://www.sec.gov/Archives/edgar/data/895421/000095010303001359/0000950103-03-001359-index.htm
#> 14   https://www.sec.gov/Archives/edgar/data/50863/000119312519142117/0001193125-19-142117-index.htm
#> 15 https://www.sec.gov/Archives/edgar/data/1316016/000095012316020279/0000950123-16-020279-index.htm
#> 16  https://www.sec.gov/Archives/edgar/data/932290/000091205700036562/0000912057-00-036562-index.htm
#> 17 https://www.sec.gov/Archives/edgar/data/1120105/000119312507253790/0001193125-07-253790-index.htm
#> 18 https://www.sec.gov/Archives/edgar/data/1442505/000095013409006072/0000950134-09-006072-index.htm
#> 19   https://www.sec.gov/Archives/edgar/data/50863/000115752305003483/0001157523-05-003483-index.htm
#> 20 https://www.sec.gov/Archives/edgar/data/1022652/000095015000000898/0000950150-00-000898-index.htm
#> 21 https://www.sec.gov/Archives/edgar/data/1442505/000095012308009987/0000950123-08-009987-index.htm
#> 22   https://www.sec.gov/Archives/edgar/data/50863/000119312515209231/0001193125-15-209231-index.htm
#> 23 https://www.sec.gov/Archives/edgar/data/1316016/000095012317004142/0000950123-17-004142-index.htm
#> 24   https://www.sec.gov/Archives/edgar/data/50863/000119312512075534/0001193125-12-075534-index.htm
#> 25 https://www.sec.gov/Archives/edgar/data/1420302/000095012309014140/0000950123-09-014140-index.htm
#> 26   https://www.sec.gov/Archives/edgar/data/50863/000115752311005759/0001157523-11-005759-index.htm
#> 27 https://www.sec.gov/Archives/edgar/data/1318281/000119312507064810/0001193125-07-064810-index.htm
#> 28 https://www.sec.gov/Archives/edgar/data/1316016/000095012317005440/0000950123-17-005440-index.htm
#> 29   https://www.sec.gov/Archives/edgar/data/50863/000005086310000050/0000050863-10-000050-index.htm
#> 30  https://www.sec.gov/Archives/edgar/data/351145/000035114504000009/0000351145-04-000009-index.htm
#> 31   https://www.sec.gov/Archives/edgar/data/50863/000005086311000020/0000050863-11-000020-index.htm
#> 32   https://www.sec.gov/Archives/edgar/data/50863/000115752311004100/0001157523-11-004100-index.htm
#> 33  https://www.sec.gov/Archives/edgar/data/895421/000095010302000756/0000950103-02-000756-index.htm
#> 34   https://www.sec.gov/Archives/edgar/data/50863/000115752304000218/0001157523-04-000218-index.htm
#> 35   https://www.sec.gov/Archives/edgar/data/50863/000005086311000020/0000050863-11-000020-index.htm
#> 36   https://www.sec.gov/Archives/edgar/data/50863/000119312515382423/0001193125-15-382423-index.htm
#> 37   https://www.sec.gov/Archives/edgar/data/50863/000089161801500166/0000891618-01-500166-index.htm
#> 38  https://www.sec.gov/Archives/edgar/data/723125/000072312507000028/0000723125-07-000028-index.htm
#> 39 https://www.sec.gov/Archives/edgar/data/1171012/000118518515000158/0001185185-15-000158-index.htm
#> 40    https://www.sec.gov/Archives/edgar/data/2488/000119312509236705/0001193125-09-236705-index.htm
#> 41   https://www.sec.gov/Archives/edgar/data/50863/000095012309044666/0000950123-09-044666-index.htm
#> 42   https://www.sec.gov/Archives/edgar/data/50863/000005086311000013/0000050863-11-000013-index.htm
#> 43  https://www.sec.gov/Archives/edgar/data/895421/000095010302000842/0000950103-02-000842-index.htm
#> 44   https://www.sec.gov/Archives/edgar/data/50863/000095013407016808/0000950134-07-016808-index.htm
#> 45   https://www.sec.gov/Archives/edgar/data/50863/000091205702009698/0000912057-02-009698-index.htm
#> 46 https://www.sec.gov/Archives/edgar/data/1001193/000095013408004919/0000950134-08-004919-index.htm
#> 47 https://www.sec.gov/Archives/edgar/data/1280363/000119312511065181/0001193125-11-065181-index.htm
#> 48   https://www.sec.gov/Archives/edgar/data/50863/000119312519096036/0001193125-19-096036-index.htm
#> 49   https://www.sec.gov/Archives/edgar/data/50863/000005086308000089/0000050863-08-000089-index.htm
#> 50  https://www.sec.gov/Archives/edgar/data/917273/000101287001503032/0001012870-01-503032-index.htm
#> 51   https://www.sec.gov/Archives/edgar/data/50863/000119312520272580/0001193125-20-272580-index.htm
#> 52  https://www.sec.gov/Archives/edgar/data/351145/000035114502000021/0000351145-02-000021-index.htm
#> 53   https://www.sec.gov/Archives/edgar/data/50863/000119312504027430/0001193125-04-027430-index.htm
#> 54   https://www.sec.gov/Archives/edgar/data/50863/000119312519023301/0001193125-19-023301-index.htm
#> 55 https://www.sec.gov/Archives/edgar/data/1316016/000095012317004138/0000950123-17-004138-index.htm
#> 56   https://www.sec.gov/Archives/edgar/data/50863/000089161801000052/0000891618-01-000052-index.htm
#> 57 https://www.sec.gov/Archives/edgar/data/1285551/000089102007000003/0000891020-07-000003-index.htm
#> 58 https://www.sec.gov/Archives/edgar/data/1318281/000119312507034882/0001193125-07-034882-index.htm
#> 59   https://www.sec.gov/Archives/edgar/data/50863/000005086307000402/0000050863-07-000402-index.htm
#> 60  https://www.sec.gov/Archives/edgar/data/723125/000072312512000163/0000723125-12-000163-index.htm
#> 61   https://www.sec.gov/Archives/edgar/data/65100/000119312504025255/0001193125-04-025255-index.htm
#> 62   https://www.sec.gov/Archives/edgar/data/50863/000005086309000213/0000050863-09-000213-index.htm
#> 63   https://www.sec.gov/Archives/edgar/data/50863/000119312517209888/0001193125-17-209888-index.htm
#> 64 https://www.sec.gov/Archives/edgar/data/1420302/000095012309017935/0000950123-09-017935-index.htm
#> 65   https://www.sec.gov/Archives/edgar/data/50863/000119312521009142/0001193125-21-009142-index.htm
#> 66   https://www.sec.gov/Archives/edgar/data/50863/000005086316000105/0000050863-16-000105-index.htm
#> 67 https://www.sec.gov/Archives/edgar/data/1120105/000119312507262676/0001193125-07-262676-index.htm
#> 68 https://www.sec.gov/Archives/edgar/data/1283073/000093041304003280/0000930413-04-003280-index.htm
#> 69   https://www.sec.gov/Archives/edgar/data/50863/000005086315000015/0000050863-15-000015-index.htm
#> 70   https://www.sec.gov/Archives/edgar/data/50863/000005086315000055/0000050863-15-000055-index.htm
#> 71 https://www.sec.gov/Archives/edgar/data/1535379/000162828017003221/0001628280-17-003221-index.htm
#> 72 https://www.sec.gov/Archives/edgar/data/1001193/000089161808000487/0000891618-08-000487-index.htm
#> 73 https://www.sec.gov/Archives/edgar/data/1285551/000089102007000003/0000891020-07-000003-index.htm
#> 74   https://www.sec.gov/Archives/edgar/data/50863/000119312519023301/0001193125-19-023301-index.htm
#> 75   https://www.sec.gov/Archives/edgar/data/50863/000089161806000089/0000891618-06-000089-index.htm
#> 76    https://www.sec.gov/Archives/edgar/data/2488/000101287001501532/0001012870-01-501532-index.htm
#> 77   https://www.sec.gov/Archives/edgar/data/50863/000115752304009185/0001157523-04-009185-index.htm
#> 78   https://www.sec.gov/Archives/edgar/data/50863/000119312515382423/0001193125-15-382423-index.htm
#> 79   https://www.sec.gov/Archives/edgar/data/50863/000005086315000009/0000050863-15-000009-index.htm
#> 80 https://www.sec.gov/Archives/edgar/data/1316016/000119312517306042/0001193125-17-306042-index.htm
#> 81 https://www.sec.gov/Archives/edgar/data/1318281/000119312507034886/0001193125-07-034886-index.htm
#> 82   https://www.sec.gov/Archives/edgar/data/50863/000005086310000050/0000050863-10-000050-index.htm
#> 83   https://www.sec.gov/Archives/edgar/data/50863/000119312520220333/0001193125-20-220333-index.htm
#> 84   https://www.sec.gov/Archives/edgar/data/50863/000005086309000116/0000050863-09-000116-index.htm
#> 85   https://www.sec.gov/Archives/edgar/data/50863/000005086316000105/0000050863-16-000105-index.htm
#> 86   https://www.sec.gov/Archives/edgar/data/50863/000119312512297577/0001193125-12-297577-index.htm
#> 87   https://www.sec.gov/Archives/edgar/data/50863/000119312505033162/0001193125-05-033162-index.htm
#> 88   https://www.sec.gov/Archives/edgar/data/50863/000005086315000044/0000050863-15-000044-index.htm
#> 89   https://www.sec.gov/Archives/edgar/data/50863/000119312521009142/0001193125-21-009142-index.htm
#> 90   https://www.sec.gov/Archives/edgar/data/50863/000005086318000027/0000050863-18-000027-index.htm
#> 91  https://www.sec.gov/Archives/edgar/data/723125/000072312515000112/0000723125-15-000112-index.htm
#> 92   https://www.sec.gov/Archives/edgar/data/50863/000005086303000053/0000050863-03-000053-index.htm
#> 93 https://www.sec.gov/Archives/edgar/data/1120105/000119312507260104/0001193125-07-260104-index.htm
#> 94 https://www.sec.gov/Archives/edgar/data/1120105/000119312507170681/0001193125-07-170681-index.htm
#> 95 https://www.sec.gov/Archives/edgar/data/1388447/000089102007000191/0000891020-07-000191-index.htm
#> 96   https://www.sec.gov/Archives/edgar/data/50863/000119312517251345/0001193125-17-251345-index.htm
#> 97 https://www.sec.gov/Archives/edgar/data/1535379/000162828017003221/0001628280-17-003221-index.htm
#> 98 https://www.sec.gov/Archives/edgar/data/1442505/000095012310073483/0000950123-10-073483-index.htm
#> 99   https://www.sec.gov/Archives/edgar/data/50863/000005086311000094/0000050863-11-000094-index.htm
#>                                                                                           index_href
#> 1    https://www.sec.gov/Archives/edgar/data/50863/000095013407016808/0000950134-07-016808-index.htm
#> 2    https://www.sec.gov/Archives/edgar/data/50863/000095013408008205/0000950134-08-008205-index.htm
#> 3   https://www.sec.gov/Archives/edgar/data/932787/000095012307010792/0000950123-07-010792-index.htm
#> 4    https://www.sec.gov/Archives/edgar/data/50863/000119312521091374/0001193125-21-091374-index.htm
#> 5   https://www.sec.gov/Archives/edgar/data/200245/000095012301002657/0000950123-01-002657-index.htm
#> 6    https://www.sec.gov/Archives/edgar/data/50863/000095013408008205/0000950134-08-008205-index.htm
#> 7    https://www.sec.gov/Archives/edgar/data/50863/000119312519142117/0001193125-19-142117-index.htm
#> 8   https://www.sec.gov/Archives/edgar/data/200245/000095012303004687/0000950123-03-004687-index.htm
#> 9   https://www.sec.gov/Archives/edgar/data/895421/000095010303001288/0000950103-03-001288-index.htm
#> 10   https://www.sec.gov/Archives/edgar/data/50863/000119312504054199/0001193125-04-054199-index.htm
#> 11  https://www.sec.gov/Archives/edgar/data/768251/000076825113000023/0000768251-13-000023-index.htm
#> 12 https://www.sec.gov/Archives/edgar/data/1022652/000109581100004115/0001095811-00-004115-index.htm
#> 13  https://www.sec.gov/Archives/edgar/data/895421/000095010303001359/0000950103-03-001359-index.htm
#> 14   https://www.sec.gov/Archives/edgar/data/50863/000119312519142117/0001193125-19-142117-index.htm
#> 15 https://www.sec.gov/Archives/edgar/data/1316016/000095012316020279/0000950123-16-020279-index.htm
#> 16  https://www.sec.gov/Archives/edgar/data/932290/000091205700036562/0000912057-00-036562-index.htm
#> 17 https://www.sec.gov/Archives/edgar/data/1120105/000119312507253790/0001193125-07-253790-index.htm
#> 18 https://www.sec.gov/Archives/edgar/data/1442505/000095013409006072/0000950134-09-006072-index.htm
#> 19   https://www.sec.gov/Archives/edgar/data/50863/000115752305003483/0001157523-05-003483-index.htm
#> 20 https://www.sec.gov/Archives/edgar/data/1022652/000095015000000898/0000950150-00-000898-index.htm
#> 21 https://www.sec.gov/Archives/edgar/data/1442505/000095012308009987/0000950123-08-009987-index.htm
#> 22   https://www.sec.gov/Archives/edgar/data/50863/000119312515209231/0001193125-15-209231-index.htm
#> 23 https://www.sec.gov/Archives/edgar/data/1316016/000095012317004142/0000950123-17-004142-index.htm
#> 24   https://www.sec.gov/Archives/edgar/data/50863/000119312512075534/0001193125-12-075534-index.htm
#> 25 https://www.sec.gov/Archives/edgar/data/1420302/000095012309014140/0000950123-09-014140-index.htm
#> 26   https://www.sec.gov/Archives/edgar/data/50863/000115752311005759/0001157523-11-005759-index.htm
#> 27 https://www.sec.gov/Archives/edgar/data/1318281/000119312507064810/0001193125-07-064810-index.htm
#> 28 https://www.sec.gov/Archives/edgar/data/1316016/000095012317005440/0000950123-17-005440-index.htm
#> 29   https://www.sec.gov/Archives/edgar/data/50863/000005086310000050/0000050863-10-000050-index.htm
#> 30  https://www.sec.gov/Archives/edgar/data/351145/000035114504000009/0000351145-04-000009-index.htm
#> 31   https://www.sec.gov/Archives/edgar/data/50863/000005086311000020/0000050863-11-000020-index.htm
#> 32   https://www.sec.gov/Archives/edgar/data/50863/000115752311004100/0001157523-11-004100-index.htm
#> 33  https://www.sec.gov/Archives/edgar/data/895421/000095010302000756/0000950103-02-000756-index.htm
#> 34   https://www.sec.gov/Archives/edgar/data/50863/000115752304000218/0001157523-04-000218-index.htm
#> 35   https://www.sec.gov/Archives/edgar/data/50863/000005086311000020/0000050863-11-000020-index.htm
#> 36   https://www.sec.gov/Archives/edgar/data/50863/000119312515382423/0001193125-15-382423-index.htm
#> 37   https://www.sec.gov/Archives/edgar/data/50863/000089161801500166/0000891618-01-500166-index.htm
#> 38  https://www.sec.gov/Archives/edgar/data/723125/000072312507000028/0000723125-07-000028-index.htm
#> 39 https://www.sec.gov/Archives/edgar/data/1171012/000118518515000158/0001185185-15-000158-index.htm
#> 40    https://www.sec.gov/Archives/edgar/data/2488/000119312509236705/0001193125-09-236705-index.htm
#> 41   https://www.sec.gov/Archives/edgar/data/50863/000095012309044666/0000950123-09-044666-index.htm
#> 42   https://www.sec.gov/Archives/edgar/data/50863/000005086311000013/0000050863-11-000013-index.htm
#> 43  https://www.sec.gov/Archives/edgar/data/895421/000095010302000842/0000950103-02-000842-index.htm
#> 44   https://www.sec.gov/Archives/edgar/data/50863/000095013407016808/0000950134-07-016808-index.htm
#> 45   https://www.sec.gov/Archives/edgar/data/50863/000091205702009698/0000912057-02-009698-index.htm
#> 46 https://www.sec.gov/Archives/edgar/data/1001193/000095013408004919/0000950134-08-004919-index.htm
#> 47 https://www.sec.gov/Archives/edgar/data/1280363/000119312511065181/0001193125-11-065181-index.htm
#> 48   https://www.sec.gov/Archives/edgar/data/50863/000119312519096036/0001193125-19-096036-index.htm
#> 49   https://www.sec.gov/Archives/edgar/data/50863/000005086308000089/0000050863-08-000089-index.htm
#> 50  https://www.sec.gov/Archives/edgar/data/917273/000101287001503032/0001012870-01-503032-index.htm
#> 51   https://www.sec.gov/Archives/edgar/data/50863/000119312520272580/0001193125-20-272580-index.htm
#> 52  https://www.sec.gov/Archives/edgar/data/351145/000035114502000021/0000351145-02-000021-index.htm
#> 53   https://www.sec.gov/Archives/edgar/data/50863/000119312504027430/0001193125-04-027430-index.htm
#> 54   https://www.sec.gov/Archives/edgar/data/50863/000119312519023301/0001193125-19-023301-index.htm
#> 55 https://www.sec.gov/Archives/edgar/data/1316016/000095012317004138/0000950123-17-004138-index.htm
#> 56   https://www.sec.gov/Archives/edgar/data/50863/000089161801000052/0000891618-01-000052-index.htm
#> 57 https://www.sec.gov/Archives/edgar/data/1285551/000089102007000003/0000891020-07-000003-index.htm
#> 58 https://www.sec.gov/Archives/edgar/data/1318281/000119312507034882/0001193125-07-034882-index.htm
#> 59   https://www.sec.gov/Archives/edgar/data/50863/000005086307000402/0000050863-07-000402-index.htm
#> 60  https://www.sec.gov/Archives/edgar/data/723125/000072312512000163/0000723125-12-000163-index.htm
#> 61   https://www.sec.gov/Archives/edgar/data/65100/000119312504025255/0001193125-04-025255-index.htm
#> 62   https://www.sec.gov/Archives/edgar/data/50863/000005086309000213/0000050863-09-000213-index.htm
#> 63   https://www.sec.gov/Archives/edgar/data/50863/000119312517209888/0001193125-17-209888-index.htm
#> 64 https://www.sec.gov/Archives/edgar/data/1420302/000095012309017935/0000950123-09-017935-index.htm
#> 65   https://www.sec.gov/Archives/edgar/data/50863/000119312521009142/0001193125-21-009142-index.htm
#> 66   https://www.sec.gov/Archives/edgar/data/50863/000005086316000105/0000050863-16-000105-index.htm
#> 67 https://www.sec.gov/Archives/edgar/data/1120105/000119312507262676/0001193125-07-262676-index.htm
#> 68 https://www.sec.gov/Archives/edgar/data/1283073/000093041304003280/0000930413-04-003280-index.htm
#> 69   https://www.sec.gov/Archives/edgar/data/50863/000005086315000015/0000050863-15-000015-index.htm
#> 70   https://www.sec.gov/Archives/edgar/data/50863/000005086315000055/0000050863-15-000055-index.htm
#> 71 https://www.sec.gov/Archives/edgar/data/1535379/000162828017003221/0001628280-17-003221-index.htm
#> 72 https://www.sec.gov/Archives/edgar/data/1001193/000089161808000487/0000891618-08-000487-index.htm
#> 73 https://www.sec.gov/Archives/edgar/data/1285551/000089102007000003/0000891020-07-000003-index.htm
#> 74   https://www.sec.gov/Archives/edgar/data/50863/000119312519023301/0001193125-19-023301-index.htm
#> 75   https://www.sec.gov/Archives/edgar/data/50863/000089161806000089/0000891618-06-000089-index.htm
#> 76    https://www.sec.gov/Archives/edgar/data/2488/000101287001501532/0001012870-01-501532-index.htm
#> 77   https://www.sec.gov/Archives/edgar/data/50863/000115752304009185/0001157523-04-009185-index.htm
#> 78   https://www.sec.gov/Archives/edgar/data/50863/000119312515382423/0001193125-15-382423-index.htm
#> 79   https://www.sec.gov/Archives/edgar/data/50863/000005086315000009/0000050863-15-000009-index.htm
#> 80 https://www.sec.gov/Archives/edgar/data/1316016/000119312517306042/0001193125-17-306042-index.htm
#> 81 https://www.sec.gov/Archives/edgar/data/1318281/000119312507034886/0001193125-07-034886-index.htm
#> 82   https://www.sec.gov/Archives/edgar/data/50863/000005086310000050/0000050863-10-000050-index.htm
#> 83   https://www.sec.gov/Archives/edgar/data/50863/000119312520220333/0001193125-20-220333-index.htm
#> 84   https://www.sec.gov/Archives/edgar/data/50863/000005086309000116/0000050863-09-000116-index.htm
#> 85   https://www.sec.gov/Archives/edgar/data/50863/000005086316000105/0000050863-16-000105-index.htm
#> 86   https://www.sec.gov/Archives/edgar/data/50863/000119312512297577/0001193125-12-297577-index.htm
#> 87   https://www.sec.gov/Archives/edgar/data/50863/000119312505033162/0001193125-05-033162-index.htm
#> 88   https://www.sec.gov/Archives/edgar/data/50863/000005086315000044/0000050863-15-000044-index.htm
#> 89   https://www.sec.gov/Archives/edgar/data/50863/000119312521009142/0001193125-21-009142-index.htm
#> 90   https://www.sec.gov/Archives/edgar/data/50863/000005086318000027/0000050863-18-000027-index.htm
#> 91  https://www.sec.gov/Archives/edgar/data/723125/000072312515000112/0000723125-15-000112-index.htm
#> 92   https://www.sec.gov/Archives/edgar/data/50863/000005086303000053/0000050863-03-000053-index.htm
#> 93 https://www.sec.gov/Archives/edgar/data/1120105/000119312507260104/0001193125-07-260104-index.htm
#> 94 https://www.sec.gov/Archives/edgar/data/1120105/000119312507170681/0001193125-07-170681-index.htm
#> 95 https://www.sec.gov/Archives/edgar/data/1388447/000089102007000191/0000891020-07-000191-index.htm
#> 96   https://www.sec.gov/Archives/edgar/data/50863/000119312517251345/0001193125-17-251345-index.htm
#> 97 https://www.sec.gov/Archives/edgar/data/1535379/000162828017003221/0001628280-17-003221-index.htm
#> 98 https://www.sec.gov/Archives/edgar/data/1442505/000095012310073483/0000950123-10-073483-index.htm
#> 99   https://www.sec.gov/Archives/edgar/data/50863/000005086311000094/0000050863-11-000094-index.htm

# }