SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(q = "*", type = "", reverse_order = FALSE, count = 10,
  page = 1, stemming = TRUE, name = "", cik = "", sic = "",
  from = "", to = "")

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

## This can be very slow running
full_text('intel')

#>    filing_date                           name
#> 1   2019-02-01            10-K for INTEL CORP
#> 2   2019-02-01 EX-10.2 of 10-K for INTEL CORP
#> 3   2019-02-01 EX-21.1 of 10-K for INTEL CORP
#> 4   2019-02-01 EX-23.1 of 10-K for INTEL CORP
#> 5   2019-02-01 EX-31.1 of 10-K for INTEL CORP
#> 6   2019-02-01 EX-31.2 of 10-K for INTEL CORP
#> 7   2019-02-01 EX-32.1 of 10-K for INTEL CORP
#> 8   2019-02-01 EX-99.1 of 10-K for INTEL CORP
#> 9   2019-02-01     XML of 10-K for INTEL CORP
#> 10  2019-02-01     XML of 10-K for INTEL CORP
#>                                                                                            href
#> 1  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-10kdocument.htm
#> 2        https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-ex102.htm
#> 3        https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-ex211.htm
#> 4        https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-ex231.htm
#> 5        https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-ex311.htm
#> 6        https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-ex312.htm
#> 7        https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-ex321.htm
#> 8        https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-ex991.htm
#> 9            https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/FilingSummary.xml
#> 10                      https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/R1.htm
#>    company_name   cik  sic
#> 1    INTEL CORP 50863 3674
#> 2    INTEL CORP 50863 3674
#> 3    INTEL CORP 50863 3674
#> 4    INTEL CORP 50863 3674
#> 5    INTEL CORP 50863 3674
#> 6    INTEL CORP 50863 3674
#> 7    INTEL CORP 50863 3674
#> 8    INTEL CORP 50863 3674
#> 9    INTEL CORP 50863 3674
#> 10   INTEL CORP 50863 3674
#>                                                                                                                                                                                                                                                                                                     content
#> 1                                                                           The reports and other publications referenced are generally available to the public and were not commissioned by Intel. INTEL UNIQUE TERMS We use specific terms throughout this document to describe our business and results.
#> 2                                                                                                                                                                                                                               Exhibit 10.2 INTEL CORPORATION 2006 EMPLOYEE STOCK PURCHASE PLAN Section 1.
#> 3                                                                        Pursuant to Item 601(b)(21)(ii) of Regulation S-K, the names of other Intel Corporation subsidiaries are omitted because, considered in the aggregate, they would not constitute a significant subsidiary as of December 29, 2018.
#> 4       Exhibit 23.1 CONSENT OF INDEPENDENT REGISTERED PUBLIC ACCOUNTING FIRM We consent to the incorporation by reference in the following Registration Statements: (1) Registration Statements (Form S-3 Nos. 333-207633 and 333-224472) of Intel Corporation, (2) Registration Statements (Form S-4 Nos.
#> 5                                                                                                                                                                                                                                  I have reviewed this annual report on Form 10-K of Intel Corporation; 2.
#> 6                                                                                                                                                                                                                                  I have reviewed this annual report on Form 10-K of Intel Corporation; 2.
#> 7      This written statement is being furnished to the Securities and Exchange Commission as an exhibit to such Form 10-K. A signed original of this statement has been provided to Intel and will be retained by Intel and furnished to the Securities and Exchange Commission or its staff upon request.
#> 8                                                                                                                                                                                                                       Boards and Systems - Server boards and small form factor systems such as Intel NUC.
#> 9  3.10.0.1 html 726 554 1 false 196 0 false 8 false false R1.htm 0001000 - Document - Document and Entity Information Sheet http:/www.intel.com/role/DocumentAndEntityInformation Document and Entity Information Cover 1 false false R2.htm 1001000 - Statement - Consolidated Statements of Income Sheet
#> 10                                                             v3.10.0.1 Document and Entity Information - USD ($) shares in Millions, $ in Billions 12 Months Ended Dec. 29, 2018 Jan. 26, 2019 Jun. 29, 2018 Intel Corporation 0000050863 -12-29 Large Accelerated Filer 10-K Dec. 29, 2018 2018 FY false
#>                                                                                     parent_href
#> 1                                                                                          <NA>
#> 2  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-10kdocument.htm
#> 3  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-10kdocument.htm
#> 4  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-10kdocument.htm
#> 5  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-10kdocument.htm
#> 6  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-10kdocument.htm
#> 7  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-10kdocument.htm
#> 8  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-10kdocument.htm
#> 9  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-10kdocument.htm
#> 10 https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/a12292018q4-10kdocument.htm
#>                                                                                         index_href
#> 1                                                                                             <NA>
#> 2  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/0000050863-19-000007-index.htm
#> 3  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/0000050863-19-000007-index.htm
#> 4  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/0000050863-19-000007-index.htm
#> 5  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/0000050863-19-000007-index.htm
#> 6  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/0000050863-19-000007-index.htm
#> 7  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/0000050863-19-000007-index.htm
#> 8  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/0000050863-19-000007-index.htm
#> 9  https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/0000050863-19-000007-index.htm
#> 10 https://www.sec.gov/Archives/edgar/data/50863/000005086319000007/0000050863-19-000007-index.htm