SEC Full-Text Search

Provides access to the SEC fillings full-text search tool.

full_text(
  q = "*",
  type = "",
  reverse_order = FALSE,
  count = 10,
  page = 1,
  stemming = TRUE,
  name = "",
  cik = "",
  sic = "",
  from = "",
  to = ""
)

Arguments

q
Search query. For details on special formatting, see the FAQ.

type
Type of forms to search - e.g. '10-K'

reverse_order
If true, order by oldest first instead of newest first

count
Number of results to return

page
Which page of results to return

stemming
Search by base words(default) or exactly as entered

name
Company name. Cannot be combined with `cik` or `sik`.

cik
Company code to search. Cannot be combined with `name` or `sic`

sic
Standard Industrial Classification of filer to search for. Cannot be combined with `cik` or `name`. Special options - 1: all, 0: Unspecified.

from
Start date. Must be in the form of `mm/dd/yyyy`. Must also specify `to`

to
End date. Must be in the form of `mm/dd/yyyy`. Must also specify `from`

Value

A dataframe list of results including the following columns -

filing_date

name

href

company_name

cik

sic

content

parent_href

index_href

Examples

# \donttest{
## This can be very slow running
full_text('intel')

#>    filing_date                                                             name
#> 1   2020-04-17                 INFORMATION TABLE of 13F-HR for CENTRAL TRUST Co
#> 2   2020-04-17                                        DEF 14A for USA TRUCK INC
#> 3   2020-04-17                    485BPOS for WILSHIRE VARIABLE INSURANCE TRUST
#> 4   2020-04-17     INFORMATION TABLE of 13F-HR for NISA INVESTMENT ADVISORS LLC
#> 5   2020-04-17 INFORMATION TABLE of 13F-HR for Bremer Bank National Association
#> 6   2020-04-17          INFORMATION TABLE of 13F-HR for DnB Asset Management AS
#> 7   2020-04-17        INFORMATION TABLE of 13F-HR for JCIC Asset Management Inc
#> 8   2020-04-17  INFORMATION TABLE of 13F-HR for Endurance Wealth Management Inc
#> 9   2020-04-17                          DEF 14A for ROYAL CARIBBEAN CRUISES LTD
#> 10  2020-04-17  INFORMATION TABLE of 13F-HR for MARCO INVESTMENT MANAGEMENT LLC
#>                                                                                                      href
#> 1          https://www.sec.gov/Archives/edgar/data/44365/000004436520000002/xslForm13F_X01/MAR202013F.xml
#> 2              https://www.sec.gov/Archives/edgar/data/883945/000088394520000016/usak-20200417xdef14a.htm
#> 3           https://www.sec.gov/Archives/edgar/data/1026708/000089418920002786/wilshirevitcombineddoc.htm
#> 4        https://www.sec.gov/Archives/edgar/data/937615/000093761520000005/xslForm13F_X01/13F03312020.xml
#> 5  https://www.sec.gov/Archives/edgar/data/1775321/000095012320003400/xslForm13F_X01/form13fInfoTable.xml
#> 6  https://www.sec.gov/Archives/edgar/data/1456670/000095012320003408/xslForm13F_X01/form13fInfoTable.xml
#> 7  https://www.sec.gov/Archives/edgar/data/1492040/000095012320003437/xslForm13F_X01/form13fInfoTable.xml
#> 8  https://www.sec.gov/Archives/edgar/data/1476225/000095012320003447/xslForm13F_X01/form13fInfoTable.xml
#> 9                   https://www.sec.gov/Archives/edgar/data/884887/000104746920002406/a2241308zdef14a.htm
#> 10 https://www.sec.gov/Archives/edgar/data/1080493/000106299320001836/xslForm13F_X01/form13fInfoTable.xml
#>                         company_name     cik  sic
#> 1                   CENTRAL TRUST Co   44365 <NA>
#> 2                      USA TRUCK INC  883945 4213
#> 3  WILSHIRE VARIABLE INSURANCE TRUST 1026708 <NA>
#> 4       NISA INVESTMENT ADVISORS LLC  937615 <NA>
#> 5   Bremer Bank National Association 1775321 <NA>
#> 6            DnB Asset Management AS 1456670 <NA>
#> 7          JCIC Asset Management Inc 1492040 <NA>
#> 8    Endurance Wealth Management Inc 1476225 <NA>
#> 9        ROYAL CARIBBEAN CRUISES LTD  884887 4400
#> 10   MARCO INVESTMENT MANAGEMENT LLC 1080493 <NA>
#>                                                                                                                                                                                                                                                                                                        content
#> 1                                                                                                                                                                                                                                                                                                             
#> 2                                               UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 SCHEDULE 14A Proxy Statement Pursuant to Section 14(a) of the Securities Exchange Act of 1934 Filed by the Registrant Filed by a Party other than the Registrant Check the appropriate
#> 3      As filed with the Securities and Exchange Commission on April 17, 2020 1933 Act Registration File No. 333-15881 1940 Act File No. 811-07917 SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 FORM N-1A REGISTRATION STATEMENT UNDER THE SECURITIES ACT OF 1933 [X] Pre-Effective Amendment No.
#> 4  AAON INC COM 000360206 14 300 SH SOLE 300 0 0 AAR CORP COM 000361105 6 345 SH SOLE 345 0 0 ACCO BRANDS CORP COM 00081T108 8 1570 SH SOLE 1570 0 0 ABM INDS INC COM 000957100 59 2430 SH SOLE 2430 0 0 AFLAC INC COM 001055102 10883 317849 SH OTR 01 317789 0 60 AGCO CORP COM 001084102 1465 31010 SH SOLE
#> 5                                                                                                                                                                                                                                                                                    3M Co COM 88579Y101 2983 
#> 6                                                                                                                                                                                                                                                            Elanco Animal Health Inc common 28414H103 1153779
#> 7                                                                                                                                                                                                                                                                   ISHARES TR CORE US AGGBD ET 464287226 136 
#> 8                                                                                                                                                                                                                                                                               3M CO COM Stock 88579Y101 501 
#> 9        UNITED STATES SECURITIES AND EXCHANGE COMMISSION Washington, D.C. 20549 SCHEDULE 14A Proxy Statement Pursuant to Section 14(a) of the Securities Exchange Act of 1934 Filed by the Registrant Filed by a Party other than the Registrant o Check the appropriate box: o Preliminary Proxy Statement o
#> 10     AFLAC INC COM 001055102 3456 100930 SH SOLE 0 0 0 AGCO CORP COM 001084102 898 19000 SH SOLE 0 0 0 AT&T INC COM 00206R102 1714 58809 SH SOLE 0 0 0 ABBOTT LABS COM 002824100 5988 75886 SH SOLE 0 0 0 ABBVIE INC COM 00287Y109 6280 82427 SH SOLE 0 0 0 ACCENTURE PLC IRELAND SHS CLASS A G1151C101 2094
#>                                                                                          parent_href
#> 1    https://www.sec.gov/Archives/edgar/data/44365/000004436520000002/xslForm13F_X01/primary_doc.xml
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4  https://www.sec.gov/Archives/edgar/data/1456670/000095012320003408/xslForm13F_X01/primary_doc.xml
#> 5  https://www.sec.gov/Archives/edgar/data/1492040/000095012320003437/xslForm13F_X01/primary_doc.xml
#> 6  https://www.sec.gov/Archives/edgar/data/1476225/000095012320003447/xslForm13F_X01/primary_doc.xml
#> 7  https://www.sec.gov/Archives/edgar/data/1080493/000106299320001836/xslForm13F_X01/primary_doc.xml
#> 8    https://www.sec.gov/Archives/edgar/data/44365/000004436520000002/xslForm13F_X01/primary_doc.xml
#> 9                                                                                               <NA>
#> 10 https://www.sec.gov/Archives/edgar/data/1775321/000095012320003400/xslForm13F_X01/primary_doc.xml
#>                                                                                           index_href
#> 1    https://www.sec.gov/Archives/edgar/data/44365/000004436520000002/0000044365-20-000002-index.htm
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4  https://www.sec.gov/Archives/edgar/data/1456670/000095012320003408/0000950123-20-003408-index.htm
#> 5  https://www.sec.gov/Archives/edgar/data/1492040/000095012320003437/0000950123-20-003437-index.htm
#> 6  https://www.sec.gov/Archives/edgar/data/1476225/000095012320003447/0000950123-20-003447-index.htm
#> 7  https://www.sec.gov/Archives/edgar/data/1080493/000106299320001836/0001062993-20-001836-index.htm
#> 8    https://www.sec.gov/Archives/edgar/data/44365/000004436520000002/0000044365-20-000002-index.htm
#> 9                                                                                               <NA>
#> 10 https://www.sec.gov/Archives/edgar/data/1775321/000095012320003400/0000950123-20-003400-index.htm

# }