SEC Header Search

Searches filing headers going back to 1994 excluding the most recent day using the interface here

header_search(q, page = 1, from = 1994, to = 2017)

Arguments

q
The search string. Documentation here

page
Which results page to return (default: 1)

from
Start year (default: 1994)

to
End year (default: Current year)

Value

A dataframe of funds found including the following columns -

company_name

filing_href

form

filing_date

size

Examples

# \donttest{
## This can be very slow running
header_search("company-name = Apple")

#>                           company_name
#> 1                 ADS APPLE VALLEY INC
#> 2                 ADS APPLE VALLEY INC
#> 3                 ADS APPLE VALLEY INC
#> 4                 ADS APPLE VALLEY INC
#> 5                 ADS APPLE VALLEY INC
#> 6     ADS APPLE VALLEY LTD PARTNERSHIP
#> 7     ADS APPLE VALLEY LTD PARTNERSHIP
#> 8     ADS APPLE VALLEY LTD PARTNERSHIP
#> 9     ADS APPLE VALLEY LTD PARTNERSHIP
#> 10    ADS APPLE VALLEY LTD PARTNERSHIP
#> 11                  Apple Allan Victor
#> 12                  Apple Allan Victor
#> 13                  Apple Allan Victor
#> 14                  Apple Allan Victor
#> 15                  Apple Allan Victor
#> 16                  Apple Allan Victor
#> 17 Apple Beach County Ltd Liability Co
#> 18             Apple Blossom REIT, LLC
#> 19         Apple Blossom Royalties LLC
#> 20       APPLE BOX PRODUCTIONS SUB INC
#> 21       APPLE BOX PRODUCTIONS SUB INC
#> 22       APPLE BOX PRODUCTIONS SUB INC
#> 23       APPLE BOX PRODUCTIONS SUB INC
#> 24       APPLE BOX PRODUCTIONS SUB INC
#> 25       APPLE BOX PRODUCTIONS SUB INC
#> 26       APPLE BOX PRODUCTIONS SUB INC
#> 27       APPLE BOX PRODUCTIONS SUB INC
#> 28                       APPLE C CHRIS
#> 29                       APPLE C CHRIS
#> 30                       APPLE C CHRIS
#> 31                       APPLE C CHRIS
#> 32                       APPLE C CHRIS
#> 33           Apple Capital Group, Inc.
#> 34                  APPLE COMPUTER INC
#> 35                  APPLE COMPUTER INC
#> 36                  APPLE COMPUTER INC
#> 37                  APPLE COMPUTER INC
#> 38                  APPLE COMPUTER INC
#> 39                  APPLE COMPUTER INC
#> 40                  APPLE COMPUTER INC
#> 41                  APPLE COMPUTER INC
#> 42                  APPLE COMPUTER INC
#> 43                  APPLE COMPUTER INC
#> 44                  APPLE COMPUTER INC
#> 45                  APPLE COMPUTER INC
#> 46                  APPLE COMPUTER INC
#> 47                  APPLE COMPUTER INC
#> 48                  APPLE COMPUTER INC
#> 49                  APPLE COMPUTER INC
#> 50                  APPLE COMPUTER INC
#> 51                  APPLE COMPUTER INC
#> 52                  APPLE COMPUTER INC
#> 53                  APPLE COMPUTER INC
#> 54                  APPLE COMPUTER INC
#> 55                  APPLE COMPUTER INC
#> 56                  APPLE COMPUTER INC
#> 57                  APPLE COMPUTER INC
#> 58                  APPLE COMPUTER INC
#> 59                  APPLE COMPUTER INC
#> 60                  APPLE COMPUTER INC
#> 61                  APPLE COMPUTER INC
#> 62                  APPLE COMPUTER INC
#> 63                  APPLE COMPUTER INC
#> 64                  APPLE COMPUTER INC
#> 65                  APPLE COMPUTER INC
#> 66                  APPLE COMPUTER INC
#> 67                  APPLE COMPUTER INC
#> 68                  APPLE COMPUTER INC
#> 69                  APPLE COMPUTER INC
#> 70                  APPLE COMPUTER INC
#> 71                  APPLE COMPUTER INC
#> 72                  APPLE COMPUTER INC
#> 73                  APPLE COMPUTER INC
#> 74                  APPLE COMPUTER INC
#> 75                  APPLE COMPUTER INC
#> 76                  APPLE COMPUTER INC
#> 77                  APPLE COMPUTER INC
#> 78                  APPLE COMPUTER INC
#> 79                  APPLE COMPUTER INC
#> 80                  APPLE COMPUTER INC
#> 81                  APPLE COMPUTER INC
#>                                                                                          filing_href
#> 1  https://www.sec.gov/Archives/edgar/data/1281227/000110465907053957/0001104659-07-053957-index.htm
#> 2  https://www.sec.gov/Archives/edgar/data/1281227/000110465907053944/0001104659-07-053944-index.htm
#> 3  https://www.sec.gov/Archives/edgar/data/1281227/000095011605002515/0000950116-05-002515-index.htm
#> 4  https://www.sec.gov/Archives/edgar/data/1281227/000095011605002291/0000950116-05-002291-index.htm
#> 5  https://www.sec.gov/Archives/edgar/data/1281227/000095011604001921/0000950116-04-001921-index.htm
#> 6  https://www.sec.gov/Archives/edgar/data/1281225/000110465907053944/0001104659-07-053944-index.htm
#> 7  https://www.sec.gov/Archives/edgar/data/1281225/000110465907053957/0001104659-07-053957-index.htm
#> 8  https://www.sec.gov/Archives/edgar/data/1281225/000095011605002515/0000950116-05-002515-index.htm
#> 9  https://www.sec.gov/Archives/edgar/data/1281225/000095011605002291/0000950116-05-002291-index.htm
#> 10 https://www.sec.gov/Archives/edgar/data/1281225/000095011604001921/0000950116-04-001921-index.htm
#> 11 https://www.sec.gov/Archives/edgar/data/1493086/000120919116138837/0001209191-16-138837-index.htm
#> 12 https://www.sec.gov/Archives/edgar/data/1493086/000120919116129022/0001209191-16-129022-index.htm
#> 13 https://www.sec.gov/Archives/edgar/data/1493086/000120919115085008/0001209191-15-085008-index.htm
#> 14 https://www.sec.gov/Archives/edgar/data/1493086/000120919110038126/0001209191-10-038126-index.htm
#> 15 https://www.sec.gov/Archives/edgar/data/1493086/000149308610000002/0001493086-10-000002-index.htm
#> 16 https://www.sec.gov/Archives/edgar/data/1493086/000149308610000001/0001493086-10-000001-index.htm
#> 17 https://www.sec.gov/Archives/edgar/data/1653690/000165369015000001/0001653690-15-000001-index.htm
#> 18 https://www.sec.gov/Archives/edgar/data/1567043/000114036113006629/0001140361-13-006629-index.htm
#> 19 https://www.sec.gov/Archives/edgar/data/1412692/999999999707041239/9999999997-07-041239-index.htm
#> 20 https://www.sec.gov/Archives/edgar/data/1235253/000119312505100880/0001193125-05-100880-index.htm
#> 21 https://www.sec.gov/Archives/edgar/data/1235253/000119312505096169/0001193125-05-096169-index.htm
#> 22 https://www.sec.gov/Archives/edgar/data/1235253/000119312505077834/0001193125-05-077834-index.htm
#> 23 https://www.sec.gov/Archives/edgar/data/1235253/000095012304002409/0000950123-04-002409-index.htm
#> 24 https://www.sec.gov/Archives/edgar/data/1235253/000095012303010784/0000950123-03-010784-index.htm
#> 25 https://www.sec.gov/Archives/edgar/data/1235253/000095012303008451/0000950123-03-008451-index.htm
#> 26 https://www.sec.gov/Archives/edgar/data/1235253/000095012303008257/0000950123-03-008257-index.htm
#> 27 https://www.sec.gov/Archives/edgar/data/1235253/000095012303007198/0000950123-03-007198-index.htm
#> 28 https://www.sec.gov/Archives/edgar/data/1166766/000118143105021839/0001181431-05-021839-index.htm
#> 29 https://www.sec.gov/Archives/edgar/data/1166766/000118143105019382/0001181431-05-019382-index.htm
#> 30 https://www.sec.gov/Archives/edgar/data/1166766/000118143104019688/0001181431-04-019688-index.htm
#> 31 https://www.sec.gov/Archives/edgar/data/1166766/000093066102003099/0000930661-02-003099-index.htm
#> 32 https://www.sec.gov/Archives/edgar/data/1166766/000093066102000355/0000930661-02-000355-index.htm
#> 33 https://www.sec.gov/Archives/edgar/data/1718804/000171880417000001/0001718804-17-000001-index.htm
#> 34  https://www.sec.gov/Archives/edgar/data/320193/000118143107001267/0001181431-07-001267-index.htm
#> 35  https://www.sec.gov/Archives/edgar/data/320193/000110465906084288/0001104659-06-084288-index.htm
#> 36  https://www.sec.gov/Archives/edgar/data/320193/000110465906084286/0001104659-06-084286-index.htm
#> 37  https://www.sec.gov/Archives/edgar/data/320193/000110465906081617/0001104659-06-081617-index.htm
#> 38  https://www.sec.gov/Archives/edgar/data/320193/000118143106067704/0001181431-06-067704-index.htm
#> 39  https://www.sec.gov/Archives/edgar/data/320193/000118143106067706/0001181431-06-067706-index.htm
#> 40  https://www.sec.gov/Archives/edgar/data/320193/000110465906069015/0001104659-06-069015-index.htm
#> 41  https://www.sec.gov/Archives/edgar/data/320193/000110465906067275/0001104659-06-067275-index.htm
#> 42  https://www.sec.gov/Archives/edgar/data/320193/000000000006048618/0000000000-06-048618-index.htm
#> 43  https://www.sec.gov/Archives/edgar/data/320193/000110465906064935/0001104659-06-064935-index.htm
#> 44  https://www.sec.gov/Archives/edgar/data/320193/000118143106053331/0001181431-06-053331-index.htm
#> 45  https://www.sec.gov/Archives/edgar/data/320193/000118143106051734/0001181431-06-051734-index.htm
#> 46  https://www.sec.gov/Archives/edgar/data/320193/000118143106051735/0001181431-06-051735-index.htm
#> 47  https://www.sec.gov/Archives/edgar/data/320193/000118143106050971/0001181431-06-050971-index.htm
#> 48  https://www.sec.gov/Archives/edgar/data/320193/000110465906058808/0001104659-06-058808-index.htm
#> 49  https://www.sec.gov/Archives/edgar/data/320193/000000000006041438/0000000000-06-041438-index.htm
#> 50  https://www.sec.gov/Archives/edgar/data/320193/000118143106048402/0001181431-06-048402-index.htm
#> 51  https://www.sec.gov/Archives/edgar/data/320193/000110465906055907/0001104659-06-055907-index.htm
#> 52  https://www.sec.gov/Archives/edgar/data/320193/000110465906053743/0001104659-06-053743-index.htm
#> 53  https://www.sec.gov/Archives/edgar/data/320193/000110465906054118/0001104659-06-054118-index.htm
#> 54  https://www.sec.gov/Archives/edgar/data/320193/000118143106046420/0001181431-06-046420-index.htm
#> 55  https://www.sec.gov/Archives/edgar/data/320193/000118143106046421/0001181431-06-046421-index.htm
#> 56  https://www.sec.gov/Archives/edgar/data/320193/000110465906051302/0001104659-06-051302-index.htm
#> 57  https://www.sec.gov/Archives/edgar/data/320193/000118143106044978/0001181431-06-044978-index.htm
#> 58  https://www.sec.gov/Archives/edgar/data/320193/000118143106044234/0001181431-06-044234-index.htm
#> 59  https://www.sec.gov/Archives/edgar/data/320193/000118143106043774/0001181431-06-043774-index.htm
#> 60  https://www.sec.gov/Archives/edgar/data/320193/000118143106043775/0001181431-06-043775-index.htm
#> 61  https://www.sec.gov/Archives/edgar/data/320193/000118143106043325/0001181431-06-043325-index.htm
#> 62  https://www.sec.gov/Archives/edgar/data/320193/000110465906047739/0001104659-06-047739-index.htm
#> 63  https://www.sec.gov/Archives/edgar/data/320193/000118143106040870/0001181431-06-040870-index.htm
#> 64  https://www.sec.gov/Archives/edgar/data/320193/000118143106033706/0001181431-06-033706-index.htm
#> 65  https://www.sec.gov/Archives/edgar/data/320193/000118143106028768/0001181431-06-028768-index.htm
#> 66  https://www.sec.gov/Archives/edgar/data/320193/000118143106028769/0001181431-06-028769-index.htm
#> 67  https://www.sec.gov/Archives/edgar/data/320193/000110465906031303/0001104659-06-031303-index.htm
#> 68  https://www.sec.gov/Archives/edgar/data/320193/000118143106027492/0001181431-06-027492-index.htm
#> 69  https://www.sec.gov/Archives/edgar/data/320193/000118143106027491/0001181431-06-027491-index.htm
#> 70  https://www.sec.gov/Archives/edgar/data/320193/000118143106027619/0001181431-06-027619-index.htm
#> 71  https://www.sec.gov/Archives/edgar/data/320193/000118143106027012/0001181431-06-027012-index.htm
#> 72  https://www.sec.gov/Archives/edgar/data/320193/000118143106024759/0001181431-06-024759-index.htm
#> 73  https://www.sec.gov/Archives/edgar/data/320193/000110465906025992/0001104659-06-025992-index.htm
#> 74  https://www.sec.gov/Archives/edgar/data/320193/000118143106023722/0001181431-06-023722-index.htm
#> 75  https://www.sec.gov/Archives/edgar/data/320193/000118143106023326/0001181431-06-023326-index.htm
#> 76  https://www.sec.gov/Archives/edgar/data/320193/000118143106023327/0001181431-06-023327-index.htm
#> 77  https://www.sec.gov/Archives/edgar/data/320193/000118143106023019/0001181431-06-023019-index.htm
#> 78  https://www.sec.gov/Archives/edgar/data/320193/000118143106022567/0001181431-06-022567-index.htm
#> 79  https://www.sec.gov/Archives/edgar/data/320193/000118143106022566/0001181431-06-022566-index.htm
#> 80  https://www.sec.gov/Archives/edgar/data/320193/000118143106022568/0001181431-06-022568-index.htm
#> 81  https://www.sec.gov/Archives/edgar/data/320193/000118143106021500/0001181431-06-021500-index.htm
#>       form filing_date    size
#> 1   POSASR  2007-07-16 1032070
#> 2   POS AM  2007-07-16 1095382
#> 3    S-3/A  2005-07-22 1497372
#> 4      S-3  2005-06-28 1961178
#> 5    S-4/A  2004-06-17 3237214
#> 6   POS AM  2007-07-16 1095382
#> 7   POSASR  2007-07-16 1032070
#> 8    S-3/A  2005-07-22 1497372
#> 9      S-3  2005-06-28 1961178
#> 10   S-4/A  2004-06-17 3237214
#> 11       4  2016-08-26    4637
#> 12       4  2016-06-21    4534
#> 13       3  2015-12-14    3570
#> 14       4  2010-07-12    6469
#> 15     3/A  2010-06-28    7186
#> 16       3  2010-06-01    7052
#> 17       D  2015-10-01    6092
#> 18       D  2013-02-14   13864
#> 19  REGDEX  2007-09-17    1603
#> 20   424B3  2005-05-09 3670621
#> 21   F-4/A  2005-05-05 5308203
#> 22     F-4  2005-04-15 5784728
#> 23   F-3/A  2004-02-26  638116
#> 24     F-3  2003-09-25  572241
#> 25   424B3  2003-07-23 4439660
#> 26   F-4/A  2003-07-16 5249895
#> 27     F-4  2003-06-18 6535344
#> 28       4  2005-04-12    7711
#> 29       4  2005-03-31    4730
#> 30       4  2004-04-06    4942
#> 31       4  2002-08-23   47911
#> 32       3  2002-02-08   11088
#> 33       D  2017-10-11    6767
#> 34       4  2007-01-04   22984
#> 35    10-K  2006-12-29 4280524
#> 36    10-Q  2006-12-29 1714502
#> 37 NT 10-K  2006-12-15   60013
#> 38       3  2006-12-05    3221
#> 39       4  2006-12-05    5712
#> 40     8-K  2006-10-27   18706
#> 41     8-K  2006-10-18  308586
#> 42  UPLOAD  2006-10-06   20693
#> 43     8-K  2006-10-04   31356
#> 44       4  2006-09-15    4942
#> 45       3  2006-09-07    4610
#> 46       4  2006-09-07   11977
#> 47       4  2006-09-01   16331
#> 48     8-K  2006-09-01   37818
#> 49  UPLOAD  2006-08-28   64802
#> 50       4  2006-08-17    5469
#> 51 CORRESP  2006-08-17   52363
#> 52 NT 10-Q  2006-08-11   34859
#> 53     8-K  2006-08-11   28764
#> 54       4  2006-08-08    5630
#> 55       4  2006-08-08    5499
#> 56     8-K  2006-08-04   31216
#> 57       4  2006-08-02    5143
#> 58       4  2006-07-28    8738
#> 59     4/A  2006-07-26   16171
#> 60       4  2006-07-26   16094
#> 61       4  2006-07-24   16094
#> 62     8-K  2006-07-19  524011
#> 63       4  2006-07-06    6682
#> 64       4  2006-05-30    5510
#> 65       4  2006-05-08   51853
#> 66       4  2006-05-08   17256
#> 67    10-Q  2006-05-05 1543637
#> 68       4  2006-05-03   52730
#> 69       4  2006-05-03   51536
#> 70       4  2006-05-03    8245
#> 71       4  2006-05-02    9595
#> 72     8-K  2006-04-21   11972
#> 73     8-K  2006-04-19  500482
#> 74       4  2006-04-13    8738
#> 75       4  2006-04-11   21252
#> 76       4  2006-04-11   48103
#> 77       4  2006-04-07   25248
#> 78       4  2006-04-05   36380
#> 79       4  2006-04-05   42050
#> 80       4  2006-04-05   28897
#> 81       4  2006-04-03   23766

# }